% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{bisection_ci_single_bound_cpp}
\alias{bisection_ci_single_bound_cpp}
\title{Single-threaded helper for computing one CI bound}
\usage{
bisection_ci_single_bound_cpp(
  pval_fn,
  nsim_exact_test,
  l,
  u,
  pval_th,
  tol,
  transform_responses,
  lower,
  num_cores = 1L
)
}
\description{
Single-threaded helper for computing one CI bound
}
\keyword{internal}
