% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_bai_distr_parallel_cpp}
\alias{compute_bai_distr_parallel_cpp}
\title{Fast Bai Adjusted T Statistic for Multiple Permutations}
\usage{
compute_bai_distr_parallel_cpp(
  y,
  w_mat,
  m_mat,
  delta,
  halves_idx,
  convex_flag,
  num_cores
)
}
\description{
Fast Bai Adjusted T Statistic for Multiple Permutations
}
