% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_kk_wilcox_distr_parallel_cpp}
\alias{compute_kk_wilcox_distr_parallel_cpp}
\title{Fast KK Wilcoxon Statistic for Multiple Permutations}
\usage{
compute_kk_wilcox_distr_parallel_cpp(
  y,
  w_mat,
  m_mat,
  delta,
  transform_code,
  is_fixed_matching,
  num_cores
)
}
\description{
Fast KK Wilcoxon Statistic for Multiple Permutations
}
