% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_abstract.R
\name{Design}
\alias{Design}
\title{An Abstract Experimental Design}
\description{
Internal method.
An abstract R6 Class encapsulating the data and functionality for an experimental design.
This class takes care of data storage and response handling.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Design-new}{\code{Design$new()}}
\item \href{#method-Design-add_subject}{\code{Design$add_subject()}}
\item \href{#method-Design-add_subject_response}{\code{Design$add_subject_response()}}
\item \href{#method-Design-add_all_subject_responses}{\code{Design$add_all_subject_responses()}}
\item \href{#method-Design-add_all_subject_assignments}{\code{Design$add_all_subject_assignments()}}
\item \href{#method-Design-is_fixed_sample_size}{\code{Design$is_fixed_sample_size()}}
\item \href{#method-Design-assert_all_subjects_arrived}{\code{Design$assert_all_subjects_arrived()}}
\item \href{#method-Design-assert_all_responses_recorded}{\code{Design$assert_all_responses_recorded()}}
\item \href{#method-Design-check_experiment_completed}{\code{Design$check_experiment_completed()}}
\item \href{#method-Design-assert_even_allocation}{\code{Design$assert_even_allocation()}}
\item \href{#method-Design-assert_fixed_sample}{\code{Design$assert_fixed_sample()}}
\item \href{#method-Design-any_censoring}{\code{Design$any_censoring()}}
\item \href{#method-Design-get_t}{\code{Design$get_t()}}
\item \href{#method-Design-get_X_raw}{\code{Design$get_X_raw()}}
\item \href{#method-Design-get_X_imp}{\code{Design$get_X_imp()}}
\item \href{#method-Design-get_X}{\code{Design$get_X()}}
\item \href{#method-Design-get_y}{\code{Design$get_y()}}
\item \href{#method-Design-get_w}{\code{Design$get_w()}}
\item \href{#method-Design-get_block_ids}{\code{Design$get_block_ids()}}
\item \href{#method-Design-get_n}{\code{Design$get_n()}}
\item \href{#method-Design-get_dead}{\code{Design$get_dead()}}
\item \href{#method-Design-get_prob_T}{\code{Design$get_prob_T()}}
\item \href{#method-Design-get_response_type}{\code{Design$get_response_type()}}
\item \href{#method-Design-duplicate}{\code{Design$duplicate()}}
\item \href{#method-Design-resample_design}{\code{Design$resample_design()}}
\item \href{#method-Design-clone}{\code{Design$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-new"></a>}}
\if{latex}{\out{\hypertarget{method-Design-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$new(
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = FALSE,
  n = NULL,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_subject"></a>}}
\if{latex}{\out{\hypertarget{method-Design-add_subject}{}}}
\subsection{Method \code{add_subject()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$add_subject(x_new, allow_new_cols = TRUE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_subject_response"></a>}}
\if{latex}{\out{\hypertarget{method-Design-add_subject_response}{}}}
\subsection{Method \code{add_subject_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$add_subject_response(t, y, dead = 1)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_all_subject_responses"></a>}}
\if{latex}{\out{\hypertarget{method-Design-add_all_subject_responses}{}}}
\subsection{Method \code{add_all_subject_responses()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$add_all_subject_responses(ys, deads = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_all_subject_assignments"></a>}}
\if{latex}{\out{\hypertarget{method-Design-add_all_subject_assignments}{}}}
\subsection{Method \code{add_all_subject_assignments()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$add_all_subject_assignments(w)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-is_fixed_sample_size"></a>}}
\if{latex}{\out{\hypertarget{method-Design-is_fixed_sample_size}{}}}
\subsection{Method \code{is_fixed_sample_size()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$is_fixed_sample_size()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_all_subjects_arrived"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_all_subjects_arrived}{}}}
\subsection{Method \code{assert_all_subjects_arrived()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_all_subjects_arrived()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_all_responses_recorded"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_all_responses_recorded}{}}}
\subsection{Method \code{assert_all_responses_recorded()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_all_responses_recorded()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-check_experiment_completed"></a>}}
\if{latex}{\out{\hypertarget{method-Design-check_experiment_completed}{}}}
\subsection{Method \code{check_experiment_completed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$check_experiment_completed()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_even_allocation"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_even_allocation}{}}}
\subsection{Method \code{assert_even_allocation()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_even_allocation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_fixed_sample"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_fixed_sample}{}}}
\subsection{Method \code{assert_fixed_sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_fixed_sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-any_censoring"></a>}}
\if{latex}{\out{\hypertarget{method-Design-any_censoring}{}}}
\subsection{Method \code{any_censoring()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$any_censoring()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_t"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_t}{}}}
\subsection{Method \code{get_t()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_t()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X_raw"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_X_raw}{}}}
\subsection{Method \code{get_X_raw()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_X_raw()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X_imp"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_X_imp}{}}}
\subsection{Method \code{get_X_imp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_X_imp()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_X}{}}}
\subsection{Method \code{get_X()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_X()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_y"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_y}{}}}
\subsection{Method \code{get_y()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_y()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_w"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_w}{}}}
\subsection{Method \code{get_w()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_w()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_block_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_block_ids}{}}}
\subsection{Method \code{get_block_ids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_block_ids()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_n"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_n}{}}}
\subsection{Method \code{get_n()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_n()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_dead"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_dead}{}}}
\subsection{Method \code{get_dead()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_dead()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_prob_T"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_prob_T}{}}}
\subsection{Method \code{get_prob_T()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_prob_T()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_response_type"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_response_type}{}}}
\subsection{Method \code{get_response_type()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_response_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-duplicate"></a>}}
\if{latex}{\out{\hypertarget{method-Design-duplicate}{}}}
\subsection{Method \code{duplicate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$duplicate(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-resample_design"></a>}}
\if{latex}{\out{\hypertarget{method-Design-resample_design}{}}}
\subsection{Method \code{resample_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$resample_design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Design-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
