% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_inference_continuous_quantile_regr_multi.R
\name{DesignInferenceContinMultiQuantileRegr}
\alias{DesignInferenceContinMultiQuantileRegr}
\title{Multivariate Quantile Regression Inference for Continuous Responses}
\description{
Fits a quantile regression for continuous responses using the treatment
indicator and all recorded covariates as predictors. The treatment effect is
reported on the response scale at quantile \code{tau}; by default
\code{tau = 0.5}, so this is median regression.
}
\examples{

## ------------------------------------------------
## Method `DesignInferenceContinMultiQuantileRegr$new`
## ------------------------------------------------

set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignBernoulli$new(n = nrow(x_dat), response_type = "continuous", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(c(1.2, 0.9, 1.5, 1.8, 2.1, 1.7, 2.6, 2.2))
infer <- DesignInferenceContinMultiQuantileRegr$new(seq_des, verbose = FALSE)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{\link[SeqExpMatch:DesignInferenceContinUnivQuantileRegr]{SeqExpMatch::DesignInferenceContinUnivQuantileRegr}} -> \code{DesignInferenceContinMultiQuantileRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceContinMultiQuantileRegr-new}{\code{DesignInferenceContinMultiQuantileRegr$new()}}
\item \href{#method-DesignInferenceContinMultiQuantileRegr-clone}{\code{DesignInferenceContinMultiQuantileRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceContinUnivQuantileRegr" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceContinUnivQuantileRegr.html#method-DesignInferenceContinUnivQuantileRegr-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceContinUnivQuantileRegr$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceContinUnivQuantileRegr" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceContinUnivQuantileRegr.html#method-DesignInferenceContinUnivQuantileRegr-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceContinUnivQuantileRegr$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceContinUnivQuantileRegr" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceContinUnivQuantileRegr.html#method-DesignInferenceContinUnivQuantileRegr-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceContinUnivQuantileRegr$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceContinMultiQuantileRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceContinMultiQuantileRegr-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multivariate quantile-regression inference object for a completed
design with a continuous response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceContinMultiQuantileRegr$new(
  des_obj,
  tau = 0.5,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{SeqDesign} object with a continuous response.}

\item{\code{tau}}{The quantile level for regression, strictly between 0 and 1. The default is
\code{tau = 0.5}.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignBernoulli$new(n = nrow(x_dat), response_type = "continuous", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(c(1.2, 0.9, 1.5, 1.8, 2.1, 1.7, 2.6, 2.2))
infer <- DesignInferenceContinMultiQuantileRegr$new(seq_des, verbose = FALSE)
infer

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceContinMultiQuantileRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceContinMultiQuantileRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceContinMultiQuantileRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
