% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_count_quasipoisson_multi.R
\name{DesignInferenceCountMultiQuasiPoissonRegr}
\alias{DesignInferenceCountMultiQuasiPoissonRegr}
\title{Multivariate Quasi-Poisson Regression Inference for Count Responses}
\description{
Fits a quasi-Poisson log-link regression for count responses using the treatment
indicator and all recorded covariates. The treatment effect is reported on the
log-rate scale and inference uses the model-based quasi-Poisson variance with an
estimated dispersion parameter.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignBernoulli$
  new(
  n = nrow(x_dat),
  response_type = "count",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 1, 2, 2, 3, 3, 4))
infer <- DesignInferenceCountMultiQuasiPoissonRegr$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceMLEorKMforGLMs} -> \code{\link[SeqExpMatch:DesignInferenceCountUnivRobustPoissonRegr]{SeqExpMatch::DesignInferenceCountUnivRobustPoissonRegr}} -> \code{\link[SeqExpMatch:DesignInferenceCountUnivQuasiPoissonRegr]{SeqExpMatch::DesignInferenceCountUnivQuasiPoissonRegr}} -> \code{DesignInferenceCountMultiQuasiPoissonRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceCountMultiQuasiPoissonRegr-new}{\code{DesignInferenceCountMultiQuasiPoissonRegr$new()}}
\item \href{#method-DesignInferenceCountMultiQuasiPoissonRegr-clone}{\code{DesignInferenceCountMultiQuasiPoissonRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceMLEorKMforGLMs" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceMLEorKMforGLMs.html#method-DesignInferenceMLEorKMforGLMs-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceMLEorKMforGLMs$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceMLEorKMforGLMs" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceMLEorKMforGLMs.html#method-DesignInferenceMLEorKMforGLMs-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceMLEorKMforGLMs$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceCountUnivRobustPoissonRegr" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceCountUnivRobustPoissonRegr.html#method-DesignInferenceCountUnivRobustPoissonRegr-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceCountUnivRobustPoissonRegr$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountMultiQuasiPoissonRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountMultiQuasiPoissonRegr-new}{}}}
\subsection{Method \code{new()}}{
Initialize a sequential experimental design estimation and test object
after the sequential design is completed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountMultiQuasiPoissonRegr$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object whose entire n subjects
are assigned and response y is recorded within.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize
the sampling during randomization-based inference and
bootstrap resampling.
The default is 1 for serial computation. For simple
estimators (e.g. mean difference and KK compound),
parallelization is achieved with zero-overhead C++ OpenMP.
For complex models (e.g. GLMs),
parallelization falls back to R's
\code{parallel::mclapply}, which incurs
session-forking overhead.}

\item{\code{verbose}}{A flag indicating whether messages should be
displayed to the user. Default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountMultiQuasiPoissonRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountMultiQuasiPoissonRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountMultiQuasiPoissonRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
