% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_inference_count_poisson_KK_cpoisson_combined_likelihood_univ.R
\name{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood}
\alias{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood}
\title{Univariate Conditional Poisson Combined-Likelihood Compound Inference for KK Designs}
\description{
Fits a compound estimator for KK matching-on-the-fly designs with count
responses using only the treatment indicator (no additional covariates).
Jointly maximises L_cond_Poisson(pairs) + L_Poisson(reservoir) via BFGS.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$
  new(
  n = nrow(x_dat),
  response_type = "count",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 1, 2, 2, 3, 3, 4))
infer <- DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$
  new(seq_des, verbose
= FALSE)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceKKPassThrough} -> \code{SeqExpMatch::DesignInferenceAbstractKKPoissonCPoissonCombinedLikelihood} -> \code{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-new}{\code{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$new()}}
\item \href{#method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_treatment_estimate}{\code{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$compute_treatment_estimate()}}
\item \href{#method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_asymp_confidence_interval}{\code{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$compute_asymp_confidence_interval()}}
\item \href{#method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{\code{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-clone}{\code{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceKKPassThrough.html#method-DesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-new}{}}}
\subsection{Method \code{new()}}{
Initialize the inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A SeqDesign object (must be a KK design).}

\item{\code{num_cores}}{Number of CPU cores for parallel processing.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Returns the estimated treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$compute_treatment_estimate(
  
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes the asymptotic confidence interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes the asymptotic p-value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null difference to test against. For any
treatment effect at all this is set to zero (the default).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountPoissonUnivKKCPoissonCombinedLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
