% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_inference_count_hurdle_poisson_KK_combined_likelihood_univ.R
\name{DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood}
\alias{DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood}
\title{Univariate KK Hurdle Poisson Combined-Likelihood Inference for Count Responses}
\description{
Fits a KK hurdle-Poisson combined-likelihood model for count responses using
only the treatment indicator in both the hurdle and positive-count
components. Matched pairs contribute pair-specific random intercepts, while
reservoir subjects contribute through a common fixed reservoir intercept. The
reported treatment effect is the treatment coefficient from the positive-count
component on the log-rate scale.
}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceKKPassThrough} -> \code{SeqExpMatch::DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood} -> \code{DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood-clone}{\code{DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceKKPassThrough.html#method-DesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood.html#method-DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood.html#method-DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood.html#method-DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood" data-id="initialize"><a href='../../SeqExpMatch/html/DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood.html#method-DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood-initialize'><code>SeqExpMatch::DesignInferenceAbstractKKHurdlePoissonCombinedLikelihood$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceCountUnivKKHurdlePoissonCombinedLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
