% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_incidence_kk_exact_zhang.R
\name{DesignInferenceIncidKKExactZhang}
\alias{DesignInferenceIncidKKExactZhang}
\title{Exact incidence inference for KK designs via Zhang's combined matched-pair / reservoir tests}
\description{
Extends \code{DesignInferenceIncidExactZhang} with the KK matched-pair exact test,
producing Zhang's combined exact p-values and confidence intervals for KK
matching-on-the-fly designs with incidence outcomes.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$
  new(
  n = nrow(x_dat),
  response_type = "incidence",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 0, 1, 0, 1, 1, 0))
infer <- DesignInferenceIncidKKExactZhang$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceAbstractZhangCombinedBase} -> \code{SeqExpMatch::DesignInferenceIncidExactZhangAbstract} -> \code{\link[SeqExpMatch:DesignInferenceIncidExactZhang]{SeqExpMatch::DesignInferenceIncidExactZhang}} -> \code{DesignInferenceIncidKKExactZhang}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceIncidKKExactZhang-new}{\code{DesignInferenceIncidKKExactZhang$new()}}
\item \href{#method-DesignInferenceIncidKKExactZhang-clone}{\code{DesignInferenceIncidKKExactZhang$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceIncidExactZhang$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidExactZhang" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInferenceIncidExactZhang.html#method-DesignInferenceIncidExactZhang-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInferenceIncidExactZhang$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidKKExactZhang-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidKKExactZhang-new}{}}}
\subsection{Method \code{new()}}{
Initialize the exact incidence inference object for a KK design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidKKExactZhang$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesignKK14 object (or subclass) with an incidence response.}

\item{\code{num_cores}}{Number of CPU cores for parallel processing.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidKKExactZhang-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidKKExactZhang-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidKKExactZhang$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
