% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_inference_incidence_binomial_identity_multi.R
\name{DesignInferenceIncidMultiBinomialIdentityRiskDiff}
\alias{DesignInferenceIncidMultiBinomialIdentityRiskDiff}
\title{Multivariate Binomial Identity-Link Regression for Binary Responses}
\description{
Fits a binomial GLM with identity link for binary (incidence) responses using
treatment and all recorded covariates as predictors. The treatment effect is
reported on the risk-difference scale.
}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceIncidConstrainedBinomialAbstract} -> \code{SeqExpMatch::DesignInferenceIncidBinomialIdentityAbstract} -> \code{\link[SeqExpMatch:DesignInferenceIncidUnivBinomialIdentityRiskDiff]{SeqExpMatch::DesignInferenceIncidUnivBinomialIdentityRiskDiff}} -> \code{DesignInferenceIncidMultiBinomialIdentityRiskDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceIncidMultiBinomialIdentityRiskDiff-clone}{\code{DesignInferenceIncidMultiBinomialIdentityRiskDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidConstrainedBinomialAbstract" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceIncidConstrainedBinomialAbstract.html#method-DesignInferenceIncidConstrainedBinomialAbstract-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceIncidConstrainedBinomialAbstract$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidConstrainedBinomialAbstract" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceIncidConstrainedBinomialAbstract.html#method-DesignInferenceIncidConstrainedBinomialAbstract-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceIncidConstrainedBinomialAbstract$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidConstrainedBinomialAbstract" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceIncidConstrainedBinomialAbstract.html#method-DesignInferenceIncidConstrainedBinomialAbstract-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceIncidConstrainedBinomialAbstract$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidConstrainedBinomialAbstract" data-id="initialize"><a href='../../SeqExpMatch/html/DesignInferenceIncidConstrainedBinomialAbstract.html#method-DesignInferenceIncidConstrainedBinomialAbstract-initialize'><code>SeqExpMatch::DesignInferenceIncidConstrainedBinomialAbstract$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidMultiBinomialIdentityRiskDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidMultiBinomialIdentityRiskDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidMultiBinomialIdentityRiskDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
