% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_incidence_KK_gcomp_multi.R
\name{DesignInferenceIncidMultiKKGCompRiskRatio}
\alias{DesignInferenceIncidMultiKKGCompRiskRatio}
\title{Multivariate G-Computation Risk-Ratio Inference for KK Designs with Binary Responses}
\description{
Fits an all-subject logistic working model for a KK incidence outcome using
treatment and all recorded covariates, then estimates the marginal risk ratio
by standardizing predicted risks under all-treated and all-control assignments
over the empirical covariate distribution. Matched pairs are treated as
clusters and reservoir subjects are treated as singletons when computing the
sandwich covariance.
}
\details{
The point estimate is returned on the risk-ratio scale. Confidence intervals
and p-values use the delta method on the log-risk-ratio scale and then map
back to the risk-ratio scale.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$
  new(
  n = nrow(x_dat),
  response_type = "incidence",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 0, 1, 0, 1, 1, 0))
infer <- DesignInferenceIncidMultiKKGCompRiskRatio$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceKKPassThrough} -> \code{SeqExpMatch::DesignInferenceAbstractKKMarginalIncid} -> \code{SeqExpMatch::DesignInferenceIncidKKGCompAbstract} -> \code{\link[SeqExpMatch:DesignInferenceIncidUnivKKGCompRiskRatio]{SeqExpMatch::DesignInferenceIncidUnivKKGCompRiskRatio}} -> \code{DesignInferenceIncidMultiKKGCompRiskRatio}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceIncidMultiKKGCompRiskRatio-new}{\code{DesignInferenceIncidMultiKKGCompRiskRatio$new()}}
\item \href{#method-DesignInferenceIncidMultiKKGCompRiskRatio-clone}{\code{DesignInferenceIncidMultiKKGCompRiskRatio$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceKKPassThrough.html#method-DesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidUnivKKGCompRiskRatio" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceIncidUnivKKGCompRiskRatio.html#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceIncidUnivKKGCompRiskRatio$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidUnivKKGCompRiskRatio" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceIncidUnivKKGCompRiskRatio.html#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceIncidUnivKKGCompRiskRatio$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidUnivKKGCompRiskRatio" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInferenceIncidUnivKKGCompRiskRatio.html#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInferenceIncidUnivKKGCompRiskRatio$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceIncidUnivKKGCompRiskRatio" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceIncidUnivKKGCompRiskRatio.html#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceIncidUnivKKGCompRiskRatio$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidMultiKKGCompRiskRatio-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidMultiKKGCompRiskRatio-new}{}}}
\subsection{Method \code{new()}}{
Initialize the multivariate KK g-computation RR inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidMultiKKGCompRiskRatio$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A completed KK \code{SeqDesign} object with an incidence response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidMultiKKGCompRiskRatio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidMultiKKGCompRiskRatio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidMultiKKGCompRiskRatio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
