% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_incidence_gcomp_univ.R
\name{DesignInferenceIncidUnivGCompRiskDiff}
\alias{DesignInferenceIncidUnivGCompRiskDiff}
\title{Univariate G-Computation Risk-Difference Inference for Binary Responses}
\description{
Fits a univariate logistic working model for an incidence outcome and then
estimates the marginal risk difference by standardizing predicted risks under
all-treated and all-control assignments over the empirical covariate
distribution.
}
\examples{

## ------------------------------------------------
## Method `DesignInferenceIncidUnivGCompRiskDiff$new`
## ------------------------------------------------

\dontrun{
seq_des = SeqDesignBernoulli$new(n = 20, response_type = "incidence")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	p_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbinom(1, 1, p_i))
}
seq_des_inf = DesignInferenceIncidUnivGCompRiskDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceIncidGCompAbstract} -> \code{DesignInferenceIncidUnivGCompRiskDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceIncidUnivGCompRiskDiff-new}{\code{DesignInferenceIncidUnivGCompRiskDiff$new()}}
\item \href{#method-DesignInferenceIncidUnivGCompRiskDiff-compute_treatment_estimate}{\code{DesignInferenceIncidUnivGCompRiskDiff$compute_treatment_estimate()}}
\item \href{#method-DesignInferenceIncidUnivGCompRiskDiff-compute_asymp_confidence_interval}{\code{DesignInferenceIncidUnivGCompRiskDiff$compute_asymp_confidence_interval()}}
\item \href{#method-DesignInferenceIncidUnivGCompRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect}{\code{DesignInferenceIncidUnivGCompRiskDiff$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-DesignInferenceIncidUnivGCompRiskDiff-compute_bootstrap_two_sided_pval}{\code{DesignInferenceIncidUnivGCompRiskDiff$compute_bootstrap_two_sided_pval()}}
\item \href{#method-DesignInferenceIncidUnivGCompRiskDiff-clone}{\code{DesignInferenceIncidUnivGCompRiskDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivGCompRiskDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivGCompRiskDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize the g-computation RD inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivGCompRiskDiff$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{SeqDesign} object with an incidence response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = SeqDesignBernoulli$new(n = 20, response_type = "incidence")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	p_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbinom(1, 1, p_i))
}
seq_des_inf = DesignInferenceIncidUnivGCompRiskDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivGCompRiskDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivGCompRiskDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the standardized marginal risk-difference estimate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivGCompRiskDiff$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivGCompRiskDiff-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivGCompRiskDiff-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval for the marginal risk difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivGCompRiskDiff$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is 1 - \code{alpha}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivGCompRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivGCompRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the marginal risk difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivGCompRiskDiff$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null risk difference to test against. The default is 0.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivGCompRiskDiff-compute_bootstrap_two_sided_pval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivGCompRiskDiff-compute_bootstrap_two_sided_pval}{}}}
\subsection{Method \code{compute_bootstrap_two_sided_pval()}}{
Computes a bootstrap two-sided p-value for the marginal risk difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivGCompRiskDiff$compute_bootstrap_two_sided_pval(
  delta = NULL,
  B = 501,
  na.rm = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null risk difference to test against. The default is 0.}

\item{\code{B}}{Number of bootstrap samples.}

\item{\code{na.rm}}{Whether to remove non-finite bootstrap replicates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivGCompRiskDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivGCompRiskDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivGCompRiskDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
