% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_incidence_KK_gcomp_univ.R
\name{DesignInferenceIncidUnivKKGCompRiskRatio}
\alias{DesignInferenceIncidUnivKKGCompRiskRatio}
\title{Univariate G-Computation Risk-Ratio Inference for KK Designs with Binary Responses}
\description{
Fits an all-subject logistic working model for a KK incidence outcome using
only the treatment indicator, then estimates the marginal risk ratio by
standardizing predicted risks under all-treated and all-control assignments
over the empirical covariate distribution. Matched pairs are treated as
clusters and reservoir subjects are treated as singletons when computing the
sandwich covariance.
}
\details{
The point estimate is returned on the risk-ratio scale. Confidence intervals
and p-values use the delta method on the log-risk-ratio scale and then map
back to the risk-ratio scale.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$
  new(
  n = nrow(x_dat),
  response_type = "incidence",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 0, 1, 0, 1, 1, 0))
infer <- DesignInferenceIncidUnivKKGCompRiskRatio$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceKKPassThrough} -> \code{SeqExpMatch::DesignInferenceAbstractKKMarginalIncid} -> \code{SeqExpMatch::DesignInferenceIncidKKGCompAbstract} -> \code{DesignInferenceIncidUnivKKGCompRiskRatio}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceIncidUnivKKGCompRiskRatio-new}{\code{DesignInferenceIncidUnivKKGCompRiskRatio$new()}}
\item \href{#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_treatment_estimate}{\code{DesignInferenceIncidUnivKKGCompRiskRatio$compute_treatment_estimate()}}
\item \href{#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_confidence_interval}{\code{DesignInferenceIncidUnivKKGCompRiskRatio$compute_asymp_confidence_interval()}}
\item \href{#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_two_sided_pval_for_treatment_effect}{\code{DesignInferenceIncidUnivKKGCompRiskRatio$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_bootstrap_two_sided_pval}{\code{DesignInferenceIncidUnivKKGCompRiskRatio$compute_bootstrap_two_sided_pval()}}
\item \href{#method-DesignInferenceIncidUnivKKGCompRiskRatio-clone}{\code{DesignInferenceIncidUnivKKGCompRiskRatio$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceKKPassThrough.html#method-DesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKGCompRiskRatio-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKGCompRiskRatio-new}{}}}
\subsection{Method \code{new()}}{
Initialize the KK g-computation RR inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKGCompRiskRatio$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed KK \code{SeqDesign} object with an incidence response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the standardized marginal risk-ratio estimate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKGCompRiskRatio$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval for the marginal risk ratio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKGCompRiskRatio$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is 1 - \code{alpha}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the marginal risk ratio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKGCompRiskRatio$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null risk ratio to test against. The default is 1.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_bootstrap_two_sided_pval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKGCompRiskRatio-compute_bootstrap_two_sided_pval}{}}}
\subsection{Method \code{compute_bootstrap_two_sided_pval()}}{
Computes a bootstrap two-sided p-value for the marginal risk ratio.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKGCompRiskRatio$compute_bootstrap_two_sided_pval(
  delta = NULL,
  B = 501,
  na.rm = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null risk ratio to test against. The default is 1.}

\item{\code{B}}{Number of bootstrap samples.}

\item{\code{na.rm}}{Whether to remove non-finite bootstrap replicates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKGCompRiskRatio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKGCompRiskRatio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKGCompRiskRatio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
