% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_inference_incidence_KK_modified_poisson_univ.R
\name{DesignInferenceIncidUnivKKModifiedPoisson}
\alias{DesignInferenceIncidUnivKKModifiedPoisson}
\title{Univariate Modified-Poisson Inference for KK Designs with Binary Responses}
\description{
Fits an all-subject modified-Poisson working model for incidence outcomes under
a KK matching-on-the-fly design using only the treatment indicator as a
predictor. Matched pairs are treated as clusters and reservoir subjects are
treated as singleton clusters when computing the sandwich covariance, so the
estimated treatment effect is a log risk ratio with cluster-robust inference.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$
  new(
  n = nrow(x_dat),
  response_type = "incidence",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 0, 1, 0, 1, 1, 0))
infer <- DesignInferenceIncidUnivKKModifiedPoisson$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceKKPassThrough} -> \code{SeqExpMatch::DesignInferenceAbstractKKMarginalIncid} -> \code{SeqExpMatch::DesignInferenceAbstractKKModifiedPoisson} -> \code{DesignInferenceIncidUnivKKModifiedPoisson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceIncidUnivKKModifiedPoisson-new}{\code{DesignInferenceIncidUnivKKModifiedPoisson$new()}}
\item \href{#method-DesignInferenceIncidUnivKKModifiedPoisson-compute_treatment_estimate}{\code{DesignInferenceIncidUnivKKModifiedPoisson$compute_treatment_estimate()}}
\item \href{#method-DesignInferenceIncidUnivKKModifiedPoisson-compute_asymp_confidence_interval}{\code{DesignInferenceIncidUnivKKModifiedPoisson$compute_asymp_confidence_interval()}}
\item \href{#method-DesignInferenceIncidUnivKKModifiedPoisson-compute_asymp_two_sided_pval_for_treatment_effect}{\code{DesignInferenceIncidUnivKKModifiedPoisson$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-DesignInferenceIncidUnivKKModifiedPoisson-clone}{\code{DesignInferenceIncidUnivKKModifiedPoisson$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceKKPassThrough.html#method-DesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKModifiedPoisson-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKModifiedPoisson-new}{}}}
\subsection{Method \code{new()}}{
Initialize a univariate modified-Poisson inference object for a completed
KK design with a binary response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKModifiedPoisson$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A completed KK \code{SeqDesign} object with an incidence response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKModifiedPoisson-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKModifiedPoisson-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the modified-Poisson estimate of the treatment effect on the
log-risk-ratio scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKModifiedPoisson$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKModifiedPoisson-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKModifiedPoisson-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval using the cluster-robust
sandwich standard error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKModifiedPoisson$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKModifiedPoisson-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKModifiedPoisson-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKModifiedPoisson$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null treatment effect on the log-risk-ratio scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceIncidUnivKKModifiedPoisson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceIncidUnivKKModifiedPoisson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceIncidUnivKKModifiedPoisson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
