% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_inference_ordinal_KK_cond_cont_ratio_multi.R
\name{DesignInferenceOrdinalMultiKKCondContRatioRegr}
\alias{DesignInferenceOrdinalMultiKKCondContRatioRegr}
\title{Multivariate Conditional Continuation-Ratio Inference for KK Designs}
\description{
Fits a conditional continuation-ratio logit model for ordinal responses under a KK
matching-on-the-fly design, adjusting for baseline covariates.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$new(n = nrow(x_dat), response_type = "ordinal", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- DesignInferenceOrdinalMultiKKCondContRatioRegr$
  new(seq_des, verbose = FALSE)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceKKPassThrough} -> \code{\link[SeqExpMatch:DesignInferenceOrdinalUnivKKCondContRatioRegr]{SeqExpMatch::DesignInferenceOrdinalUnivKKCondContRatioRegr}} -> \code{DesignInferenceOrdinalMultiKKCondContRatioRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceOrdinalMultiKKCondContRatioRegr-new}{\code{DesignInferenceOrdinalMultiKKCondContRatioRegr$new()}}
\item \href{#method-DesignInferenceOrdinalMultiKKCondContRatioRegr-clone}{\code{DesignInferenceOrdinalMultiKKCondContRatioRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceKKPassThrough.html#method-DesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceOrdinalUnivKKCondContRatioRegr" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceOrdinalUnivKKCondContRatioRegr.html#method-DesignInferenceOrdinalUnivKKCondContRatioRegr-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceOrdinalUnivKKCondContRatioRegr$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceOrdinalUnivKKCondContRatioRegr" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceOrdinalUnivKKCondContRatioRegr.html#method-DesignInferenceOrdinalUnivKKCondContRatioRegr-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceOrdinalUnivKKCondContRatioRegr$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceOrdinalUnivKKCondContRatioRegr" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceOrdinalUnivKKCondContRatioRegr.html#method-DesignInferenceOrdinalUnivKKCondContRatioRegr-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceOrdinalUnivKKCondContRatioRegr$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalMultiKKCondContRatioRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalMultiKKCondContRatioRegr-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multivariate conditional continuation-ratio inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalMultiKKCondContRatioRegr$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A SeqDesign object (must be a KK design).}

\item{\code{num_cores}}{Number of CPU cores.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalMultiKKCondContRatioRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalMultiKKCondContRatioRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalMultiKKCondContRatioRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
