% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/design_inference_ordinal_stereotype_probit_multi.R
\name{DesignInferenceOrdinalMultiStereotypeProbitRegr}
\alias{DesignInferenceOrdinalMultiStereotypeProbitRegr}
\title{Multivariate Stereotype Probit Inference for Ordinal Responses}
\description{
Stereotype probit inference for ordinal responses using treatment plus observed covariates.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignBernoulli$
  new(
  n = nrow(x_dat),
  response_type = "ordinal",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- DesignInferenceOrdinalMultiStereotypeProbitRegr$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{\link[SeqExpMatch:DesignInferenceOrdinalUniStereotypeProbitRegr]{SeqExpMatch::DesignInferenceOrdinalUniStereotypeProbitRegr}} -> \code{DesignInferenceOrdinalMultiStereotypeProbitRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceOrdinalMultiStereotypeProbitRegr-new}{\code{DesignInferenceOrdinalMultiStereotypeProbitRegr$new()}}
\item \href{#method-DesignInferenceOrdinalMultiStereotypeProbitRegr-clone}{\code{DesignInferenceOrdinalMultiStereotypeProbitRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceOrdinalUniStereotypeProbitRegr" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceOrdinalUniStereotypeProbitRegr.html#method-DesignInferenceOrdinalUniStereotypeProbitRegr-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceOrdinalUniStereotypeProbitRegr$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceOrdinalUniStereotypeProbitRegr" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceOrdinalUniStereotypeProbitRegr.html#method-DesignInferenceOrdinalUniStereotypeProbitRegr-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceOrdinalUniStereotypeProbitRegr$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceOrdinalUniStereotypeProbitRegr" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceOrdinalUniStereotypeProbitRegr.html#method-DesignInferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceOrdinalUniStereotypeProbitRegr$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalMultiStereotypeProbitRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalMultiStereotypeProbitRegr-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multivariate stereotype probit inference object for a
completed ordinal sequential design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalMultiStereotypeProbitRegr$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{SeqDesign} object with an ordinal response.}

\item{\code{num_cores}}{Number of CPU cores for bootstrap/randomization helpers.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalMultiStereotypeProbitRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalMultiStereotypeProbitRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalMultiStereotypeProbitRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
