% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_ordinal_ridit.R
\name{DesignInferenceOrdinalRidit}
\alias{DesignInferenceOrdinalRidit}
\title{Ridit Analysis for Ordinal Responses}
\description{
Performs Ridit analysis (Relative to an Identified Distribution unit) for
comparing two groups on an ordinal scale. Ridit analysis provides a
distribution-free way to estimate the probability that a randomly selected
member of the treatment group has a better outcome than a randomly selected
member of the control group.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignBernoulli$new(n = nrow(x_dat), response_type = "ordinal", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- DesignInferenceOrdinalRidit$
  new(seq_des, verbose = FALSE)
infer

}
\section{Super class}{
\code{SeqExpMatch::DesignInference} -> \code{DesignInferenceOrdinalRidit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceOrdinalRidit-new}{\code{DesignInferenceOrdinalRidit$new()}}
\item \href{#method-DesignInferenceOrdinalRidit-compute_treatment_estimate}{\code{DesignInferenceOrdinalRidit$compute_treatment_estimate()}}
\item \href{#method-DesignInferenceOrdinalRidit-get_mean_ridit_treatment}{\code{DesignInferenceOrdinalRidit$get_mean_ridit_treatment()}}
\item \href{#method-DesignInferenceOrdinalRidit-get_ridit_scores}{\code{DesignInferenceOrdinalRidit$get_ridit_scores()}}
\item \href{#method-DesignInferenceOrdinalRidit-compute_asymp_confidence_interval}{\code{DesignInferenceOrdinalRidit$compute_asymp_confidence_interval()}}
\item \href{#method-DesignInferenceOrdinalRidit-compute_asymp_two_sided_pval_for_treatment_effect}{\code{DesignInferenceOrdinalRidit$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-DesignInferenceOrdinalRidit-clone}{\code{DesignInferenceOrdinalRidit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalRidit-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalRidit-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Ridit analysis inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalRidit$new(
  seq_des_obj,
  reference = "control",
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object whose entire n subjects are assigned and
response y is recorded within.}

\item{\code{reference}}{The group to use as the "Identified Distribution" (reference).
Must be one of "control", "treatment", or "pooled". Default is "control".}

\item{\code{num_cores}}{The number of CPU cores to use.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalRidit-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalRidit-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Returns the estimated treatment effect (Mean Ridit - 0.5).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalRidit$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The numeric estimate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalRidit-get_mean_ridit_treatment"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalRidit-get_mean_ridit_treatment}{}}}
\subsection{Method \code{get_mean_ridit_treatment()}}{
Returns the Mean Ridit for the treatment group.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalRidit$get_mean_ridit_treatment()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The numeric Mean Ridit.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalRidit-get_ridit_scores"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalRidit-get_ridit_scores}{}}}
\subsection{Method \code{get_ridit_scores()}}{
Returns the ridit scores for all subjects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalRidit$get_ridit_scores()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector of scores.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalRidit-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalRidit-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes the asymptotic confidence interval for the treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalRidit$compute_asymp_confidence_interval(alpha = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Significance level.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector of length 2.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalRidit-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalRidit-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes the p-value for the null hypothesis that Mean Ridit = 0.5.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalRidit$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null value (centered at 0, so delta=0 means Ridit=0.5).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The p-value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalRidit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalRidit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalRidit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
