% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_ordinal_KK_glmm_probit_univ.R
\name{DesignInferenceOrdinalUnivKKGLMMProbit}
\alias{DesignInferenceOrdinalUnivKKGLMMProbit}
\title{Univariate Probit CLMM Inference for KK Designs with Ordinal Response}
\description{
Fits a cumulative-link mixed model via \code{ordinal::clmm} with a probit
link for ordinal responses under a KK matching-on-the-fly design using only
the treatment indicator as a fixed effect and a random intercept for each
matched pair. Reservoir subjects each form their own singleton cluster.
Inference is based on the Wald Z-statistic.
}
\details{
This class requires the \pkg{ordinal} package, which is listed under
\code{Suggests} and is not installed automatically with \pkg{SeqExpMatch}.
Install \pkg{ordinal} manually before using this class.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$
  new(
  n = nrow(x_dat),
  response_type = "ordinal",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- DesignInferenceOrdinalUnivKKGLMMProbit$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferenceKKPassThrough} -> \code{SeqExpMatch::DesignInferenceAbstractKKOrdinalCLMM} -> \code{DesignInferenceOrdinalUnivKKGLMMProbit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferenceOrdinalUnivKKGLMMProbit-new}{\code{DesignInferenceOrdinalUnivKKGLMMProbit$new()}}
\item \href{#method-DesignInferenceOrdinalUnivKKGLMMProbit-clone}{\code{DesignInferenceOrdinalUnivKKGLMMProbit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferenceKKPassThrough.html#method-DesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceAbstractKKOrdinalCLMM" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferenceAbstractKKOrdinalCLMM.html#method-DesignInferenceAbstractKKOrdinalCLMM-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferenceAbstractKKOrdinalCLMM$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceAbstractKKOrdinalCLMM" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferenceAbstractKKOrdinalCLMM.html#method-DesignInferenceAbstractKKOrdinalCLMM-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferenceAbstractKKOrdinalCLMM$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferenceAbstractKKOrdinalCLMM" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferenceAbstractKKOrdinalCLMM.html#method-DesignInferenceAbstractKKOrdinalCLMM-compute_treatment_estimate'><code>SeqExpMatch::DesignInferenceAbstractKKOrdinalCLMM$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalUnivKKGLMMProbit-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalUnivKKGLMMProbit-new}{}}}
\subsection{Method \code{new()}}{
Initialize a univariate ordinal probit CLMM inference object for a
completed KK design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalUnivKKGLMMProbit$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A SeqDesign object (must be a KK design) whose entire
n subjects are assigned and whose ordinal response y is
recorded.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize the
sampling during randomization-based inference and bootstrap
resampling.}

\item{\code{verbose}}{Whether to print progress messages. Default is
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferenceOrdinalUnivKKGLMMProbit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferenceOrdinalUnivKKGLMMProbit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferenceOrdinalUnivKKGLMMProbit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
