% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_proportion_gcomp_multi.R
\name{DesignInferencePropMultiGCompMeanDiff}
\alias{DesignInferencePropMultiGCompMeanDiff}
\title{Multivariate G-Computation Mean-Difference Inference for Proportion Responses}
\description{
Fits a multivariate fractional-logit working model for a proportion outcome
and then estimates the marginal mean difference by standardizing predicted
mean proportions under all-treated and all-control assignments over the
empirical covariate distribution.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignBernoulli$
  new(
  n = nrow(x_dat),
  response_type = "proportion",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0.10, 0.25, 0.20, 0.40, 0.35, 0.55, 0.60, 0.75))
infer <- DesignInferencePropMultiGCompMeanDiff$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferencePropGCompAbstract} -> \code{\link[SeqExpMatch:DesignInferencePropUniGCompMeanDiff]{SeqExpMatch::DesignInferencePropUniGCompMeanDiff}} -> \code{DesignInferencePropMultiGCompMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferencePropMultiGCompMeanDiff-new}{\code{DesignInferencePropMultiGCompMeanDiff$new()}}
\item \href{#method-DesignInferencePropMultiGCompMeanDiff-clone}{\code{DesignInferencePropMultiGCompMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferencePropGCompAbstract" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferencePropGCompAbstract.html#method-DesignInferencePropGCompAbstract-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferencePropGCompAbstract$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferencePropUniGCompMeanDiff" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/DesignInferencePropUniGCompMeanDiff.html#method-DesignInferencePropUniGCompMeanDiff-compute_asymp_confidence_interval'><code>SeqExpMatch::DesignInferencePropUniGCompMeanDiff$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferencePropUniGCompMeanDiff" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInferencePropUniGCompMeanDiff.html#method-DesignInferencePropUniGCompMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInferencePropUniGCompMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferencePropUniGCompMeanDiff" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/DesignInferencePropUniGCompMeanDiff.html#method-DesignInferencePropUniGCompMeanDiff-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::DesignInferencePropUniGCompMeanDiff$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferencePropUniGCompMeanDiff" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/DesignInferencePropUniGCompMeanDiff.html#method-DesignInferencePropUniGCompMeanDiff-compute_treatment_estimate'><code>SeqExpMatch::DesignInferencePropUniGCompMeanDiff$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropMultiGCompMeanDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropMultiGCompMeanDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize the multivariate g-computation mean-difference inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropMultiGCompMeanDiff$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{SeqDesign} object with a proportion response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropMultiGCompMeanDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropMultiGCompMeanDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropMultiGCompMeanDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
