% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_inference_proportion_gcomp_univ.R
\name{DesignInferencePropUniGCompMeanDiff}
\alias{DesignInferencePropUniGCompMeanDiff}
\title{Univariate G-Computation Mean-Difference Inference for Proportion Responses}
\description{
Fits a univariate fractional-logit working model for a proportion outcome and
then estimates the marginal mean difference by standardizing predicted mean
proportions under all-treated and all-control assignments over the empirical
covariate distribution.
}
\examples{

## ------------------------------------------------
## Method `DesignInferencePropUniGCompMeanDiff$new`
## ------------------------------------------------

\dontrun{
seq_des = SeqDesignBernoulli$new(n = 20, response_type = "proportion")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	mu_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbeta(1, shape1 = 8 * mu_i, shape2 = 8 * (1 - mu_i)))
}
seq_des_inf = DesignInferencePropUniGCompMeanDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\section{Super classes}{
\code{SeqExpMatch::DesignInference} -> \code{SeqExpMatch::DesignInferencePropGCompAbstract} -> \code{DesignInferencePropUniGCompMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignInferencePropUniGCompMeanDiff-new}{\code{DesignInferencePropUniGCompMeanDiff$new()}}
\item \href{#method-DesignInferencePropUniGCompMeanDiff-compute_treatment_estimate}{\code{DesignInferencePropUniGCompMeanDiff$compute_treatment_estimate()}}
\item \href{#method-DesignInferencePropUniGCompMeanDiff-compute_asymp_confidence_interval}{\code{DesignInferencePropUniGCompMeanDiff$compute_asymp_confidence_interval()}}
\item \href{#method-DesignInferencePropUniGCompMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect}{\code{DesignInferencePropUniGCompMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-DesignInferencePropUniGCompMeanDiff-compute_bootstrap_two_sided_pval}{\code{DesignInferencePropUniGCompMeanDiff$compute_bootstrap_two_sided_pval()}}
\item \href{#method-DesignInferencePropUniGCompMeanDiff-clone}{\code{DesignInferencePropUniGCompMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::DesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::DesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::DesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::DesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::DesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::DesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-duplicate'><code>SeqExpMatch::DesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/DesignInference.html#method-DesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::DesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="DesignInferencePropGCompAbstract" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/DesignInferencePropGCompAbstract.html#method-DesignInferencePropGCompAbstract-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::DesignInferencePropGCompAbstract$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropUniGCompMeanDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropUniGCompMeanDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize the g-computation mean-difference inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropUniGCompMeanDiff$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A completed \code{SeqDesign} object with a proportion response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = SeqDesignBernoulli$new(n = 20, response_type = "proportion")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	mu_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbeta(1, shape1 = 8 * mu_i, shape2 = 8 * (1 - mu_i)))
}
seq_des_inf = DesignInferencePropUniGCompMeanDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropUniGCompMeanDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropUniGCompMeanDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the standardized marginal mean-difference estimate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropUniGCompMeanDiff$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropUniGCompMeanDiff-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropUniGCompMeanDiff-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval for the marginal mean difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropUniGCompMeanDiff$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is 1 - \code{alpha}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropUniGCompMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropUniGCompMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the marginal mean difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropUniGCompMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null mean difference to test against. The default is 0.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropUniGCompMeanDiff-compute_bootstrap_two_sided_pval"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropUniGCompMeanDiff-compute_bootstrap_two_sided_pval}{}}}
\subsection{Method \code{compute_bootstrap_two_sided_pval()}}{
Computes a bootstrap two-sided p-value for the marginal mean difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropUniGCompMeanDiff$compute_bootstrap_two_sided_pval(
  delta = 0,
  B = 501,
  na.rm = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null mean difference to test against. The default is 0.}

\item{\code{B}}{Number of bootstrap samples.}

\item{\code{na.rm}}{Whether to remove non-finite bootstrap replicates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignInferencePropUniGCompMeanDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DesignInferencePropUniGCompMeanDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignInferencePropUniGCompMeanDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
