% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_seq_one_by_one_bernoulli.R
\name{DesignSeqOneByOneBernoulli}
\alias{DesignSeqOneByOneBernoulli}
\title{A completely randomized / Bernoulli Sequential Design}
\description{
A completely randomized / Bernoulli Sequential Design

A completely randomized / Bernoulli Sequential Design
}
\details{
An R6 Class encapsulating the data and functionality for a sequential experimental design.
}
\section{Super classes}{
\\code{EDI::Design}
-> \\code{EDI::DesignSeqOneByOne}
-> \code{DesignSeqOneByOneBernoulli}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignSeqOneByOneBernoulli-new}{\code{DesignSeqOneByOneBernoulli$new()}}
\item \code{DesignSeqOneByOneBernoulli$assign_wt()}
\item \code{DesignSeqOneByOneBernoulli$draw_ws_according_to_design()}
\item \href{#method-DesignSeqOneByOneBernoulli-clone}{\code{DesignSeqOneByOneBernoulli$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Design$add_all_subject_assignments()</code></li>
<li><code>EDI::Design$add_all_subject_responses()</code></li>
<li><code>EDI::Design$add_subject()</code></li>
<li><code>EDI::Design$add_subject_response()</code></li>
<li><code>EDI::Design$any_censoring()</code></li>
<li><code>EDI::Design$assert_all_responses_recorded()</code></li>
<li><code>EDI::Design$assert_all_subjects_arrived()</code></li>
<li><code>EDI::Design$assert_even_allocation()</code></li>
<li><code>EDI::Design$assert_fixed_sample()</code></li>
<li><code>EDI::Design$check_experiment_completed()</code></li>
<li><code>EDI::Design$duplicate()</code></li>
<li><code>EDI::Design$get_X()</code></li>
<li><code>EDI::Design$get_X_imp()</code></li>
<li><code>EDI::Design$get_X_raw()</code></li>
<li><code>EDI::Design$get_block_ids()</code></li>
<li><code>EDI::Design$get_dead()</code></li>
<li><code>EDI::Design$get_n()</code></li>
<li><code>EDI::Design$get_prob_T()</code></li>
<li><code>EDI::Design$get_response_type()</code></li>
<li><code>EDI::Design$get_t()</code></li>
<li><code>EDI::Design$get_w()</code></li>
<li><code>EDI::Design$get_y()</code></li>
<li><code>EDI::Design$is_fixed_sample_size()</code></li>
<li><code>EDI::Design$resample_design()</code></li>
<li><code>EDI::DesignSeqOneByOne$add_subject_to_experiment_and_assign()</code></li>
<li><code>EDI::DesignSeqOneByOne$print_current_subject_assignment()</code></li>
<li><code>EDI::DesignSeqOneByOne$supports_resampling()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneBernoulli-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a Bernoulli sequential experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignSeqOneByOneBernoulli$new(
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{The data type of response values which must be one of the following:
"continuous" (the default),
"incidence",
"proportion",
"count",
"survival",
"ordinal".}

\item{\code{prob_T}}{The probability of the treatment assignment. This defaults to \code{0.5}.}

\item{\code{include_is_missing_as_a_new_feature}}{If missing data is present in a variable,
should we include another dummy variable for its missingness? The default is \code{TRUE}.}

\item{\code{n}}{The sample size (if fixed). Default is \code{NULL}.}

\item{\code{num_cores}}{The number of CPU cores to use.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DesignSeqOneByOneBernoulli` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneBernoulli-assign_wt"></a>}}
\subsection{Method \code{assign_wt()}}{
Assign the next subject to a treatment group
\subsection{Usage}{
\preformatted{DesignSeqOneByOneBernoulli$assign_wt()}
}

\subsection{Returns}{
The treatment assignment (0 or 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneBernoulli-draw_ws_according_to_design"></a>}}
\subsection{Method \code{draw_ws_according_to_design()}}{
Draw multiple treatment assignment vectors according to Bernoulli randomization.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneBernoulli$draw_ws_according_to_design(r = 100)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{The number of designs to draw.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of size n x r.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneBernoulli-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneBernoulli$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
