% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_seq_one_by_one_efron.R
\name{DesignSeqOneByOneEfron}
\alias{DesignSeqOneByOneEfron}
\title{Efron's (1971) Biased Coin Sequential Design}
\description{
Efron's (1971) Biased Coin Sequential Design

Efron's (1971) Biased Coin Sequential Design
}
\details{
An R6 Class encapsulating the data and functionality for Efron's biased coin
sequential experimental design.
}
\section{Super classes}{
\\code{EDI::Design}
-> \\code{EDI::DesignSeqOneByOne}
-> \code{DesignSeqOneByOneEfron}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignSeqOneByOneEfron-new}{\code{DesignSeqOneByOneEfron$new()}}
\item \href{#method-DesignSeqOneByOneEfron-assign_wt}{\code{DesignSeqOneByOneEfron$assign_wt()}}
\item \code{DesignSeqOneByOneEfron$draw_ws_according_to_design()}
\item \href{#method-DesignSeqOneByOneEfron-clone}{\code{DesignSeqOneByOneEfron$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Design$add_all_subject_assignments()</code></li>
<li><code>EDI::Design$add_all_subject_responses()</code></li>
<li><code>EDI::Design$add_subject()</code></li>
<li><code>EDI::Design$add_subject_response()</code></li>
<li><code>EDI::Design$any_censoring()</code></li>
<li><code>EDI::Design$assert_all_responses_recorded()</code></li>
<li><code>EDI::Design$assert_all_subjects_arrived()</code></li>
<li><code>EDI::Design$assert_even_allocation()</code></li>
<li><code>EDI::Design$assert_fixed_sample()</code></li>
<li><code>EDI::Design$check_experiment_completed()</code></li>
<li><code>EDI::Design$duplicate()</code></li>
<li><code>EDI::Design$get_X()</code></li>
<li><code>EDI::Design$get_X_imp()</code></li>
<li><code>EDI::Design$get_X_raw()</code></li>
<li><code>EDI::Design$get_block_ids()</code></li>
<li><code>EDI::Design$get_dead()</code></li>
<li><code>EDI::Design$get_n()</code></li>
<li><code>EDI::Design$get_prob_T()</code></li>
<li><code>EDI::Design$get_response_type()</code></li>
<li><code>EDI::Design$get_t()</code></li>
<li><code>EDI::Design$get_w()</code></li>
<li><code>EDI::Design$get_y()</code></li>
<li><code>EDI::Design$is_fixed_sample_size()</code></li>
<li><code>EDI::Design$resample_design()</code></li>
<li><code>EDI::DesignSeqOneByOne$add_subject_to_experiment_and_assign()</code></li>
<li><code>EDI::DesignSeqOneByOne$print_current_subject_assignment()</code></li>
<li><code>EDI::DesignSeqOneByOne$supports_resampling()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneEfron-new"></a>}}
\subsection{Method \code{new()}}{
Initialize an Efron biased coin sequential experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignSeqOneByOneEfron$new(
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE,
  weighted_coin_prob = 2/3
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{"continuous", "incidence", "proportion", "count", "survival", or
"ordinal".}

\item{\code{prob_T}}{Probability of treatment assignment.}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size.}

\item{\code{num_cores}}{The number of CPU cores to use.}

\item{\code{verbose}}{A flag for verbosity.}

\item{\code{weighted_coin_prob}}{The probability of assigning to the under-represented group.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DesignSeqOneByOneEfron` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneEfron-assign_wt"></a>}}
\subsection{Method \code{assign_wt()}}{
Assign the next subject to a treatment group
\subsection{Usage}{
\preformatted{DesignSeqOneByOneEfron$assign_wt()}
}

\subsection{Returns}{
The treatment assignment (0 or 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneEfron-draw_ws_according_to_design"></a>}}
\subsection{Method \code{draw_ws_according_to_design()}}{
Draw multiple treatment assignment vectors according to Efron's design.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneEfron$draw_ws_according_to_design(r = 100)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{The number of designs to draw.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of size n x r.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneEfron-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneEfron$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
