% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_seq_one_by_one_KK14.R
\name{DesignSeqOneByOneKK14}
\alias{DesignSeqOneByOneKK14}
\title{A Sequential Design}
\description{
A Sequential Design

A Sequential Design
}
\details{
An R6 Class encapsulating the data and functionality for a sequential experimental design.
This class takes care of data initialization and sequential assignments. The class object
should be saved securely after each assignment e.g. on an encrypted cloud server.
}
\section{Super classes}{
\\code{EDI::Design}
-> \\code{EDI::DesignSeqOneByOne}
-> \code{DesignSeqOneByOneKK14}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignSeqOneByOneKK14-new}{\code{DesignSeqOneByOneKK14$new()}}
\item \href{#method-DesignSeqOneByOneKK14-assign_wt}{\code{DesignSeqOneByOneKK14$assign_wt()}}
\item \code{DesignSeqOneByOneKK14$draw_ws_according_to_design()}
\item \href{#method-DesignSeqOneByOneKK14-clone}{\code{DesignSeqOneByOneKK14$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Design$add_all_subject_assignments()</code></li>
<li><code>EDI::Design$add_all_subject_responses()</code></li>
<li><code>EDI::Design$add_subject()</code></li>
<li><code>EDI::Design$add_subject_response()</code></li>
<li><code>EDI::Design$any_censoring()</code></li>
<li><code>EDI::Design$assert_all_responses_recorded()</code></li>
<li><code>EDI::Design$assert_all_subjects_arrived()</code></li>
<li><code>EDI::Design$assert_even_allocation()</code></li>
<li><code>EDI::Design$assert_fixed_sample()</code></li>
<li><code>EDI::Design$check_experiment_completed()</code></li>
<li><code>EDI::Design$duplicate()</code></li>
<li><code>EDI::Design$get_X()</code></li>
<li><code>EDI::Design$get_X_imp()</code></li>
<li><code>EDI::Design$get_X_raw()</code></li>
<li><code>EDI::Design$get_block_ids()</code></li>
<li><code>EDI::Design$get_dead()</code></li>
<li><code>EDI::Design$get_n()</code></li>
<li><code>EDI::Design$get_prob_T()</code></li>
<li><code>EDI::Design$get_response_type()</code></li>
<li><code>EDI::Design$get_t()</code></li>
<li><code>EDI::Design$get_w()</code></li>
<li><code>EDI::Design$get_y()</code></li>
<li><code>EDI::Design$is_fixed_sample_size()</code></li>
<li><code>EDI::Design$resample_design()</code></li>
<li><code>EDI::DesignSeqOneByOne$add_subject_to_experiment_and_assign()</code></li>
<li><code>EDI::DesignSeqOneByOne$print_current_subject_assignment()</code></li>
<li><code>EDI::DesignSeqOneByOne$supports_resampling()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneKK14-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a KK14 sequential experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignSeqOneByOneKK14$new(
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE,
  lambda = NULL,
  t_0_pct = NULL,
  morrison = FALSE,
  p = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{"continuous", "incidence", "proportion", "count", "survival", or
"ordinal".}

\item{\code{prob_T}}{Probability of treatment assignment.}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size.}

\item{\code{num_cores}}{The number of CPU cores to use.}

\item{\code{verbose}}{A flag for verbosity.}

\item{\code{lambda}}{The penalty parameter for covariate imbalance.}

\item{\code{t_0_pct}}{The percentage of subjects to allocate before matching begins.}

\item{\code{morrison}}{If TRUE, use Morrison's method for matching.}

\item{\code{p}}{The number of covariates to use for matching.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DesignSeqOneByOneKK14` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneKK14-assign_wt"></a>}}
\subsection{Method \code{assign_wt()}}{
Assign the next subject to a treatment group
\subsection{Usage}{
\preformatted{DesignSeqOneByOneKK14$assign_wt()}
}

\subsection{Returns}{
The treatment assignment (0 or 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneKK14-draw_ws_according_to_design"></a>}}
\subsection{Method \code{draw_ws_according_to_design()}}{
Draw multiple treatment assignment vectors according to KK14 design.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneKK14$draw_ws_according_to_design(r = 100)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{The number of designs to draw.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of size n x r.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneKK14-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneKK14$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
