% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_seq_one_by_one_random_block_size.R
\name{DesignSeqOneByOneRandomBlockSize}
\alias{DesignSeqOneByOneRandomBlockSize}
\title{A sequential blocking allocation design with random block sizes}
\description{
A sequential blocking allocation design with random block sizes

A sequential blocking allocation design with random block sizes
}
\details{
An R6 Class encapsulating the data and functionality for a sequential blocking
experimental design
where block sizes are randomly chosen from a specified set for each stratum. This design
is commonly used in clinical trials to prevent predictability of treatment assignments.
}
\section{Super classes}{
\\code{EDI::Design}
-> \\code{EDI::DesignSeqOneByOne}
-> \code{DesignSeqOneByOneRandomBlockSize}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{DesignSeqOneByOneRandomBlockSize$new()}
\item \code{DesignSeqOneByOneRandomBlockSize$assign_wt()}
\item \code{DesignSeqOneByOneRandomBlockSize$clone()}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Design$add_all_subject_assignments()</code></li>
<li><code>EDI::Design$add_all_subject_responses()</code></li>
<li><code>EDI::Design$add_subject()</code></li>
<li><code>EDI::Design$add_subject_response()</code></li>
<li><code>EDI::Design$any_censoring()</code></li>
<li><code>EDI::Design$assert_all_responses_recorded()</code></li>
<li><code>EDI::Design$assert_all_subjects_arrived()</code></li>
<li><code>EDI::Design$assert_even_allocation()</code></li>
<li><code>EDI::Design$assert_fixed_sample()</code></li>
<li><code>EDI::Design$check_experiment_completed()</code></li>
<li><code>EDI::Design$duplicate()</code></li>
<li><code>EDI::Design$get_X()</code></li>
<li><code>EDI::Design$get_X_imp()</code></li>
<li><code>EDI::Design$get_X_raw()</code></li>
<li><code>EDI::Design$get_block_ids()</code></li>
<li><code>EDI::Design$get_dead()</code></li>
<li><code>EDI::Design$get_n()</code></li>
<li><code>EDI::Design$get_prob_T()</code></li>
<li><code>EDI::Design$get_response_type()</code></li>
<li><code>EDI::Design$get_t()</code></li>
<li><code>EDI::Design$get_w()</code></li>
<li><code>EDI::Design$get_y()</code></li>
<li><code>EDI::Design$is_fixed_sample_size()</code></li>
<li><code>EDI::Design$resample_design()</code></li>
<li><code>EDI::DesignSeqOneByOne$add_subject_to_experiment_and_assign()</code></li>
<li><code>EDI::DesignSeqOneByOne$draw_ws_according_to_design()</code></li>
<li><code>EDI::DesignSeqOneByOne$print_current_subject_assignment()</code></li>
<li><code>EDI::DesignSeqOneByOne$supports_resampling()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneRandomBlockSize-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a random block size sequential experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignSeqOneByOneRandomBlockSize$new(
  strata_cols = NULL,
  block_sizes = c(4, 6, 8),
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strata_cols}}{A character vector of column names to use for stratification. If
NULL, simple blocking is used.}

\item{\code{block_sizes}}{A vector of positive integers representing the possible block sizes
  to choose from.
Each must be a multiple of the inverse of \code{prob_T} to ensure integer
treatment/control counts.}

\item{\code{response_type}}{The data type of response values which must be one of the following:
"continuous" (the default),
"incidence",
"proportion",
"count",
"survival",
"ordinal".}

\item{\code{prob_T}}{The probability of the treatment assignment. This defaults to \code{0.5}.}

\item{\code{include_is_missing_as_a_new_feature}}{If missing data is present in a variable,
should we include another dummy variable for its missingness? Default is \code{TRUE}.}

\item{\code{n}}{The sample size (if fixed). Default is \code{NULL} for not fixed.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize the sampling.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed. Default is
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DesignSeqOneByOneRandomBlockSize` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneRandomBlockSize-assign_wt"></a>}}
\subsection{Method \code{assign_wt()}}{
Assign the next subject to a treatment group
\subsection{Usage}{
\preformatted{DesignSeqOneByOneRandomBlockSize$assign_wt()}
}

\subsection{Returns}{
The treatment assignment (0 or 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneRandomBlockSize-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneRandomBlockSize$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
