% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_seq_one_by_one_urn.R
\name{DesignSeqOneByOneUrn}
\alias{DesignSeqOneByOneUrn}
\title{Wei's Urn Sequential Design}
\description{
Wei's Urn Sequential Design

Wei's Urn Sequential Design
}
\details{
An R6 Class encapsulating the data and functionality for Wei's Urn sequential
experimental design.
This design uses an adaptive urn model where the probability of assignment to a group
decreases as the number of subjects in that group increases.
}
\section{Super classes}{
\\code{EDI::Design}
-> \\code{EDI::DesignSeqOneByOne}
-> \code{DesignSeqOneByOneUrn}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DesignSeqOneByOneUrn-new}{\code{DesignSeqOneByOneUrn$new()}}
\item \href{#method-DesignSeqOneByOneUrn-assign_wt}{\code{DesignSeqOneByOneUrn$assign_wt()}}
\item \href{#method-DesignSeqOneByOneUrn-clone}{\code{DesignSeqOneByOneUrn$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Design$add_all_subject_assignments()</code></li>
<li><code>EDI::Design$add_all_subject_responses()</code></li>
<li><code>EDI::Design$add_subject()</code></li>
<li><code>EDI::Design$add_subject_response()</code></li>
<li><code>EDI::Design$any_censoring()</code></li>
<li><code>EDI::Design$assert_all_responses_recorded()</code></li>
<li><code>EDI::Design$assert_all_subjects_arrived()</code></li>
<li><code>EDI::Design$assert_even_allocation()</code></li>
<li><code>EDI::Design$assert_fixed_sample()</code></li>
<li><code>EDI::Design$check_experiment_completed()</code></li>
<li><code>EDI::Design$duplicate()</code></li>
<li><code>EDI::Design$get_X()</code></li>
<li><code>EDI::Design$get_X_imp()</code></li>
<li><code>EDI::Design$get_X_raw()</code></li>
<li><code>EDI::Design$get_block_ids()</code></li>
<li><code>EDI::Design$get_dead()</code></li>
<li><code>EDI::Design$get_n()</code></li>
<li><code>EDI::Design$get_prob_T()</code></li>
<li><code>EDI::Design$get_response_type()</code></li>
<li><code>EDI::Design$get_t()</code></li>
<li><code>EDI::Design$get_w()</code></li>
<li><code>EDI::Design$get_y()</code></li>
<li><code>EDI::Design$is_fixed_sample_size()</code></li>
<li><code>EDI::Design$resample_design()</code></li>
<li><code>EDI::DesignSeqOneByOne$add_subject_to_experiment_and_assign()</code></li>
<li><code>EDI::DesignSeqOneByOne$draw_ws_according_to_design()</code></li>
<li><code>EDI::DesignSeqOneByOne$print_current_subject_assignment()</code></li>
<li><code>EDI::DesignSeqOneByOne$supports_resampling()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneUrn-new"></a>}}
\subsection{Method \code{new()}}{
Initialize an Urn sequential experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DesignSeqOneByOneUrn$new(
  alpha = 1,
  beta = 1,
  response_type = "continuous",
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The initial number of balls of each type (Treatment/Control) in the urn.}

\item{\code{beta}}{The number of balls of the opposite type to add to the urn after an assignment.}

\item{\code{response_type}}{The data type of response values.}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size.}

\item{\code{num_cores}}{The number of CPU cores to use.}

\item{\code{verbose}}{A flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `DesignSeqOneByOneUrn` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneUrn-assign_wt"></a>}}
\subsection{Method \code{assign_wt()}}{
Assign the next subject to a treatment group
\subsection{Usage}{
\preformatted{DesignSeqOneByOneUrn$assign_wt()}
}

\subsection{Returns}{
The treatment assignment (0 or 1)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DesignSeqOneByOneUrn-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{DesignSeqOneByOneUrn$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
