% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDI.R
\docType{package}
\name{EDI}
\alias{EDI-package}
\alias{EDI}
\title{Experimental Design and Inference}
\description{
EDI
}
\details{
Provides comprehensive support for many fixed and sequential experimental designs and many
infererential methods (parametric, nonparametric, exact) for response types continuous, 
incidence, count, proportion, survival  (with censoring) and ordinal. Supports automatic 
missing data imputation, parallelization, provides robustness fallbacks and is
optimized with C++.
}
\references{
Adam Kapelner and Abba Krieger A Matching Procedure for Sequential
  Experiments that Iteratively Learns which Covariates Improve Power, Arxiv
  2010.05980
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kapelner/matching_on_the_fly_designs_R_package_and_paper_repr}
}

}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{design}
\keyword{pair-matching}
