% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fit_helpers.R
\name{fast_logistic_regression_weighted_cpp}
\alias{fast_logistic_regression_weighted_cpp}
\title{Fast Weighted Logistic Regression (C++ Backend)}
\arguments{
\item{X}{A numeric matrix of predictor variables. It is assumed that an intercept column
(e.g., a column of ones) is already included in \code{X} if desired.}

\item{y}{A numeric vector of the response variable, expected to be binary (0 or 1).}

\item{weights}{A numeric vector of weights for each observation.}

\item{maxit}{Maximum number of iterations for the IRLS algorithm. Defaults to 100.}

\item{tol}{Convergence tolerance. Defaults to 1e-8.}
}
\value{
A list containing the following components:
\describe{
\item{b}{A numeric vector of the estimated logistic regression coefficients.}
\item{mu}{The fitted values.}
\item{XtWX}{The XtWX matrix at the final iteration.}
\item{converged}{A logical value indicating whether the IRLS algorithm converged.}
}
}
\description{
Fast Weighted Logistic Regression (C++ Backend)
}
