% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_logistic_regression_with_var_cpp}
\alias{fast_logistic_regression_with_var_cpp}
\title{Fast Logistic Regression with variance using Rcpp and IRLS}
\usage{
fast_logistic_regression_with_var_cpp(Xmm, y, j = 2L)
}
\arguments{
\item{Xmm}{Design matrix.}

\item{y}{Response vector.}

\item{j}{Index of the coefficient for which to compute the variance.}
}
\value{
A list with coefficients and specific treatment variance.
}
\description{
Fast Logistic Regression with variance using Rcpp and IRLS
}
