% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_neg_bin_with_censoring_with_sd_cpp}
\alias{fast_neg_bin_with_censoring_with_sd_cpp}
\title{Fast Negative Binomial Regression with censoring and standard deviations using Rcpp and L-BFGS}
\usage{
fast_neg_bin_with_censoring_with_sd_cpp(
  X,
  y,
  dead,
  maxit = 100L,
  eps_f = 1e-08,
  eps_g = 1e-05
)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response vector.}

\item{dead}{Censoring indicator (1=event, 0=censored).}

\item{maxit}{Maximum iterations.}

\item{eps_f}{Convergence tolerance for function value.}

\item{eps_g}{Convergence tolerance for gradient.}
}
\value{
A list with coefficients, theta, log-likelihood, and Hessian.
}
\description{
Fast Negative Binomial Regression with censoring and standard deviations using Rcpp and L-BFGS
}
