% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_neg_bin_with_var_cpp}
\alias{fast_neg_bin_with_var_cpp}
\title{Fast Negative Binomial Regression with standard deviations using Rcpp and L-BFGS}
\usage{
fast_neg_bin_with_var_cpp(X, y, maxit = 100L, eps_f = 1e-08, eps_g = 1e-05)
}
\arguments{
\item{X}{Design matrix.}

\item{y}{Response vector.}

\item{maxit}{Maximum iterations.}

\item{eps_f}{Convergence tolerance for function value.}

\item{eps_g}{Convergence tolerance for gradient.}
}
\value{
A list with coefficients, theta, log-likelihood, and Hessian.
}
\description{
Fast Negative Binomial Regression with standard deviations using Rcpp and L-BFGS
}
