% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fit_helpers.R
\name{fast_negbin_regression}
\alias{fast_negbin_regression}
\title{Fast Negative Binomial Regression (R Wrapper)}
\usage{
fast_negbin_regression(Xmm, y)
}
\arguments{
\item{Xmm}{A numeric matrix of predictor variables. It is assumed that an intercept column
(e.g., a column of ones) is already included in \code{Xmm} if desired.}

\item{y}{A numeric vector of the response variable, representing count data.}
}
\value{
A list containing the following component:
\item{b}{A numeric vector of the estimated negative binomial regression coefficients.}
}
\description{
This function provides a fast implementation of negative binomial regression by wrapping
a C++ backend function. It is designed to efficiently estimate regression coefficients
for count data, assuming all observations are "not censored" (i.e., `dead = 1`).
}
