% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fit_helpers.R
\name{fast_ols_cpp}
\alias{fast_ols_cpp}
\title{Fast Ordinary Least Squares Regression}
\arguments{
\item{X}{A numeric matrix of predictors.}

\item{y}{A numeric vector of the response variable.}
}
\value{
A list containing the following components:
\item{b}{A numeric vector of the estimated coefficients.}
\item{XtX}{The matrix X'X.}
}
\description{
This function performs a fast Ordinary Least Squares (OLS) regression using Eigen.
}
