% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_fit_helpers.R
\name{fast_quasipoisson_regression_with_var_cpp}
\alias{fast_quasipoisson_regression_with_var_cpp}
\title{Fast Quasi-Poisson Regression with Variance Calculation (C++ Backend)}
\arguments{
\item{Xmm}{A numeric matrix of predictor variables. It is assumed that an intercept column
(e.g., a column of ones) is already included in \code{Xmm} if desired.}

\item{y}{A numeric vector of the response variable, expected to be counts.}

\item{j}{The index of the coefficient to compute the variance for. Defaults to 2.}

\item{maxit}{Maximum number of iterations for the IRLS algorithm. Defaults to 100.}

\item{tol}{Convergence tolerance. Defaults to 1e-8.}
}
\value{
A list containing the following components:
\describe{
\item{b}{A numeric vector of the obtained Poisson regression coefficients.}
\item{ssq_b_j}{The squared standard error (variance) of the j-th estimated coefficient.}
\item{ssq_b_2}{The squared standard error (variance) of the second estimated coefficient.}
\item{dispersion}{The estimated dispersion parameter.}
\item{mu}{The fitted values.}
\item{converged}{A logical value indicating whether the IRLS algorithm converged.}
}
}
\description{
Fast Quasi-Poisson Regression with Variance Calculation (C++ Backend)
}
