% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_weibull_regression}
\alias{fast_weibull_regression}
\title{Fast Weibull Regression using Rcpp and L-BFGS}
\usage{
fast_weibull_regression(y, dead, X)
}
\arguments{
\item{y}{Numeric vector of survival times.}

\item{dead}{Integer vector of event indicators (1=event, 0=censored).}

\item{X}{Model matrix.}
}
\value{
A list with coefficients, log(sigma), standard errors, vcov, etc.
}
\description{
Fast Weibull Regression using Rcpp and L-BFGS
}
