% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fixed_abstract.R
\name{FixedDesign}
\alias{FixedDesign}
\title{A Fixed Design}
\description{
An abstract R6 Class encapsulating the data and functionality for a fixed experimental design.
This class takes care of whole-experiment randomization.
}
\keyword{internal}
\section{Super class}{
\code{\link[EDI:Design]{EDI::Design}} -> \code{FixedDesign}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FixedDesign-new}{\code{FixedDesign$new()}}
\item \href{#method-FixedDesign-randomize}{\code{FixedDesign$randomize()}}
\item \href{#method-FixedDesign-supports_resampling}{\code{FixedDesign$supports_resampling()}}
\item \href{#method-FixedDesign-redraw_w_according_to_design}{\code{FixedDesign$redraw_w_according_to_design()}}
\item \href{#method-FixedDesign-draw_ws_according_to_design}{\code{FixedDesign$draw_ws_according_to_design()}}
\item \href{#method-FixedDesign-clone}{\code{FixedDesign$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="add_all_subject_assignments"><a href='../../EDI/html/Design.html#method-Design-add_all_subject_assignments'><code>EDI::Design$add_all_subject_assignments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="add_all_subject_responses"><a href='../../EDI/html/Design.html#method-Design-add_all_subject_responses'><code>EDI::Design$add_all_subject_responses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="add_subject"><a href='../../EDI/html/Design.html#method-Design-add_subject'><code>EDI::Design$add_subject()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="add_subject_response"><a href='../../EDI/html/Design.html#method-Design-add_subject_response'><code>EDI::Design$add_subject_response()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="any_censoring"><a href='../../EDI/html/Design.html#method-Design-any_censoring'><code>EDI::Design$any_censoring()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_all_subjects_arrived"><a href='../../EDI/html/Design.html#method-Design-assert_all_subjects_arrived'><code>EDI::Design$assert_all_subjects_arrived()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_even_allocation"><a href='../../EDI/html/Design.html#method-Design-assert_even_allocation'><code>EDI::Design$assert_even_allocation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_experiment_completed"><a href='../../EDI/html/Design.html#method-Design-assert_experiment_completed'><code>EDI::Design$assert_experiment_completed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_fixed_sample"><a href='../../EDI/html/Design.html#method-Design-assert_fixed_sample'><code>EDI::Design$assert_fixed_sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="check_experiment_completed"><a href='../../EDI/html/Design.html#method-Design-check_experiment_completed'><code>EDI::Design$check_experiment_completed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="duplicate"><a href='../../EDI/html/Design.html#method-Design-duplicate'><code>EDI::Design$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_X"><a href='../../EDI/html/Design.html#method-Design-get_X'><code>EDI::Design$get_X()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_X_imp"><a href='../../EDI/html/Design.html#method-Design-get_X_imp'><code>EDI::Design$get_X_imp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_X_raw"><a href='../../EDI/html/Design.html#method-Design-get_X_raw'><code>EDI::Design$get_X_raw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_dead"><a href='../../EDI/html/Design.html#method-Design-get_dead'><code>EDI::Design$get_dead()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_n"><a href='../../EDI/html/Design.html#method-Design-get_n'><code>EDI::Design$get_n()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_prob_T"><a href='../../EDI/html/Design.html#method-Design-get_prob_T'><code>EDI::Design$get_prob_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_response_type"><a href='../../EDI/html/Design.html#method-Design-get_response_type'><code>EDI::Design$get_response_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_t"><a href='../../EDI/html/Design.html#method-Design-get_t'><code>EDI::Design$get_t()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_w"><a href='../../EDI/html/Design.html#method-Design-get_w'><code>EDI::Design$get_w()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_y"><a href='../../EDI/html/Design.html#method-Design-get_y'><code>EDI::Design$get_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="is_fixed_sample_size"><a href='../../EDI/html/Design.html#method-Design-is_fixed_sample_size'><code>EDI::Design$is_fixed_sample_size()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesign-new"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesign-new}{}}}
\subsection{Method \code{new()}}{
Initialize a fixed experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesign$new(
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{"continuous", "incidence", "proportion", "count", "survival", or "ordinal".}

\item{\code{prob_T}}{Probability of treatment assignment.}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size.}

\item{\code{num_cores}}{The number of CPU cores.}

\item{\code{verbose}}{A flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FixedDesign` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesign-randomize"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesign-randomize}{}}}
\subsection{Method \code{randomize()}}{
Randomize treatment assignments for the entire experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesign$randomize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesign-supports_resampling"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesign-supports_resampling}{}}}
\subsection{Method \code{supports_resampling()}}{
Check if the design supports resampling.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesign$supports_resampling()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if supported.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesign-redraw_w_according_to_design"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesign-redraw_w_according_to_design}{}}}
\subsection{Method \code{redraw_w_according_to_design()}}{
Redraw treatment assignments according to the design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesign$redraw_w_according_to_design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesign-draw_ws_according_to_design"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesign-draw_ws_according_to_design}{}}}
\subsection{Method \code{draw_ws_according_to_design()}}{
Draw multiple treatment assignment vectors.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesign$draw_ws_according_to_design(r = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{The number of designs to draw.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of size n x r.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesign-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesign-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesign$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
