% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fixed_a_optimal.R
\name{FixedDesignAOptimal}
\alias{FixedDesignAOptimal}
\title{An A-optimal Search Fixed Design}
\description{
An A-optimal Search Fixed Design

An A-optimal Search Fixed Design
}
\details{
An R6 Class encapsulating the data and functionality for a fixed A-optimal experimental design.
This design searches for an allocation that minimizes the trace of the inverse
information matrix
(equivalent to minimizing the average variance of the parameter estimates).
}
\section{Super classes}{
\\code{EDI::Design}
-> \\code{EDI::FixedDesign}
-> \code{FixedDesignAOptimal}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FixedDesignAOptimal-new}{\code{FixedDesignAOptimal$new()}}
\item \code{FixedDesignAOptimal$draw_ws_according_to_design()}
\item \href{#method-FixedDesignAOptimal-clone}{\code{FixedDesignAOptimal$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Design$add_all_subject_assignments()</code></li>
<li><code>EDI::Design$add_all_subject_responses()</code></li>
<li><code>EDI::Design$add_subject()</code></li>
<li><code>EDI::Design$add_subject_response()</code></li>
<li><code>EDI::Design$any_censoring()</code></li>
<li><code>EDI::Design$assert_all_responses_recorded()</code></li>
<li><code>EDI::Design$assert_all_subjects_arrived()</code></li>
<li><code>EDI::Design$assert_even_allocation()</code></li>
<li><code>EDI::Design$assert_fixed_sample()</code></li>
<li><code>EDI::Design$check_experiment_completed()</code></li>
<li><code>EDI::Design$duplicate()</code></li>
<li><code>EDI::Design$get_X()</code></li>
<li><code>EDI::Design$get_X_imp()</code></li>
<li><code>EDI::Design$get_X_raw()</code></li>
<li><code>EDI::Design$get_block_ids()</code></li>
<li><code>EDI::Design$get_dead()</code></li>
<li><code>EDI::Design$get_n()</code></li>
<li><code>EDI::Design$get_prob_T()</code></li>
<li><code>EDI::Design$get_response_type()</code></li>
<li><code>EDI::Design$get_t()</code></li>
<li><code>EDI::Design$get_w()</code></li>
<li><code>EDI::Design$get_y()</code></li>
<li><code>EDI::Design$is_fixed_sample_size()</code></li>
<li><code>EDI::Design$resample_design()</code></li>
<li><code>EDI::FixedDesign$randomize()</code></li>
<li><code>EDI::FixedDesign$supports_resampling()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignAOptimal-new"></a>}}
\subsection{Method \code{new()}}{
Initialize an A-optimal search fixed experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignAOptimal$new(
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{"continuous", "incidence", "proportion", "count", "survival", or
"ordinal".}

\item{\code{prob_T}}{Probability of treatment assignment (default 0.5).}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{Sample size (if fixed).}

\item{\code{num_cores}}{Number of CPU cores.}

\item{\code{verbose}}{Flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FixedDesignAOptimal` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignAOptimal-draw_ws_according_to_design"></a>}}
\subsection{Method \code{draw_ws_according_to_design()}}{
Draw treatment assignments according to the A-optimal search fixed design.
\subsection{Usage}{
\preformatted{FixedDesignAOptimal$draw_ws_according_to_design(r = 100)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{Number of designs to draw.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of size n x r.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignAOptimal-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{FixedDesignAOptimal$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
