% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_restricted_mean_se_diff}
\alias{get_restricted_mean_se_diff}
\title{Calculates the standard error of the difference in restricted mean survival times.}
\usage{
get_restricted_mean_se_diff(y, dead, w)
}
\arguments{
\item{y}{Numeric vector of survival times.}

\item{dead}{Integer vector of event indicators (1=event, 0=censored).}

\item{w}{Integer vector of treatment assignments (1=treatment, 0=control).}
}
\value{
The standard error of the difference.
}
\description{
Calculates the standard error of the difference in restricted mean survival times.
}
\keyword{internal}
