% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_survival_stat_diff}
\alias{get_survival_stat_diff}
\title{Calculates the difference in a survival statistic (median or restricted mean)
between two groups (treatment vs. control).}
\usage{
get_survival_stat_diff(y, dead, w, requested_stat)
}
\arguments{
\item{y}{Numeric vector of survival times.}

\item{dead}{Integer vector of event indicators (1=event, 0=censored).}

\item{w}{Integer vector of treatment assignments (1=treatment, 0=control).}

\item{requested_stat}{A string, either "median" or "restricted_mean".}
}
\value{
The difference in the statistic (treatment - control).
}
\description{
Calculates the difference in a survival statistic (median or restricted mean)
between two groups (treatment vs. control).
}
\keyword{internal}
