% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_survival_KK_clayton_copula_combined_likelihood_abstract.R
\name{InferenceAbstractKKClaytonCopulaCombinedLikelihood}
\alias{InferenceAbstractKKClaytonCopulaCombinedLikelihood}
\title{Abstract class for Clayton Copula Combined-Likelihood Inference}
\description{
Abstract class for Clayton Copula Combined-Likelihood Inference

Abstract class for Clayton Copula Combined-Likelihood Inference
}
\details{
Fits a single joint likelihood over all KK design data for survival responses using
a Clayton survival copula with Weibull AFT margins for matched pairs and standard
Weibull AFT singleton contributions for reservoir subjects. The treatment effect is
reported on the AFT log-time-ratio scale.


This keeps the package's established \code{CombinedLikelihood} terminology even though
the matched and reservoir parts are combined through a joint copula-based likelihood
rather than a classical Gaussian-style likelihood. The matched-pair contribution uses
a Clayton copula with censored bivariate survival contributions; reservoir subjects
contribute ordinary univariate Weibull AFT terms.
}
\keyword{internal}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{InferenceAbstractKKClaytonCopulaCombinedLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-new}{\code{InferenceAbstractKKClaytonCopulaCombinedLikelihood$new()}}
\item \href{#method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_treatment_estimate}{\code{InferenceAbstractKKClaytonCopulaCombinedLikelihood$compute_treatment_estimate()}}
\item \href{#method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_asymp_confidence_interval}{\code{InferenceAbstractKKClaytonCopulaCombinedLikelihood$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceAbstractKKClaytonCopulaCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-duplicate}{\code{InferenceAbstractKKClaytonCopulaCombinedLikelihood$duplicate()}}
\item \href{#method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-clone}{\code{InferenceAbstractKKClaytonCopulaCombinedLikelihood$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClaytonCopulaCombinedLikelihood$new(
  des_obj,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClaytonCopulaCombinedLikelihood$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClaytonCopulaCombinedLikelihood$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClaytonCopulaCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-duplicate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-duplicate}{}}}
\subsection{Method \code{duplicate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClaytonCopulaCombinedLikelihood$duplicate(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClaytonCopulaCombinedLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClaytonCopulaCombinedLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
