% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_incidence_KK_modified_poisson_abstract.R
\name{InferenceAbstractKKModifiedPoisson}
\alias{InferenceAbstractKKModifiedPoisson}
\title{Abstract class for all-subject modified-Poisson inference in KK designs}
\description{
Abstract class for all-subject modified-Poisson inference in KK designs

Abstract class for all-subject modified-Poisson inference in KK designs
}
\keyword{internal}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \\code{EDI::InferenceKKPassThrough}
-> \\code{EDI::InferenceAbstractKKMarginalIncid}
-> \code{InferenceAbstractKKModifiedPoisson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceAbstractKKModifiedPoisson$compute_treatment_estimate()}
\item \code{InferenceAbstractKKModifiedPoisson$compute_asymp_confidence_interval()}
\item \code{InferenceAbstractKKModifiedPoisson$compute_asymp_two_sided_pval_for_treatment_effect()}
\item \code{InferenceAbstractKKModifiedPoisson$clone()}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceAbstractKKMarginalIncid$initialize()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKModifiedPoisson-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Compute treatment estimate
\subsection{Usage}{
\preformatted{InferenceAbstractKKModifiedPoisson$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Compute asymp confidence interval
\subsection{Usage}{
\preformatted{InferenceAbstractKKModifiedPoisson$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Description for alpha}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Compute asymp two sided pval for treatment effect
\subsection{Usage}{
\preformatted{InferenceAbstractKKModifiedPoisson$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{Description for delta}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKModifiedPoisson-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceAbstractKKModifiedPoisson$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
