% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_continuous_KK_robust_regr_ivwc_abstract.R
\name{InferenceAbstractKKRobustRegrIVWC}
\alias{InferenceAbstractKKRobustRegrIVWC}
\title{Abstract class for Robust-Regression IVWC Compound Inference for KK Designs}
\description{
Fits a variance-weighted compound estimator for KK matching-on-the-fly designs
with continuous responses using robust linear regression (`MASS::rlm`) for the
matched-pair and reservoir components separately.
}
\keyword{internal}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{\link[EDI:InferenceKKPassThrough]{EDI::InferenceKKPassThrough}} -> \code{EDI::InferenceKKPassThroughCompound} -> \code{InferenceAbstractKKRobustRegrIVWC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceAbstractKKRobustRegrIVWC-new}{\code{InferenceAbstractKKRobustRegrIVWC$new()}}
\item \href{#method-InferenceAbstractKKRobustRegrIVWC-compute_treatment_estimate}{\code{InferenceAbstractKKRobustRegrIVWC$compute_treatment_estimate()}}
\item \href{#method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_confidence_interval}{\code{InferenceAbstractKKRobustRegrIVWC$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceAbstractKKRobustRegrIVWC$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceAbstractKKRobustRegrIVWC-duplicate}{\code{InferenceAbstractKKRobustRegrIVWC$duplicate()}}
\item \href{#method-InferenceAbstractKKRobustRegrIVWC-clone}{\code{InferenceAbstractKKRobustRegrIVWC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKRobustRegrIVWC-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKRobustRegrIVWC-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKRobustRegrIVWC$new(
  des_obj,
  method = "MM",
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKRobustRegrIVWC-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKRobustRegrIVWC-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKRobustRegrIVWC$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKRobustRegrIVWC$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKRobustRegrIVWC$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKRobustRegrIVWC-duplicate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKRobustRegrIVWC-duplicate}{}}}
\subsection{Method \code{duplicate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKRobustRegrIVWC$duplicate(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKRobustRegrIVWC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKRobustRegrIVWC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKRobustRegrIVWC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
