% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_survival_KK_strat_cox_combined_likelihood_abstract.R
\name{InferenceAbstractKKStratCoxCombinedLikelihood}
\alias{InferenceAbstractKKStratCoxCombinedLikelihood}
\title{Abstract class for Stratified Cox Combined-Likelihood Compound Inference}
\description{
Fits a single joint Cox model over all KK design data for survival responses.
Reservoir subjects (m_vec == 0) share stratum 0 (common baseline hazard);
matched subjects belong to their own pair stratum (strata 1..m), which conditions
away each pair's baseline hazard identically to stratified Cox.

The combined partial likelihood is therefore:
L = L_strat_pairs(beta_T, beta_xs) x L_cox_reservoir(beta_T, beta_xs)
and is maximised by a single coxph call with strata(stratum).
}
\keyword{internal}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{EDI::InferenceKKPassThrough} -> \code{InferenceAbstractKKStratCoxCombinedLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceAbstractKKStratCoxCombinedLikelihood-new}{\code{InferenceAbstractKKStratCoxCombinedLikelihood$new()}}
\item \href{#method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_treatment_estimate}{\code{InferenceAbstractKKStratCoxCombinedLikelihood$compute_treatment_estimate()}}
\item \href{#method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_asymp_confidence_interval}{\code{InferenceAbstractKKStratCoxCombinedLikelihood$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceAbstractKKStratCoxCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceAbstractKKStratCoxCombinedLikelihood-clone}{\code{InferenceAbstractKKStratCoxCombinedLikelihood$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKStratCoxCombinedLikelihood-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKStratCoxCombinedLikelihood-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKStratCoxCombinedLikelihood$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKStratCoxCombinedLikelihood$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKStratCoxCombinedLikelihood$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKStratCoxCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKStratCoxCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKStratCoxCombinedLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKStratCoxCombinedLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKStratCoxCombinedLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
