% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_KK_compound_mean_diff.R
\name{InferenceAllKKCompoundMeanDiff}
\alias{InferenceAllKKCompoundMeanDiff}
\title{Inference based on Maximum Likelihood for KK designs}
\description{
Inference based on Maximum Likelihood for KK designs

Inference based on Maximum Likelihood for KK designs
}
\details{
Inference for mean difference
}
\examples{
\dontrun{
seq_des = DesignSeqOneByOneKK14$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceAllKKCompoundMeanDiff$
  new(seq_des)
seq_des_inf$compute_treatment_estimate()
seq_des_inf$compute_asymp_confidence_interval()
seq_des_inf$compute_asymp_two_sided_pval_for_treatment_effect()
}

}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \\code{EDI::InferenceKKPassThrough}
-> \\code{EDI::InferenceKKPassThroughCompound}
-> \code{InferenceAllKKCompoundMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceAllKKCompoundMeanDiff$compute_treatment_estimate()}
\item \code{InferenceAllKKCompoundMeanDiff$compute_asymp_confidence_interval()}
\item \code{InferenceAllKKCompoundMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect()}
\item \code{InferenceAllKKCompoundMeanDiff$compute_confidence_interval_rand()}
\item \code{InferenceAllKKCompoundMeanDiff$clone()}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceKKPassThrough$initialize()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllKKCompoundMeanDiff-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the appropriate estimate for compound mean difference across pairs and reservoir
\subsection{Usage}{
\preformatted{InferenceAllKKCompoundMeanDiff$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The setting-appropriate (see description) numeric estimate of the treatment effect
}
}
\if{html}{\out{<hr>}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1-alpha level frequentist confidence interval

Here we use the theory that MLE's computed for GLM's are asymptotically normal (except in
the case
of estimat_type "median difference" where a nonparametric bootstrap confidence interval
(see the \code{controlTest::quantileControlTest} method)
is employed. Hence these confidence intervals are asymptotically valid and thus approximate
for any sample size.
\subsection{Usage}{
\preformatted{InferenceAllKKCompoundMeanDiff$compute_asymp_confidence_interval(alpha = 0.05)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A (1 - alpha)-sized frequentist confidence interval for the treatment effect
}
}
\if{html}{\out{<hr>}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a 2-sided p-value
\subsection{Usage}{
\preformatted{InferenceAllKKCompoundMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null difference to test against. For any treatment effect at all this is
set to zero (the default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The approximate frequentist p-value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllKKCompoundMeanDiff-compute_confidence_interval_rand"></a>}}
\subsection{Method \code{compute_confidence_interval_rand()}}{
Computes a 1-alpha level frequentist confidence interval for the randomization test
\subsection{Usage}{
\preformatted{InferenceAllKKCompoundMeanDiff$compute_confidence_interval_rand(
  alpha = 0.05,
  r = 501,
  pval_epsilon = 0.005,
  show_progress = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}

\item{\code{r}}{The number of randomization vectors. The default is 501.}

\item{\code{pval_epsilon}}{The bisection algorithm tolerance. The default is 0.005.}

\item{\code{show_progress}}{Show a text progress indicator.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 1 - alpha sized frequentist confidence interval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllKKCompoundMeanDiff-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceAllKKCompoundMeanDiff$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
