% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract_asymp.R
\name{InferenceAsymp}
\alias{InferenceAsymp}
\title{Asymptotic Inference}
\description{
Asymptotic Inference

Asymptotic Inference
}
\details{
Abstract class for asymptotic inference.
}
\keyword{internal}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \code{InferenceAsymp}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceAsymp$compute_asymp_confidence_interval()}
\item \code{InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()}
\item \code{InferenceAsymp$compute_treatment_estimate()}
\item \href{#method-InferenceAsymp-clone}{\code{InferenceAsymp$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::Inference$initialize()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAsymp-compute_asymp_confidence_interval"></a>}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes an asymptotic confidence interval based on Wald-type tests.
\subsection{Usage}{
\preformatted{InferenceAsymp$compute_asymp_confidence_interval(alpha = 0.05)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Significance level 1 - \code{alpha}. Default 0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A Wald-type confidence interval.
}
}
\if{html}{\out{<hr>}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes an asymptotic two-sided p-value for the treatment effect.
\subsection{Usage}{
\preformatted{InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect(delta = 0)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{Null treatment effect to test against. Default 0.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The asymptotic p-value.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAsymp-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Abstract method to compute the treatment estimate.
\subsection{Usage}{
\preformatted{InferenceAsymp$compute_treatment_estimate(estimate_only = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A scalar treatment estimate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAsymp-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceAsymp$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
