% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_continuous_KK21_bai.R
\name{InferenceBaiAdjustedTKK21}
\alias{InferenceBaiAdjustedTKK21}
\title{Inference based on Maximum Likelihood for KK designs}
\description{
Inference for mean difference
}
\examples{

## ------------------------------------------------
## Method `InferenceBaiAdjustedTKK21$new`
## ------------------------------------------------

set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneKK21$new(n = nrow(x_dat), response_type = "continuous", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(c(1.2, 0.9, 1.5, 1.8, 2.1, 1.7, 2.6, 2.2))
infer <- InferenceBaiAdjustedTKK21$new(seq_des, verbose = FALSE)
infer

}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{EDI::InferenceBoot} -> \code{EDI::InferenceAsymp} -> \code{EDI::InferenceKKPassThrough} -> \code{EDI::InferenceKKPassThroughCompound} -> \code{EDI::InferenceBaiAdjustedT} -> \code{InferenceBaiAdjustedTKK21}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceBaiAdjustedTKK21-new}{\code{InferenceBaiAdjustedTKK21$new()}}
\item \href{#method-InferenceBaiAdjustedTKK21-clone}{\code{InferenceBaiAdjustedTKK21$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBaiAdjustedT" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceBaiAdjustedT.html#method-InferenceBaiAdjustedT-compute_asymp_confidence_interval'><code>EDI::InferenceBaiAdjustedT$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBaiAdjustedT" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceBaiAdjustedT.html#method-InferenceBaiAdjustedT-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceBaiAdjustedT$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBaiAdjustedT" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceBaiAdjustedT.html#method-InferenceBaiAdjustedT-compute_treatment_estimate'><code>EDI::InferenceBaiAdjustedT$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceBaiAdjustedTKK21-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceBaiAdjustedTKK21-new}{}}}
\subsection{Method \code{new()}}{
Initialize a sequential experimental design estimation and test object
after the sequential design is completed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceBaiAdjustedTKK21$new(
  des_obj,
  num_cores = 1,
  verbose = TRUE,
  convex_flag = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A DesignSeqOneByOne object whose entire n subjects
are assigned and response y is recorded within.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize
the sampling during randomization-based inference and
bootstrap resampling.
The default is 1 for serial computation. For simple
estimators (e.g. mean difference and KK compound),
parallelization is achieved with zero-overhead C++ OpenMP.
For complex models (e.g. GLMs),
parallelization falls back to R's
\code{parallel::mclapply}, which incurs
session-forking overhead.}

\item{\code{verbose}}{A flag indicating whether messages should be
displayed to the user. Default is \code{TRUE}.}

\item{\code{convex_flag}}{A flag indicating whether the estimator should use a convex
  combination of the Bai et al
                                        matched pairs estimate with the reservoir estimate,
or just the Bai et al estimate by its self.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneKK21$new(n = nrow(x_dat), response_type = "continuous", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(c(1.2, 0.9, 1.5, 1.8, 2.1, 1.7, 2.6, 2.2))
infer <- InferenceBaiAdjustedTKK21$new(seq_des, verbose = FALSE)
infer

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceBaiAdjustedTKK21-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceBaiAdjustedTKK21-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceBaiAdjustedTKK21$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
