% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract_boot.R
\name{InferenceBoot}
\alias{InferenceBoot}
\title{Bootstrap-based Inference}
\description{
Bootstrap-based Inference

Bootstrap-based Inference
}
\details{
Abstract class for bootstrap-based inference.
}
\keyword{internal}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{InferenceBoot}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T}{\code{InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()}}
\item \href{#method-InferenceBoot-compute_bootstrap_two_sided_pval}{\code{InferenceBoot$compute_bootstrap_two_sided_pval()}}
\item \href{#method-InferenceBoot-compute_bootstrap_confidence_interval}{\code{InferenceBoot$compute_bootstrap_confidence_interval()}}
\item \href{#method-InferenceBoot-clone}{\code{InferenceBoot$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_treatment_estimate"><a href='../..//html/Inference.html#method-Inference-compute_treatment_estimate'><code>::Inference$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="initialize"><a href='../..//html/Inference.html#method-Inference-initialize'><code>::Inference$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T}{}}}
\subsection{Method \code{approximate_bootstrap_distribution_beta_hat_T()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceBoot$approximate_bootstrap_distribution_beta_hat_T(
  B = 501,
  show_progress = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceBoot-compute_bootstrap_two_sided_pval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceBoot-compute_bootstrap_two_sided_pval}{}}}
\subsection{Method \code{compute_bootstrap_two_sided_pval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceBoot$compute_bootstrap_two_sided_pval(
  delta = 0,
  B = 501,
  na.rm = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceBoot-compute_bootstrap_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceBoot-compute_bootstrap_confidence_interval}{}}}
\subsection{Method \code{compute_bootstrap_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceBoot$compute_bootstrap_confidence_interval(
  alpha = 0.05,
  B = 501,
  type = "percentile",
  na.rm = TRUE,
  show_progress = TRUE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceBoot-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceBoot-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceBoot$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
