% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_continuous_KK_gls_multi.R
\name{InferenceContinMultGLS}
\alias{InferenceContinMultGLS}
\title{Generalized Least Squares Inference based on Maximum Likelihood for KK designs}
\description{
Generalized Least Squares Inference based on Maximum Likelihood for KK designs

Generalized Least Squares Inference based on Maximum Likelihood for KK designs
}
\details{
The methods that support confidence intervals and testing for the mean difference
in continuous response types for sequential experimental designs using Generalized Least
Squares (GLS).
This model assumes a shared correlation within matched pairs but independence for subjects in
the reservoir.
When no matched pairs exist (pure reservoir design), the model degenerates to OLS.


The GLS model is fit by REML, which provides unbiased estimates of the within-pair
correlation
\eqn{\rho} and the residual variance \eqn{\sigma^2}. The residual degrees of freedom for
t-based inference on the treatment effect are set to \eqn{n - p - 1}, where \eqn{p} is the
number of fixed-effect parameters (intercept + treatment + covariates) and the extra \eqn{-1}
accounts for the estimated \eqn{\rho}. Treating \eqn{\rho} as known would yield \eqn{n - p},
which slightly overstates precision; the conservative \eqn{n - p - 1} correction is used
instead.
When the model falls back to OLS (no matched pairs), the standard OLS residual df \eqn{n - p}
is used since no correlation parameter is estimated.

This class requires the \pkg{nlme} package, which is listed in Suggests and is not
installed automatically with \pkg{SeqExpMatch}.
Install \pkg{nlme} before using this class.
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{::InferenceKKPassThrough} -> \code{InferenceContinMultGLS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceContinMultGLS-new}{\code{InferenceContinMultGLS$new()}}
\item \href{#method-InferenceContinMultGLS-compute_treatment_estimate}{\code{InferenceContinMultGLS$compute_treatment_estimate()}}
\item \href{#method-InferenceContinMultGLS-compute_asymp_confidence_interval}{\code{InferenceContinMultGLS$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceContinMultGLS-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceContinMultGLS$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceContinMultGLS-clone}{\code{InferenceContinMultGLS$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultGLS-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultGLS-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultGLS$new(des_obj, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultGLS-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultGLS-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultGLS$compute_treatment_estimate(estimate_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultGLS-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultGLS-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultGLS$compute_asymp_confidence_interval(alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultGLS-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultGLS-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultGLS$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultGLS-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultGLS-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultGLS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
