% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_continuous_KK_lin_combined_likelihood_multi.R
\name{InferenceContinMultiKKLinCombinedLikelihood}
\alias{InferenceContinMultiKKLinCombinedLikelihood}
\title{Lin (2013) Combined-Likelihood Inference for KK Designs with Continuous Responses}
\description{
Lin (2013) Combined-Likelihood Inference for KK Designs with Continuous Responses

Lin (2013) Combined-Likelihood Inference for KK Designs with Continuous Responses
}
\details{
Fits a stacked Lin (2013) style covariate-adjusted estimator for KK
matching-on-the-fly designs with continuous responses using both matched pairs
and reservoir observations in one combined fit.


The class name uses `CombinedLikelihood` to match the package terminology used
for other KK stacked estimators. Statistically, this implementation uses a
single stacked OLS fit with HC2 heteroskedasticity-robust standard errors,
not a literal homoscedastic Gaussian likelihood.

The stacked parameterization uses:
\itemize{
  \item reservoir rows with columns \eqn{(\alpha, \tau, \eta, \gamma)}
        corresponding to \eqn{1}, treatment, centered covariates, and
        \eqn{(w - 1/2) x_c};
  \item matched-pair rows with columns \eqn{(0, 1, \Delta x, \bar x_c)},
        scaled by \eqn{1 / \sqrt{2}} to parallel the existing KK combined OLS
        weighting.
}
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{::InferenceKKPassThrough} -> \code{::InferenceKKPassThroughCompound} -> \code{InferenceContinMultiKKLinCombinedLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceContinMultiKKLinCombinedLikelihood-new}{\code{InferenceContinMultiKKLinCombinedLikelihood$new()}}
\item \href{#method-InferenceContinMultiKKLinCombinedLikelihood-compute_treatment_estimate}{\code{InferenceContinMultiKKLinCombinedLikelihood$compute_treatment_estimate()}}
\item \href{#method-InferenceContinMultiKKLinCombinedLikelihood-compute_asymp_confidence_interval}{\code{InferenceContinMultiKKLinCombinedLikelihood$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceContinMultiKKLinCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceContinMultiKKLinCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceContinMultiKKLinCombinedLikelihood-clone}{\code{InferenceContinMultiKKLinCombinedLikelihood$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinCombinedLikelihood-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinCombinedLikelihood-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinCombinedLikelihood$new(des_obj, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinCombinedLikelihood-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinCombinedLikelihood-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinCombinedLikelihood$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinCombinedLikelihood-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinCombinedLikelihood-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinCombinedLikelihood$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinCombinedLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinCombinedLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinCombinedLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
