% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_continuous_KK_lin_ivwc_multi.R
\name{InferenceContinMultiKKLinIVWC}
\alias{InferenceContinMultiKKLinIVWC}
\title{Lin (2013) IVWC Inference for KK Designs with Continuous Responses}
\description{
Lin (2013) IVWC Inference for KK Designs with Continuous Responses

Lin (2013) IVWC Inference for KK Designs with Continuous Responses
}
\details{
Fits a Lin (2013) style covariate-adjusted estimator for KK
matching-on-the-fly designs with continuous responses by combining matched-pair
and reservoir estimators via inverse-variance weighting.

The reservoir component uses the usual Lin regression with centered covariates
and treatment-by-centered-covariate interactions. The matched-pair component
uses pair differences together with pair covariate differences and centered pair
means, which is the algebraically equivalent paired Lin regression.
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{::InferenceKKPassThrough} -> \code{::InferenceKKPassThroughCompound} -> \code{InferenceContinMultiKKLinIVWC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceContinMultiKKLinIVWC-new}{\code{InferenceContinMultiKKLinIVWC$new()}}
\item \href{#method-InferenceContinMultiKKLinIVWC-compute_treatment_estimate}{\code{InferenceContinMultiKKLinIVWC$compute_treatment_estimate()}}
\item \href{#method-InferenceContinMultiKKLinIVWC-compute_asymp_confidence_interval}{\code{InferenceContinMultiKKLinIVWC$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceContinMultiKKLinIVWC-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceContinMultiKKLinIVWC$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceContinMultiKKLinIVWC-clone}{\code{InferenceContinMultiKKLinIVWC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinIVWC-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinIVWC-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinIVWC$new(des_obj, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinIVWC-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinIVWC-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinIVWC$compute_treatment_estimate(estimate_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinIVWC-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinIVWC-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinIVWC$compute_asymp_confidence_interval(alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinIVWC-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinIVWC-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinIVWC$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKLinIVWC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKLinIVWC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKLinIVWC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
