% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_continuous_KK_robust_regr_ivwc_multi.R
\name{InferenceContinMultiKKRobustRegrIVWC}
\alias{InferenceContinMultiKKRobustRegrIVWC}
\title{Multivariate Robust-Regression IVWC Compound Inference for KK Designs}
\description{
Fits a variance-weighted compound estimator for KK matching-on-the-fly designs
with continuous responses using robust regression for matched-pair differences
and reservoir outcomes, with treatment and all recorded covariates as predictors.
}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{\link[EDI:InferenceKKPassThrough]{EDI::InferenceKKPassThrough}} -> \code{\link[EDI:InferenceKKPassThroughCompound]{EDI::InferenceKKPassThroughCompound}} -> \code{\link[EDI:InferenceAbstractKKRobustRegrIVWC]{EDI::InferenceAbstractKKRobustRegrIVWC}} -> \code{InferenceContinMultiKKRobustRegrIVWC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceContinMultiKKRobustRegrIVWC-clone}{\code{InferenceContinMultiKKRobustRegrIVWC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>EDI::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKRobustRegrIVWC" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceAbstractKKRobustRegrIVWC.html#method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_confidence_interval'><code>EDI::InferenceAbstractKKRobustRegrIVWC$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKRobustRegrIVWC" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceAbstractKKRobustRegrIVWC.html#method-InferenceAbstractKKRobustRegrIVWC-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceAbstractKKRobustRegrIVWC$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKRobustRegrIVWC" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceAbstractKKRobustRegrIVWC.html#method-InferenceAbstractKKRobustRegrIVWC-compute_treatment_estimate'><code>EDI::InferenceAbstractKKRobustRegrIVWC$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKRobustRegrIVWC" data-id="duplicate"><a href='../../EDI/html/InferenceAbstractKKRobustRegrIVWC.html#method-InferenceAbstractKKRobustRegrIVWC-duplicate'><code>EDI::InferenceAbstractKKRobustRegrIVWC$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKRobustRegrIVWC" data-id="initialize"><a href='../../EDI/html/InferenceAbstractKKRobustRegrIVWC.html#method-InferenceAbstractKKRobustRegrIVWC-initialize'><code>EDI::InferenceAbstractKKRobustRegrIVWC$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiKKRobustRegrIVWC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiKKRobustRegrIVWC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiKKRobustRegrIVWC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
