% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_continuous_KK_quantile_regr_combined_likelihood.R
\name{InferenceContinMultKKQuantileRegrCombinedLikelihood}
\alias{InferenceContinMultKKQuantileRegrCombinedLikelihood}
\title{Quantile Regression Combined-Likelihood Compound Estimator for KK Designs (Continuous)}
\description{
Quantile Regression Combined-Likelihood Compound Estimator for KK Designs (Continuous)

Quantile Regression Combined-Likelihood Compound Estimator for KK Designs (Continuous)
}
\details{
Fits the combined stacked quantile regression (matched-pair differences + reservoir)
using the treatment indicator and all recorded covariates for continuous responses.
Minimises the joint check-function loss over both data sources simultaneously.
Inference is based on the stacked combined-likelihood quantile-regression fit.
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{::InferenceKKPassThrough} -> \code{::InferenceKKPassThroughCompound} -> \code{::InferenceAbstractQuantileRandCI} -> \code{::InferenceAbstractKKQuantileRegrCombinedLikelihood} -> \code{InferenceContinMultKKQuantileRegrCombinedLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceContinMultKKQuantileRegrCombinedLikelihood-new}{\code{InferenceContinMultKKQuantileRegrCombinedLikelihood$new()}}
\item \href{#method-InferenceContinMultKKQuantileRegrCombinedLikelihood-compute_treatment_estimate}{\code{InferenceContinMultKKQuantileRegrCombinedLikelihood$compute_treatment_estimate()}}
\item \href{#method-InferenceContinMultKKQuantileRegrCombinedLikelihood-clone}{\code{InferenceContinMultKKQuantileRegrCombinedLikelihood$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractQuantileRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceAbstractQuantileRandCI.html#method-InferenceAbstractQuantileRandCI-compute_confidence_interval_rand'><code>::InferenceAbstractQuantileRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractKKQuantileRegrCombinedLikelihood" data-id="compute_asymp_confidence_interval"><a href='../..//html/InferenceAbstractKKQuantileRegrCombinedLikelihood.html#method-InferenceAbstractKKQuantileRegrCombinedLikelihood-compute_asymp_confidence_interval'><code>::InferenceAbstractKKQuantileRegrCombinedLikelihood$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractKKQuantileRegrCombinedLikelihood" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../..//html/InferenceAbstractKKQuantileRegrCombinedLikelihood.html#method-InferenceAbstractKKQuantileRegrCombinedLikelihood-compute_asymp_two_sided_pval_for_treatment_effect'><code>::InferenceAbstractKKQuantileRegrCombinedLikelihood$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultKKQuantileRegrCombinedLikelihood-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultKKQuantileRegrCombinedLikelihood-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultKKQuantileRegrCombinedLikelihood$new(
  des_obj,
  tau = 0.5,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultKKQuantileRegrCombinedLikelihood-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultKKQuantileRegrCombinedLikelihood-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultKKQuantileRegrCombinedLikelihood$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultKKQuantileRegrCombinedLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultKKQuantileRegrCombinedLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultKKQuantileRegrCombinedLikelihood$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
