% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_continuous_ols_multi.R
\name{InferenceContinMultOLS}
\alias{InferenceContinMultOLS}
\title{Simple Mean Difference Inference based on Maximum Likelihood}
\description{
Simple Mean Difference Inference based on Maximum Likelihood

Simple Mean Difference Inference based on Maximum Likelihood
}
\details{
The methods that support confidence intervals and testing for the mean difference
in all response types (except Weibull with censoring)
sequential experimental design estimation and test object
after the sequential design is completed.
}
\examples{

## ------------------------------------------------
## Method `InferenceContinMultOLS$compute_treatment_estimate`
## ------------------------------------------------

\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLS$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}


## ------------------------------------------------
## Method `InferenceContinMultOLS$compute_asymp_confidence_interval`
## ------------------------------------------------

\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLS$new(seq_des)
seq_des_inf$compute_asymp_confidence_interval()
}


## ------------------------------------------------
## Method `InferenceContinMultOLS$compute_asymp_two_sided_pval_for_treatment_effect`
## ------------------------------------------------

\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLS$new(seq_des)
seq_des_inf$compute_asymp_two_sided_pval_for_treatment_effect()
}

}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \code{InferenceContinMultOLS}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceContinMultOLS-new}{\code{InferenceContinMultOLS$new()}}
\item \code{InferenceContinMultOLS$compute_treatment_estimate()}
\item \code{InferenceContinMultOLS$compute_asymp_confidence_interval()}
\item \code{InferenceContinMultOLS$compute_asymp_two_sided_pval_for_treatment_effect()}
\item \href{#method-InferenceContinMultOLS-clone}{\code{InferenceContinMultOLS$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultOLS-new"></a>}}
\subsection{Method \code{new()}}{
Initialize a sequential experimental design estimation and test object
after the sequential design is completed.
\subsection{Usage}{
\preformatted{InferenceContinMultOLS$new(des_obj, num_cores = 1, verbose = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A DesignSeqOneByOne object whose entire n subjects
are assigned and response y is recorded within.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize
the sampling during randomization-based inference and
bootstrap resampling.
The default is 1 for serial computation. For simple
estimators (e.g. mean difference and KK compound),
parallelization is achieved with zero-overhead C++ OpenMP.
For complex models (e.g. GLMs),
parallelization falls back to R's
\code{parallel::mclapply}, which incurs
session-forking overhead.}

\item{\code{verbose}}{A flag indicating whether messages should be
displayed to the user. Default is \code{TRUE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultOLS-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the appropriate estimate
\subsection{Usage}{
\preformatted{InferenceContinMultOLS$compute_treatment_estimate(estimate_only = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The setting-appropriate (see description) numeric estimate of the treatment effect
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLS$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultOLS-compute_asymp_confidence_interval"></a>}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1-alpha level frequentist confidence interval
differently for all response types, estimate types, and
test types.

Here we use the theory that MLE's computed for GLM's are asymptotically normal.
Hence these confidence intervals are asymptotically valid
and thus approximate for any sample size.
\subsection{Usage}{
\preformatted{InferenceContinMultOLS$compute_asymp_confidence_interval(alpha = 0.05)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A (1 - alpha)-sized frequentist confidence interval for the treatment effect
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLS$new(seq_des)
seq_des_inf$compute_asymp_confidence_interval()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a 2-sided p-value
\subsection{Usage}{
\preformatted{InferenceContinMultOLS$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null difference to test against. For any
treatment effect at all this is set to zero (the default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The approximate frequentist p-value
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLS$new(seq_des)
seq_des_inf$compute_asymp_two_sided_pval_for_treatment_effect()
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultOLS-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceContinMultOLS$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
