% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_continuous_KK_robust_regr_ivwc_univ.R
\name{InferenceContinUnivKKRobustRegrIVWC}
\alias{InferenceContinUnivKKRobustRegrIVWC}
\title{Univariate Robust-Regression IVWC Compound Inference for KK Designs}
\description{
Fits a variance-weighted compound estimator for KK matching-on-the-fly designs
with continuous responses using robust regression for matched-pair differences
and reservoir outcomes, with treatment as the only predictor.
}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \\code{EDI::InferenceKKPassThrough}
-> \\code{EDI::InferenceKKPassThroughCompound}
-> \\code{EDI::InferenceAbstractKKRobustRegrIVWC}
-> \code{InferenceContinUnivKKRobustRegrIVWC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceContinUnivKKRobustRegrIVWC$clone()}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceAbstractKKRobustRegrIVWC$compute_asymp_confidence_interval()</code></li>
<li>
<code>EDI::InferenceAbstractKKRobustRegrIVWC$
compute_asymp_two_sided_pval_for_treatment_effect()</code></li>
</li>
<li><code>EDI::InferenceAbstractKKRobustRegrIVWC$compute_treatment_estimate()</code></li>
<li><code>EDI::InferenceAbstractKKRobustRegrIVWC$duplicate()</code></li>
<li><code>EDI::InferenceAbstractKKRobustRegrIVWC$initialize()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinUnivKKRobustRegrIVWC-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceContinUnivKKRobustRegrIVWC$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
