% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_azriel.R
\name{InferenceIncidAzriel}
\alias{InferenceIncidAzriel}
\title{Azriel Blocked Incidence Inference}
\description{
Azriel Blocked Incidence Inference

Azriel Blocked Incidence Inference
}
\details{
Unadjusted blocked-design incidence inference using the simple mean-difference
point estimate with a block-stratified standard error.
}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \\code{EDI::InferenceAllSimpleMeanDiff}
-> \code{InferenceIncidAzriel}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidAzriel-new}{\code{InferenceIncidAzriel$new()}}
\item \href{#method-InferenceIncidAzriel-clone}{\code{InferenceIncidAzriel$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_confidence_interval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::InferenceAllSimpleMeanDiff$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceAllSimpleMeanDiff$compute_treatment_estimate()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidAzriel-new"></a>}}
\subsection{Method \code{new()}}{
Initialize Azriel blocked-design incidence inference.
\subsection{Usage}{
\preformatted{InferenceIncidAzriel$new(des_obj, num_cores = 1, verbose = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed design object.}

\item{\code{num_cores}}{Number of CPU cores to use.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{InferenceIncidAzriel} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidAzriel-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceIncidAzriel$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
