\name{InferenceIncidenceSimplePropDiffPooled}
\alias{InferenceIncidenceSimplePropDiffPooled}
\title{Simple Incidence Proportion Difference Inference}
\usage{
InferenceIncidenceSimplePropDiffPooled$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE,
  make_fork_cluster = NULL
)
}
\description{
Unadjusted incidence inference using the simple treated-minus-control
proportion difference with pooled equal-variance t inference.
}
