% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_KK_gcomp_multi.R
\name{InferenceIncidMultiKKGCompRiskRatio}
\alias{InferenceIncidMultiKKGCompRiskRatio}
\title{Multivariate G-Computation Risk-Ratio Inference for KK Designs with Binary Responses}
\description{
Fits an all-subject logistic working model for a KK incidence outcome using
treatment and all recorded covariates, then estimates the marginal risk
ratio by standardizing predicted risks under all-treated and all-control
assignments over the empirical covariate distribution. Matched pairs are
treated as clusters and reservoir subjects are treated as singletons when
computing the sandwich covariance.
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{::InferenceKKPassThrough} -> \code{::InferenceAbstractKKMarginalIncid} -> \code{::InferenceIncidKKGCompAbstract} -> \code{::InferenceIncidUnivKKGCompRiskRatio} -> \code{InferenceIncidMultiKKGCompRiskRatio}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidMultiKKGCompRiskRatio-clone}{\code{InferenceIncidMultiKKGCompRiskRatio$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractKKMarginalIncid" data-id="initialize"><a href='../..//html/InferenceAbstractKKMarginalIncid.html#method-InferenceAbstractKKMarginalIncid-initialize'><code>::InferenceAbstractKKMarginalIncid$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidKKGCompAbstract" data-id="compute_asymp_confidence_interval"><a href='../..//html/InferenceIncidKKGCompAbstract.html#method-InferenceIncidKKGCompAbstract-compute_asymp_confidence_interval'><code>::InferenceIncidKKGCompAbstract$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidKKGCompAbstract" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../..//html/InferenceIncidKKGCompAbstract.html#method-InferenceIncidKKGCompAbstract-compute_asymp_two_sided_pval_for_treatment_effect'><code>::InferenceIncidKKGCompAbstract$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidKKGCompAbstract" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceIncidKKGCompAbstract.html#method-InferenceIncidKKGCompAbstract-compute_bootstrap_two_sided_pval'><code>::InferenceIncidKKGCompAbstract$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidKKGCompAbstract" data-id="compute_treatment_estimate"><a href='../..//html/InferenceIncidKKGCompAbstract.html#method-InferenceIncidKKGCompAbstract-compute_treatment_estimate'><code>::InferenceIncidKKGCompAbstract$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidMultiKKGCompRiskRatio-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidMultiKKGCompRiskRatio-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidMultiKKGCompRiskRatio$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
