% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_KK_glmm_multi.R
\name{InferenceIncidMultiKKGLMM}
\alias{InferenceIncidMultiKKGLMM}
\title{Multivariate GLMM Inference for KK Designs with Binary Response}
\description{
Fits a Generalized Linear Mixed Model (GLMM) using the \pkg{glmmTMB} fitter for binary
(incidence) responses under a KK matching-on-the-fly design using the treatment
indicator and all recorded covariates as fixed-effect predictors. The matched-pair
strata enter the model as a subject-level random intercept \code{(1 | group_id)},
which accounts for within-pair correlation. Reservoir subjects each receive a
unique singleton group. The treatment estimate is the conditional
(subject-specific) log-odds ratio; inference uses the Wald Z-statistic from the
GLMM fixed-effects table.
}
\details{
This class requires the \pkg{glmmTMB} package, which is listed in Suggests
and is not installed automatically with \pkg{SeqExpMatch}.
Install \pkg{glmmTMB} before using this class.
}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{\link[EDI:InferenceAbstractKKGLMM]{EDI::InferenceAbstractKKGLMM}} -> \code{InferenceIncidMultiKKGLMM}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidMultiKKGLMM-clone}{\code{InferenceIncidMultiKKGLMM$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>EDI::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="approximate_randomization_distribution_beta_hat_T"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-approximate_randomization_distribution_beta_hat_T'><code>EDI::InferenceRand$approximate_randomization_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKGLMM" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceAbstractKKGLMM.html#method-InferenceAbstractKKGLMM-compute_asymp_confidence_interval'><code>EDI::InferenceAbstractKKGLMM$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKGLMM" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceAbstractKKGLMM.html#method-InferenceAbstractKKGLMM-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceAbstractKKGLMM$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKGLMM" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceAbstractKKGLMM.html#method-InferenceAbstractKKGLMM-compute_confidence_interval_rand'><code>EDI::InferenceAbstractKKGLMM$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKGLMM" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceAbstractKKGLMM.html#method-InferenceAbstractKKGLMM-compute_treatment_estimate'><code>EDI::InferenceAbstractKKGLMM$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKGLMM" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceAbstractKKGLMM.html#method-InferenceAbstractKKGLMM-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceAbstractKKGLMM$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKGLMM" data-id="initialize"><a href='../../EDI/html/InferenceAbstractKKGLMM.html#method-InferenceAbstractKKGLMM-initialize'><code>EDI::InferenceAbstractKKGLMM$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidMultiKKGLMM-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidMultiKKGLMM-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidMultiKKGLMM$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
