% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_KK_modified_poisson_multi.R
\name{InferenceIncidMultiKKModifiedPoisson}
\alias{InferenceIncidMultiKKModifiedPoisson}
\title{Multivariate Modified-Poisson Inference for KK Designs with Binary Responses}
\description{
Fits an all-subject modified-Poisson working model for incidence outcomes under
a KK matching-on-the-fly design using treatment and all recorded covariates as
predictors. Matched pairs are treated as clusters and reservoir subjects are
treated as singleton clusters when computing the sandwich covariance, so the
estimated treatment effect is a log risk ratio with cluster-robust inference.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneKK14$
  new(
  n = nrow(x_dat),
  response_type = "incidence",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 0, 1, 0, 1, 1, 0))
infer <- InferenceIncidMultiKKModifiedPoisson$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{EDI::InferenceBoot} -> \code{EDI::InferenceAsymp} -> \code{EDI::InferenceKKPassThrough} -> \code{EDI::InferenceAbstractKKMarginalIncid} -> \code{EDI::InferenceAbstractKKModifiedPoisson} -> \code{\link[EDI:InferenceIncidUnivKKModifiedPoisson]{EDI::InferenceIncidUnivKKModifiedPoisson}} -> \code{InferenceIncidMultiKKModifiedPoisson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidMultiKKModifiedPoisson-new}{\code{InferenceIncidMultiKKModifiedPoisson$new()}}
\item \href{#method-InferenceIncidMultiKKModifiedPoisson-clone}{\code{InferenceIncidMultiKKModifiedPoisson$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidUnivKKModifiedPoisson" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceIncidUnivKKModifiedPoisson.html#method-InferenceIncidUnivKKModifiedPoisson-compute_asymp_confidence_interval'><code>EDI::InferenceIncidUnivKKModifiedPoisson$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidUnivKKModifiedPoisson" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceIncidUnivKKModifiedPoisson.html#method-InferenceIncidUnivKKModifiedPoisson-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceIncidUnivKKModifiedPoisson$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidUnivKKModifiedPoisson" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceIncidUnivKKModifiedPoisson.html#method-InferenceIncidUnivKKModifiedPoisson-compute_treatment_estimate'><code>EDI::InferenceIncidUnivKKModifiedPoisson$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidMultiKKModifiedPoisson-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidMultiKKModifiedPoisson-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multivariate modified-Poisson inference object for a completed
KK design with a binary response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidMultiKKModifiedPoisson$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed KK \code{DesignSeqOneByOne} object with an incidence response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidMultiKKModifiedPoisson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidMultiKKModifiedPoisson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidMultiKKModifiedPoisson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
