% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_modified_poisson_multi.R
\name{InferenceIncidMultiModifiedPoisson}
\alias{InferenceIncidMultiModifiedPoisson}
\title{Multivariate Modified Poisson Inference for Binary Responses}
\description{
Fits the classic modified Poisson model for binary (incidence) responses under
non-KK designs using treatment and all recorded covariates as predictors in a
Poisson log-link working model with Huber-White sandwich variance. The
treatment effect is reported on the log-risk-ratio scale.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$
  new(
  n = nrow(x_dat),
  response_type = "incidence",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 0, 1, 0, 1, 1, 0))
infer <- InferenceIncidMultiModifiedPoisson$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{::InferenceIncidUnivModifiedPoisson} -> \code{InferenceIncidMultiModifiedPoisson}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidMultiModifiedPoisson-clone}{\code{InferenceIncidMultiModifiedPoisson$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidUnivModifiedPoisson" data-id="compute_asymp_confidence_interval"><a href='../..//html/InferenceIncidUnivModifiedPoisson.html#method-InferenceIncidUnivModifiedPoisson-compute_asymp_confidence_interval'><code>::InferenceIncidUnivModifiedPoisson$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidUnivModifiedPoisson" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../..//html/InferenceIncidUnivModifiedPoisson.html#method-InferenceIncidUnivModifiedPoisson-compute_asymp_two_sided_pval_for_treatment_effect'><code>::InferenceIncidUnivModifiedPoisson$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidUnivModifiedPoisson" data-id="compute_treatment_estimate"><a href='../..//html/InferenceIncidUnivModifiedPoisson.html#method-InferenceIncidUnivModifiedPoisson-compute_treatment_estimate'><code>::InferenceIncidUnivModifiedPoisson$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceIncidUnivModifiedPoisson" data-id="initialize"><a href='../..//html/InferenceIncidUnivModifiedPoisson.html#method-InferenceIncidUnivModifiedPoisson-initialize'><code>::InferenceIncidUnivModifiedPoisson$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidMultiModifiedPoisson-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidMultiModifiedPoisson-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidMultiModifiedPoisson$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
