% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_gcomp_univ.R
\name{InferenceIncidUnivGCompRiskDiff}
\alias{InferenceIncidUnivGCompRiskDiff}
\title{Univariate G-Computation Risk-Difference Inference for Binary Responses}
\description{
Fits a univariate logistic working model for an incidence outcome and then
estimates the marginal risk difference by standardizing predicted risks under
all-treated and all-control assignments over the empirical covariate
distribution.
}
\examples{

## ------------------------------------------------
## Method `InferenceIncidUnivGCompRiskDiff$new`
## ------------------------------------------------

\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 20, response_type = "incidence")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	p_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbinom(1, 1, p_i))
}
seq_des_inf = InferenceIncidUnivGCompRiskDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{EDI::InferenceBoot} -> \code{EDI::InferenceAsymp} -> \code{EDI::InferenceIncidGCompAbstract} -> \code{InferenceIncidUnivGCompRiskDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidUnivGCompRiskDiff-new}{\code{InferenceIncidUnivGCompRiskDiff$new()}}
\item \href{#method-InferenceIncidUnivGCompRiskDiff-compute_treatment_estimate}{\code{InferenceIncidUnivGCompRiskDiff$compute_treatment_estimate()}}
\item \href{#method-InferenceIncidUnivGCompRiskDiff-compute_asymp_confidence_interval}{\code{InferenceIncidUnivGCompRiskDiff$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceIncidUnivGCompRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceIncidUnivGCompRiskDiff$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceIncidUnivGCompRiskDiff-compute_bootstrap_two_sided_pval}{\code{InferenceIncidUnivGCompRiskDiff$compute_bootstrap_two_sided_pval()}}
\item \href{#method-InferenceIncidUnivGCompRiskDiff-clone}{\code{InferenceIncidUnivGCompRiskDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidUnivGCompRiskDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidUnivGCompRiskDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize the g-computation RD inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidUnivGCompRiskDiff$new(des_obj, num_cores = 1, verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{DesignSeqOneByOne} object with an incidence response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 20, response_type = "incidence")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	p_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbinom(1, 1, p_i))
}
seq_des_inf = InferenceIncidUnivGCompRiskDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidUnivGCompRiskDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidUnivGCompRiskDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the standardized marginal risk-difference estimate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidUnivGCompRiskDiff$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidUnivGCompRiskDiff-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidUnivGCompRiskDiff-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval for the marginal risk difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidUnivGCompRiskDiff$compute_asymp_confidence_interval(alpha = 0.05)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is 1 - \code{alpha}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidUnivGCompRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidUnivGCompRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the marginal risk difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidUnivGCompRiskDiff$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null risk difference to test against. The default is 0.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidUnivGCompRiskDiff-compute_bootstrap_two_sided_pval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidUnivGCompRiskDiff-compute_bootstrap_two_sided_pval}{}}}
\subsection{Method \code{compute_bootstrap_two_sided_pval()}}{
Computes a bootstrap two-sided p-value for the marginal risk difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidUnivGCompRiskDiff$compute_bootstrap_two_sided_pval(
  delta = NULL,
  B = 501,
  na.rm = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null risk difference to test against. The default is 0.}

\item{\code{B}}{Number of bootstrap samples.}

\item{\code{na.rm}}{Whether to remove non-finite bootstrap replicates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidUnivGCompRiskDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidUnivGCompRiskDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidUnivGCompRiskDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
