% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_KK_clogit_ivwc_univ.R
\name{InferenceIncidUnivKKClogitIVWC}
\alias{InferenceIncidUnivKKClogitIVWC}
\title{Univariate Conditional Logistic Regression Inference for KK Designs with Binary Response}
\description{
Fits a compound estimator for KK matching-on-the-fly designs with binary (incidence)
responses using only the treatment indicator (no additional covariates). For matched
pairs, a conditional logistic regression model is used (via the internal
\code{clogit_helper}, which processes discordant pairs and fits logistic regression
on within-pair signed differences). For reservoir subjects, a standard logistic
regression is used. The two estimates are combined via a variance-weighted linear
combination, analogous to \code{InferenceContinMultOLSKK}. If clogit fails
(e.g. no discordant pairs), the estimator falls back to logistic regression on the
reservoir only.
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{::InferenceKKPassThrough} -> \code{::InferenceAbstractKKClogitIVWC} -> \code{InferenceIncidUnivKKClogitIVWC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidUnivKKClogitIVWC-clone}{\code{InferenceIncidUnivKKClogitIVWC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractKKClogitIVWC" data-id="compute_asymp_confidence_interval"><a href='../..//html/InferenceAbstractKKClogitIVWC.html#method-InferenceAbstractKKClogitIVWC-compute_asymp_confidence_interval'><code>::InferenceAbstractKKClogitIVWC$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractKKClogitIVWC" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../..//html/InferenceAbstractKKClogitIVWC.html#method-InferenceAbstractKKClogitIVWC-compute_asymp_two_sided_pval_for_treatment_effect'><code>::InferenceAbstractKKClogitIVWC$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractKKClogitIVWC" data-id="compute_treatment_estimate"><a href='../..//html/InferenceAbstractKKClogitIVWC.html#method-InferenceAbstractKKClogitIVWC-compute_treatment_estimate'><code>::InferenceAbstractKKClogitIVWC$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceAbstractKKClogitIVWC" data-id="initialize"><a href='../..//html/InferenceAbstractKKClogitIVWC.html#method-InferenceAbstractKKClogitIVWC-initialize'><code>::InferenceAbstractKKClogitIVWC$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidUnivKKClogitIVWC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidUnivKKClogitIVWC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidUnivKKClogitIVWC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
