% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract_mle_or_KM_for_GLMs.R
\name{InferenceMLEorKMforGLMs}
\alias{InferenceMLEorKMforGLMs}
\title{GLM and Kaplan-Meier Inference}
\description{
GLM and Kaplan-Meier Inference

GLM and Kaplan-Meier Inference
}
\details{
Abstract class providing MLE/KM-based inference methods for GLM and survival models.
}
\keyword{internal}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \code{InferenceMLEorKMforGLMs}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceMLEorKMforGLMs$compute_treatment_estimate()}
\item \href{#method-InferenceMLEorKMforGLMs-clone}{\code{InferenceMLEorKMforGLMs$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::Inference$initialize()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_confidence_interval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceMLEorKMforGLMs-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the treatment estimate using the underlying model.
\subsection{Usage}{
\preformatted{InferenceMLEorKMforGLMs$compute_treatment_estimate(estimate_only = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceMLEorKMforGLMs-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceMLEorKMforGLMs$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
