% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_stereotype_probit_multi.R
\name{InferenceOrdinalMultiStereotypeProbitRegr}
\alias{InferenceOrdinalMultiStereotypeProbitRegr}
\title{Multivariate Stereotype Probit Inference for Ordinal Responses}
\description{
Stereotype probit inference for ordinal responses using treatment plus observed covariates.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$
  new(
  n = nrow(x_dat),
  response_type = "ordinal",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- InferenceOrdinalMultiStereotypeProbitRegr$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{\link[EDI:InferenceOrdinalUniStereotypeProbitRegr]{EDI::InferenceOrdinalUniStereotypeProbitRegr}} -> \code{InferenceOrdinalMultiStereotypeProbitRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceOrdinalMultiStereotypeProbitRegr-clone}{\code{InferenceOrdinalMultiStereotypeProbitRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceOrdinalUniStereotypeProbitRegr" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceOrdinalUniStereotypeProbitRegr.html#method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_confidence_interval'><code>EDI::InferenceOrdinalUniStereotypeProbitRegr$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceOrdinalUniStereotypeProbitRegr" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceOrdinalUniStereotypeProbitRegr.html#method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceOrdinalUniStereotypeProbitRegr$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceOrdinalUniStereotypeProbitRegr" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceOrdinalUniStereotypeProbitRegr.html#method-InferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate'><code>EDI::InferenceOrdinalUniStereotypeProbitRegr$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceOrdinalUniStereotypeProbitRegr" data-id="initialize"><a href='../../EDI/html/InferenceOrdinalUniStereotypeProbitRegr.html#method-InferenceOrdinalUniStereotypeProbitRegr-initialize'><code>EDI::InferenceOrdinalUniStereotypeProbitRegr$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalMultiStereotypeProbitRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalMultiStereotypeProbitRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalMultiStereotypeProbitRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
