% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_ordinal_partial_proportional_odds_multi.R
\name{InferenceOrdinalPartialProportionalOdds}
\alias{InferenceOrdinalPartialProportionalOdds}
\title{Partial Proportional-Odds Inference for Ordinal Responses}
\description{
Backward-compatible alias for the multivariate partial proportional-odds
estimator.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$new(n = nrow(x_dat), response_type = "ordinal", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- InferenceOrdinalPartialProportionalOdds$
  new(
  seq_des,
  nonparallel = "x1",
  verbose = FALSE
)
infer
}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{EDI::InferenceBoot} -> \code{EDI::InferenceAsymp} -> \code{\link[EDI:InferenceOrdinalPartialProportionalOddsAbstract]{EDI::InferenceOrdinalPartialProportionalOddsAbstract}} -> \code{\link[EDI:InferenceOrdinalMultiPartialProportionalOddsRegr]{EDI::InferenceOrdinalMultiPartialProportionalOddsRegr}} -> \code{InferenceOrdinalPartialProportionalOdds}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceOrdinalPartialProportionalOdds-new}{\code{InferenceOrdinalPartialProportionalOdds$new()}}
\item \href{#method-InferenceOrdinalPartialProportionalOdds-clone}{\code{InferenceOrdinalPartialProportionalOdds$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceOrdinalPartialProportionalOddsAbstract" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceOrdinalPartialProportionalOddsAbstract.html#method-InferenceOrdinalPartialProportionalOddsAbstract-compute_asymp_confidence_interval'><code>EDI::InferenceOrdinalPartialProportionalOddsAbstract$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceOrdinalPartialProportionalOddsAbstract" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceOrdinalPartialProportionalOddsAbstract.html#method-InferenceOrdinalPartialProportionalOddsAbstract-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceOrdinalPartialProportionalOddsAbstract$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceOrdinalPartialProportionalOddsAbstract" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceOrdinalPartialProportionalOddsAbstract.html#method-InferenceOrdinalPartialProportionalOddsAbstract-compute_treatment_estimate'><code>EDI::InferenceOrdinalPartialProportionalOddsAbstract$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalPartialProportionalOdds-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalPartialProportionalOdds-new}{}}}
\subsection{Method \code{new()}}{
Initialize the backward-compatible multivariate PPO estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalPartialProportionalOdds$new(
  des_obj,
  nonparallel = character(0),
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{DesignSeqOneByOne} object with an ordinal
response.}

\item{\code{nonparallel}}{Covariate names that should vary across thresholds.}

\item{\code{num_cores}}{Number of CPU cores for bootstrap/randomization helpers.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalPartialProportionalOdds-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalPartialProportionalOdds-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalPartialProportionalOdds$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
