% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_stereotype_probit_univ.R
\name{InferenceOrdinalUniStereotypeProbitRegr}
\alias{InferenceOrdinalUniStereotypeProbitRegr}
\title{Stereotype Probit Inference for Ordinal Responses
Stereotype probit inference for ordinal responses. The model exposes a common
treatment coefficient with category-specific score weights, and inference is
based on a probit link.}
\description{
Stereotype Probit Inference for Ordinal Responses
Stereotype probit inference for ordinal responses. The model exposes a common
treatment coefficient with category-specific score weights, and inference is
based on a probit link.

Stereotype Probit Inference for Ordinal Responses
Stereotype probit inference for ordinal responses. The model exposes a common
treatment coefficient with category-specific score weights, and inference is
based on a probit link.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$
  new(
  n = nrow(x_dat),
  response_type = "ordinal",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- InferenceOrdinalUniStereotypeProbitRegr$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{InferenceOrdinalUniStereotypeProbitRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceOrdinalUniStereotypeProbitRegr-new}{\code{InferenceOrdinalUniStereotypeProbitRegr$new()}}
\item \href{#method-InferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate}{\code{InferenceOrdinalUniStereotypeProbitRegr$compute_treatment_estimate()}}
\item \href{#method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_confidence_interval}{\code{InferenceOrdinalUniStereotypeProbitRegr$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceOrdinalUniStereotypeProbitRegr$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceOrdinalUniStereotypeProbitRegr-clone}{\code{InferenceOrdinalUniStereotypeProbitRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalUniStereotypeProbitRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalUniStereotypeProbitRegr-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalUniStereotypeProbitRegr$new(des_obj, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalUniStereotypeProbitRegr$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalUniStereotypeProbitRegr$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalUniStereotypeProbitRegr-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalUniStereotypeProbitRegr$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalUniStereotypeProbitRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalUniStereotypeProbitRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalUniStereotypeProbitRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
