% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_proportion_fractional_logit_univ.R
\name{InferencePropUniFractionalLogit}
\alias{InferencePropUniFractionalLogit}
\title{Univariate Fractional Logit Inference for Proportion Responses}
\description{
Univariate Fractional Logit Inference for Proportion Responses

Univariate Fractional Logit Inference for Proportion Responses
}
\details{
Fits a fractional logit model for proportion responses using a binomial logit
quasi-likelihood with sandwich-robust variance. The treatment effect is reported
on the log-odds scale.


This is the standard Papke-Wooldridge style fractional-response model for
outcomes in \eqn{[0, 1]}. The point estimate is the treatment log-odds
coefficient from the working logit model, and inference uses a robust sandwich
covariance rather than the model-based binomial variance.
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{InferencePropUniFractionalLogit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferencePropUniFractionalLogit-new}{\code{InferencePropUniFractionalLogit$new()}}
\item \href{#method-InferencePropUniFractionalLogit-compute_treatment_estimate}{\code{InferencePropUniFractionalLogit$compute_treatment_estimate()}}
\item \href{#method-InferencePropUniFractionalLogit-compute_asymp_confidence_interval}{\code{InferencePropUniFractionalLogit$compute_asymp_confidence_interval()}}
\item \href{#method-InferencePropUniFractionalLogit-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferencePropUniFractionalLogit$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferencePropUniFractionalLogit-clone}{\code{InferencePropUniFractionalLogit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferencePropUniFractionalLogit-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferencePropUniFractionalLogit-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferencePropUniFractionalLogit$new(des_obj, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferencePropUniFractionalLogit-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferencePropUniFractionalLogit-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferencePropUniFractionalLogit$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferencePropUniFractionalLogit-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferencePropUniFractionalLogit-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferencePropUniFractionalLogit$compute_asymp_confidence_interval(alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferencePropUniFractionalLogit-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferencePropUniFractionalLogit-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferencePropUniFractionalLogit$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferencePropUniFractionalLogit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferencePropUniFractionalLogit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferencePropUniFractionalLogit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
