% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_survival_coxph_multi.R
\name{InferenceSurvivalMultiCoxPHRegr}
\alias{InferenceSurvivalMultiCoxPHRegr}
\title{Simple Mean Difference Inference based on Maximum Likelihood}
\description{
Simple Mean Difference Inference based on Maximum Likelihood

Simple Mean Difference Inference based on Maximum Likelihood
}
\details{
The methods that support confidence intervals and testing for the mean difference
in all response types (except Weibull with censoring)
sequential experimental design estimation and test object
after the sequential design is completed.
}
\examples{

## ------------------------------------------------
## Method `InferenceSurvivalMultiCoxPHRegr$compute_treatment_estimate`
## ------------------------------------------------

\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "survival")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(
  ys = c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43),
  deads = c(1L, 0L, 1L, 1L, 0L, 1L)
)

seq_des_inf = InferenceSurvivalMultiCoxPHRegr$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \\code{EDI::InferenceMLEorKMforGLMs}
-> \\code{EDI::InferenceSurvivalUniCoxPHRegr}
-> \code{InferenceSurvivalMultiCoxPHRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceSurvivalMultiCoxPHRegr$compute_treatment_estimate()}
\item \code{InferenceSurvivalMultiCoxPHRegr$clone()}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::Inference$initialize()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_confidence_interval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceSurvivalMultiCoxPHRegr-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the appropriate estimate
\subsection{Usage}{
\preformatted{InferenceSurvivalMultiCoxPHRegr$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The setting-appropriate (see description) numeric estimate of the treatment effect
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "survival")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(
  ys = c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43),
  deads = c(1L, 0L, 1L, 1L, 0L, 1L)
)

seq_des_inf = InferenceSurvivalMultiCoxPHRegr$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceSurvivalMultiCoxPHRegr-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceSurvivalMultiCoxPHRegr$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
