% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{Any real number}
}
\value{
Its corresponding inverse logit value between 0 and 1 non inclusive
}
\description{
Calculates the inverse logit i.e., 1 / (1 + exp(-x))
}
