% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{logit}
\alias{logit}
\title{Logit}
\usage{
logit(p)
}
\arguments{
\item{p}{The value between 0 and 1 non inclusive}
}
\value{
Its corresponding logit value as a real number
}
\description{
Calculates the logit i.e., log(p / (1 - p))
}
