% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{mean_cpp}
\alias{mean_cpp}
\title{Fast Mean Calculation}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The mean of the vector.
}
\description{
Calculates the mean of a numeric vector using Rcpp for speed.
}
\examples{
\dontrun{
x <- rnorm(100)
mean_cpp(x)
}
}
