% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newcombe_independent_ci_cpp}
\alias{newcombe_independent_ci_cpp}
\title{Newcombe Hybrid Score Interval for Independent Proportions (Method 10)}
\usage{
newcombe_independent_ci_cpp(x1, n1, x2, n2, alpha)
}
\description{
Newcombe Hybrid Score Interval for Independent Proportions (Method 10)
}
\keyword{internal}
