% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{newcombe_paired_ci_cpp}
\alias{newcombe_paired_ci_cpp}
\title{Newcombe Hybrid Score Interval for Paired Proportions}
\usage{
newcombe_paired_ci_cpp(n11, n10, n01, n00, alpha)
}
\description{
Newcombe Hybrid Score Interval for Paired Proportions
}
\keyword{internal}
