% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pocock_simon_assign_and_update_cpp}
\alias{pocock_simon_assign_and_update_cpp}
\title{Pocock-Simon Minimization Assignment and Update}
\usage{
pocock_simon_assign_and_update_cpp(
  counts,
  subject_levels_idx,
  weights,
  p_best,
  prob_T
)
}
\arguments{
\item{counts}{A matrix of dimensions (sum of levels) x (number of treatments).
Modified in place.}

\item{subject_levels_idx}{An integer vector of indices indicating which rows of the counts matrix
the current subject belongs to.}

\item{weights}{A numeric vector of weights for each covariate.}

\item{p_best}{The probability of assigning the treatment that minimizes the imbalance.}

\item{prob_T}{Target probability for treatment (usually 0.5).}
}
\value{
The assigned treatment (0 or 1).
}
\description{
Pocock-Simon Minimization Assignment and Update
}
