% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pocock_simon_redraw_w_cpp}
\alias{pocock_simon_redraw_w_cpp}
\title{Pocock-Simon Minimization Redraw Assignments}
\usage{
pocock_simon_redraw_w_cpp(
  x_levels_matrix,
  num_levels_total,
  weights,
  p_best,
  prob_T
)
}
\arguments{
\item{x_levels_matrix}{A matrix where each row is a subject and each column is the row 
index in counts for that covariate.}

\item{num_levels_total}{Total number of levels across all covariates.}

\item{weights}{A numeric vector of weights for each covariate.}

\item{p_best}{The probability of assigning the treatment that minimizes the imbalance.}

\item{prob_T}{Target probability for treatment (usually 0.5).}
}
\value{
An integer vector of treatment assignments.
}
\description{
Pocock-Simon Minimization Redraw Assignments
}
