% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{robust_survreg}
\alias{robust_survreg}
\title{Robust Survival Regression}
\usage{
robust_survreg(
  y,
  dead,
  cov_matrix_or_vector,
  dist = "weibull",
  num_max_iter = 50
)
}
\arguments{
\item{y}{The response vector}

\item{dead}{The censoring vector (1 if dead/uncensored and 0 if censored)}

\item{cov_matrix_or_vector}{The model matrix}

\item{dist}{The parametric distribution form (default is Weibull)}

\item{num_max_iter}{Maximum # of iterations to repeat (default is 50)}
}
\value{
The Survival regression model object
}
\description{
Performs Survival regression from the original response and censoring vectors 
that if it fails, keeps trying with a different initialization point until a maximum number of iterations
}
