% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{robust_survreg_with_surv_object}
\alias{robust_survreg_with_surv_object}
\title{Robust Survival Regression}
\usage{
robust_survreg_with_surv_object(
  surv_object,
  cov_matrix_or_vector,
  dist = "weibull",
  num_max_iter = 50
)
}
\arguments{
\item{surv_object}{The survival object (built from the response vector and censoring vector)}

\item{cov_matrix_or_vector}{The model matrix}

\item{dist}{The parametric distribution form (default is Weibull)}

\item{num_max_iter}{Maximum # of iterations to repeat (default is 50)}
}
\value{
The Survival regression model object
}
\description{
Performs Survival regression from a survival::Surv object 
that if it fails, keeps trying with a different initialization point until a maximum number of iterations
}
