% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helpers.R
\name{sample_mode}
\alias{sample_mode}
\title{Sample Mode from an array}
\usage{
sample_mode(data)
}
\arguments{
\item{data}{The array of numbers}
}
\value{
The sample mode
}
\description{
Compute sample mode from an array of numbers
}
\examples{
sample_mode(c(1, 2, 2, 3, 3, 3))
}
