% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference__all_KK_compound_mean_diff.R
\name{SeqDesignInferenceAllKKCompoundMeanDiff}
\alias{SeqDesignInferenceAllKKCompoundMeanDiff}
\title{Inference based on Maximum Likelihood for KK designs}
\description{
Inference for mean difference
}
\examples{

## ------------------------------------------------
## Method `SeqDesignInferenceAllKKCompoundMeanDiff$compute_treatment_estimate`
## ------------------------------------------------

seq_des = SeqDesign$new(n = 6, p = 10, design = "CRD", response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = SeqDesignInferenceAllKKCompoundMeanDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
	
Compute confidence interval


## ------------------------------------------------
## Method `SeqDesignInferenceAllKKCompoundMeanDiff$compute_mle_confidence_interval`
## ------------------------------------------------

seq_des = SeqDesign$new(n = 6, p = 10, design = "CRD")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = SeqDesignInferenceAllKKCompoundMeanDiff$new(seq_des, test_type = "MLE-or-KM-based")
seq_des_inf$compute_confidence_interval()

Compute p-value


## ------------------------------------------------
## Method `SeqDesignInferenceAllKKCompoundMeanDiff$compute_mle_two_sided_pval_for_treatment_effect`
## ------------------------------------------------

seq_des = SeqDesign$new(n = 6, p = 10, design = "CRD")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = SeqDesignInferenceAllKKCompoundMeanDiff$new(seq_des)
seq_des_inf$compute_two_sided_pval_for_treatment_effect()
		
}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferenceMLEorKMKK} -> \code{SeqDesignInferenceAllKKCompoundMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceAllKKCompoundMeanDiff-new}{\code{SeqDesignInferenceAllKKCompoundMeanDiff$new()}}
\item \href{#method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_treatment_estimate}{\code{SeqDesignInferenceAllKKCompoundMeanDiff$compute_treatment_estimate()}}
\item \href{#method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_mle_confidence_interval}{\code{SeqDesignInferenceAllKKCompoundMeanDiff$compute_mle_confidence_interval()}}
\item \href{#method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_mle_two_sided_pval_for_treatment_effect}{\code{SeqDesignInferenceAllKKCompoundMeanDiff$compute_mle_two_sided_pval_for_treatment_effect()}}
\item \href{#method-SeqDesignInferenceAllKKCompoundMeanDiff-clone}{\code{SeqDesignInferenceAllKKCompoundMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceMLEorKMKK" data-id="approximate_boostrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInferenceMLEorKMKK.html#method-SeqDesignInferenceMLEorKMKK-approximate_boostrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInferenceMLEorKMKK$approximate_boostrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceAllKKCompoundMeanDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceAllKKCompoundMeanDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize a sequential experimental design estimation and test object after the sequential design is completed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceAllKKCompoundMeanDiff$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object whose entire n subjects are assigned and response y is recorded within.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize the sampling during randomization-based inference 
(which is very slow). The default is 1 for serial computation. This parameter is ignored
for \code{test_type = "MLE-or-KM-based"}.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed to the user. Default is \code{TRUE}

Compute treatment effect}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the appropriate estimate for compound mean difference across pairs and reservoir
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceAllKKCompoundMeanDiff$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The setting-appropriate (see description) numeric estimate of the treatment effect
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{seq_des = SeqDesign$new(n = 6, p = 10, design = "CRD", response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = SeqDesignInferenceAllKKCompoundMeanDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
	
Compute confidence interval

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_mle_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_mle_confidence_interval}{}}}
\subsection{Method \code{compute_mle_confidence_interval()}}{
Computes a 1-alpha level frequentist confidence interval

Here we use the theory that MLE's computed for GLM's are asymptotically normal (except in the case 
of estimat_type "median difference" where a nonparametric bootstrap confidence interval (see the \code{controlTest::quantileControlTest} method)
is employed. Hence these confidence intervals are asymptotically valid and thus approximate for any sample size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceAllKKCompoundMeanDiff$compute_mle_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is 1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A (1 - alpha)-sized frequentist confidence interval for the treatment effect
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{seq_des = SeqDesign$new(n = 6, p = 10, design = "CRD")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = SeqDesignInferenceAllKKCompoundMeanDiff$new(seq_des, test_type = "MLE-or-KM-based")
seq_des_inf$compute_confidence_interval()

Compute p-value

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_mle_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceAllKKCompoundMeanDiff-compute_mle_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_mle_two_sided_pval_for_treatment_effect()}}{
Computes a 2-sided p-value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceAllKKCompoundMeanDiff$compute_mle_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null difference to test against. For any treatment effect at all this is set to zero (the default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The approximate frequentist p-value
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{seq_des = SeqDesign$new(n = 6, p = 10, design = "CRD")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = SeqDesignInferenceAllKKCompoundMeanDiff$new(seq_des)
seq_des_inf$compute_two_sided_pval_for_treatment_effect()
		
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceAllKKCompoundMeanDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceAllKKCompoundMeanDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceAllKKCompoundMeanDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
