% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_for_inference_KK.R
\name{SeqDesignInferenceHelperKK}
\alias{SeqDesignInferenceHelperKK}
\title{Helper class for KK designs}
\description{
An abstract R6 Class that aids in tests and confidence intervals all KK sequential designs.

Initialize a helper object
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceHelperKK-new}{\code{SeqDesignInferenceHelperKK$new()}}
\item \href{#method-SeqDesignInferenceHelperKK-get_post_matching_data}{\code{SeqDesignInferenceHelperKK$get_post_matching_data()}}
\item \href{#method-SeqDesignInferenceHelperKK-get_data_frame_with_matching_dummies}{\code{SeqDesignInferenceHelperKK$get_data_frame_with_matching_dummies()}}
\item \href{#method-SeqDesignInferenceHelperKK-clone}{\code{SeqDesignInferenceHelperKK$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceHelperKK-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceHelperKK-new}{}}}
\subsection{Method \code{new()}}{
Initialize a helper object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceHelperKK$new(seq_des_obj, X)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object whose entire n subjects are assigned and response y is recorded within.}

\item{\code{X}}{A data.table object of the subjects' covariates in the study
Compute matching statistics}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceHelperKK-get_post_matching_data"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceHelperKK-get_post_matching_data}{}}}
\subsection{Method \code{get_post_matching_data()}}{
Compute matching statistics for KK designs	
Compute model matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceHelperKK$get_post_matching_data()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceHelperKK-get_data_frame_with_matching_dummies"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceHelperKK-get_data_frame_with_matching_dummies}{}}}
\subsection{Method \code{get_data_frame_with_matching_dummies()}}{
Compute model matrix with assignment and match indicies
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceHelperKK$get_data_frame_with_matching_dummies()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceHelperKK-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceHelperKK-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceHelperKK$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
