% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_exact_zhang.R
\name{SeqDesignInferenceIncidExactZhang}
\alias{SeqDesignInferenceIncidExactZhang}
\title{Exact incidence inference via Zhang's combined reservoir test}
\description{
Provides incidence-specific exact inference based on the Zhang (2026) combined
method. For CRD designs this uses the reservoir exact test only.
}
\details{
This class provides the \code{compute_exact_two_sided_pval_for_treatment_effect} and
\code{compute_exact_confidence_interval} methods. Standard methods like
\code{compute_treatment_estimate} and \code{compute_mle_confidence_interval}
are not implemented for this class.
}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferenceAbstractZhangCombinedBase} -> \code{SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract} -> \code{SeqDesignInferenceIncidExactZhang}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceIncidExactZhang-new}{\code{SeqDesignInferenceIncidExactZhang$new()}}
\item \href{#method-SeqDesignInferenceIncidExactZhang-clone}{\code{SeqDesignInferenceIncidExactZhang$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_mle_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_mle_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_mle_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_mle_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_mle_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_mle_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_treatment_estimate'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceIncidExactZhangAbstract" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInferenceIncidExactZhangAbstract.html#method-SeqDesignInferenceIncidExactZhangAbstract-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInferenceIncidExactZhangAbstract$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceIncidExactZhang-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceIncidExactZhang-new}{}}}
\subsection{Method \code{new()}}{
Initialize the exact incidence inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceIncidExactZhang$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object with an incidence response.}

\item{\code{num_cores}}{Number of CPU cores for parallel processing.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceIncidExactZhang-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceIncidExactZhang-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceIncidExactZhang$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
