% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_risk_diff_univ.R
\name{SeqDesignInferenceIncidUnivRiskDiff}
\alias{SeqDesignInferenceIncidUnivRiskDiff}
\title{Univariate Risk-Difference Regression for Binary Responses}
\description{
Fits a direct risk-difference estimator for binary (incidence) responses
using a linear probability model with HC2
heteroskedasticity-robust variance. The treatment effect is reported on the
risk-difference scale.
}
\details{
This class provides a direct additive treatment-effect model for binary
outcomes. With only the treatment indicator in the design matrix, the point
estimate coincides with the empirical risk difference.
}
\examples{

## ------------------------------------------------
## Method `SeqDesignInferenceIncidUnivRiskDiff$new`
## ------------------------------------------------

\dontrun{
seq_des = SeqDesignCRD$new(n = 20, response_type = "incidence")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	p_i = plogis(-0.8 + 0.5 * w_i)
	seq_des$add_subject_response(i, rbinom(1, 1, p_i))
}
seq_des_inf = SeqDesignInferenceIncidUnivRiskDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\section{Super class}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqDesignInferenceIncidUnivRiskDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceIncidUnivRiskDiff-new}{\code{SeqDesignInferenceIncidUnivRiskDiff$new()}}
\item \href{#method-SeqDesignInferenceIncidUnivRiskDiff-compute_treatment_estimate}{\code{SeqDesignInferenceIncidUnivRiskDiff$compute_treatment_estimate()}}
\item \href{#method-SeqDesignInferenceIncidUnivRiskDiff-compute_asymp_confidence_interval}{\code{SeqDesignInferenceIncidUnivRiskDiff$compute_asymp_confidence_interval()}}
\item \href{#method-SeqDesignInferenceIncidUnivRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect}{\code{SeqDesignInferenceIncidUnivRiskDiff$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-SeqDesignInferenceIncidUnivRiskDiff-clone}{\code{SeqDesignInferenceIncidUnivRiskDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceIncidUnivRiskDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceIncidUnivRiskDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize a risk-difference inference object for a completed
design with a binary response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceIncidUnivRiskDiff$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A completed \code{SeqDesign} object with an
incidence response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and
randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = SeqDesignCRD$new(n = 20, response_type = "incidence")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	p_i = plogis(-0.8 + 0.5 * w_i)
	seq_des$add_subject_response(i, rbinom(1, 1, p_i))
}
seq_des_inf = SeqDesignInferenceIncidUnivRiskDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceIncidUnivRiskDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceIncidUnivRiskDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the direct risk-difference estimate of the treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceIncidUnivRiskDiff$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceIncidUnivRiskDiff-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceIncidUnivRiskDiff-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval using the HC2 robust
standard error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceIncidUnivRiskDiff$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is
1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceIncidUnivRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceIncidUnivRiskDiff-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceIncidUnivRiskDiff$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null treatment effect on the risk-difference scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceIncidUnivRiskDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceIncidUnivRiskDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceIncidUnivRiskDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
