% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_gcomp_multi.R
\name{SeqDesignInferenceOrdinalMultiGCompMeanDiff}
\alias{SeqDesignInferenceOrdinalMultiGCompMeanDiff}
\title{Multivariate G-Computation for Ordinal Responses}
\description{
Fits a proportional odds model adjusting for all baseline covariates, then
standardizes to estimate the marginal mean difference.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignCRD$
  new(
  n = nrow(x_dat),
  response_type = "ordinal",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- SeqDesignInferenceOrdinalMultiGCompMeanDiff$
  new(
  seq_des,
  verbose = FALSE
)
infer$
  compute_treatment_estimate()
}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferenceOrdinalGCompAbstract} -> \code{SeqDesignInferenceOrdinalMultiGCompMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceOrdinalMultiGCompMeanDiff-clone}{\code{SeqDesignInferenceOrdinalMultiGCompMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalGCompAbstract" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalGCompAbstract.html#method-SeqDesignInferenceOrdinalGCompAbstract-compute_asymp_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceOrdinalGCompAbstract$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalGCompAbstract" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalGCompAbstract.html#method-SeqDesignInferenceOrdinalGCompAbstract-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInferenceOrdinalGCompAbstract$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalGCompAbstract" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalGCompAbstract.html#method-SeqDesignInferenceOrdinalGCompAbstract-compute_treatment_estimate'><code>SeqExpMatch::SeqDesignInferenceOrdinalGCompAbstract$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalGCompAbstract" data-id="initialize"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalGCompAbstract.html#method-SeqDesignInferenceOrdinalGCompAbstract-initialize'><code>SeqExpMatch::SeqDesignInferenceOrdinalGCompAbstract$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalMultiGCompMeanDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalMultiGCompMeanDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalMultiGCompMeanDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
