% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_KK_cond_adj_cat_logit_multi.R
\name{SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr}
\alias{SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr}
\title{Multivariate Conditional Adjacent-Category Inference for KK Designs}
\description{
Fits a conditional adjacent-category logit model for ordinal responses under a KK
matching-on-the-fly design, adjusting for baseline covariates.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignKK14$new(n = nrow(x_dat), response_type = "ordinal", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr$
  new(seq_des, verbose = FALSE)
infer

}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferenceKKPassThrough} -> \code{\link[SeqExpMatch:SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr]{SeqExpMatch::SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr}} -> \code{SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr-new}{\code{SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr$new()}}
\item \href{#method-SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr-clone}{\code{SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInferenceKKPassThrough.html#method-SeqDesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr" data-id="compute_asymp_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr.html#method-SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr-compute_asymp_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr.html#method-SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr-compute_asymp_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr.html#method-SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr-compute_treatment_estimate'><code>SeqExpMatch::SeqDesignInferenceOrdinalUnivKKCondAdjCatLogitRegr$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multivariate conditional adjacent-category inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object (must be a KK design).}

\item{\code{num_cores}}{Number of CPU cores.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalMultiKKCondAdjCatLogitRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
