% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_KK_glmm_probit_multi.R
\name{SeqDesignInferenceOrdinalMultiKKGLMMProbit}
\alias{SeqDesignInferenceOrdinalMultiKKGLMMProbit}
\title{Multivariate Probit CLMM Inference for KK Designs with Ordinal Response}
\description{
Fits a cumulative-link mixed model via \code{ordinal::clmm} with a probit
link for ordinal responses under a KK matching-on-the-fly design using the
treatment indicator and all recorded covariates as fixed effects. Matched
pairs enter as random intercepts and reservoir subjects each receive a
singleton group. Inference is based on the Wald Z-statistic.
}
\details{
This class requires the \pkg{ordinal} package, which is listed under
\code{Suggests} and is not installed automatically with \pkg{SeqExpMatch}.
Install it manually with \code{install.packages("ordinal")} before using this class.
}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferenceKKPassThrough} -> \code{SeqExpMatch::SeqDesignInferenceAbstractKKOrdinalCLMM} -> \code{SeqDesignInferenceOrdinalMultiKKGLMMProbit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceOrdinalMultiKKGLMMProbit-new}{\code{SeqDesignInferenceOrdinalMultiKKGLMMProbit$new()}}
\item \href{#method-SeqDesignInferenceOrdinalMultiKKGLMMProbit-clone}{\code{SeqDesignInferenceOrdinalMultiKKGLMMProbit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInferenceKKPassThrough.html#method-SeqDesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceAbstractKKOrdinalCLMM" data-id="compute_mle_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceAbstractKKOrdinalCLMM.html#method-SeqDesignInferenceAbstractKKOrdinalCLMM-compute_mle_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceAbstractKKOrdinalCLMM$compute_mle_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceAbstractKKOrdinalCLMM" data-id="compute_mle_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInferenceAbstractKKOrdinalCLMM.html#method-SeqDesignInferenceAbstractKKOrdinalCLMM-compute_mle_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInferenceAbstractKKOrdinalCLMM$compute_mle_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceAbstractKKOrdinalCLMM" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/SeqDesignInferenceAbstractKKOrdinalCLMM.html#method-SeqDesignInferenceAbstractKKOrdinalCLMM-compute_treatment_estimate'><code>SeqExpMatch::SeqDesignInferenceAbstractKKOrdinalCLMM$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalMultiKKGLMMProbit-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalMultiKKGLMMProbit-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multivariate ordinal probit CLMM inference object for a
completed KK design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalMultiKKGLMMProbit$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object (must be a KK design) whose entire
n subjects are assigned and whose ordinal response y is
recorded.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize the
sampling during randomization-based inference and bootstrap
resampling.}

\item{\code{verbose}}{Whether to print progress messages. Default is
\code{FALSE}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalMultiKKGLMMProbit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalMultiKKGLMMProbit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalMultiKKGLMMProbit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
