% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_ordinal_partial_proportional_odds_multi.R
\name{SeqDesignInferenceOrdinalPartialProportionalOdds}
\alias{SeqDesignInferenceOrdinalPartialProportionalOdds}
\title{Partial Proportional-Odds Inference for Ordinal Responses}
\description{
Backward-compatible alias for the multivariate partial proportional-odds
estimator.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignCRD$
  new(
    n = nrow(x_dat),
    response_type = "ordinal",
    verbose = FALSE
  )
for (i in seq_len(nrow(x_dat))) {
  seq_des$
    add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- SeqDesignInferenceOrdinalPartialProportionalOdds$
  new(
    seq_des,
    nonparallel = "x1",
    verbose = FALSE
  )
infer
}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{\link[SeqExpMatch:SeqDesignInferenceOrdinalPartialProportionalOddsAbstract]{SeqExpMatch::SeqDesignInferenceOrdinalPartialProportionalOddsAbstract}} -> \code{\link[SeqExpMatch:SeqDesignInferenceOrdinalMultiPartialProportionalOddsRegr]{SeqExpMatch::SeqDesignInferenceOrdinalMultiPartialProportionalOddsRegr}} -> \code{SeqDesignInferenceOrdinalPartialProportionalOdds}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceOrdinalPartialProportionalOdds-new}{\code{SeqDesignInferenceOrdinalPartialProportionalOdds$new()}}
\item \href{#method-SeqDesignInferenceOrdinalPartialProportionalOdds-clone}{\code{SeqDesignInferenceOrdinalPartialProportionalOdds$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalPartialProportionalOddsAbstract" data-id="compute_mle_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalPartialProportionalOddsAbstract.html#method-SeqDesignInferenceOrdinalPartialProportionalOddsAbstract-compute_mle_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceOrdinalPartialProportionalOddsAbstract$compute_mle_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalPartialProportionalOddsAbstract" data-id="compute_mle_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalPartialProportionalOddsAbstract.html#method-SeqDesignInferenceOrdinalPartialProportionalOddsAbstract-compute_mle_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInferenceOrdinalPartialProportionalOddsAbstract$compute_mle_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceOrdinalPartialProportionalOddsAbstract" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/SeqDesignInferenceOrdinalPartialProportionalOddsAbstract.html#method-SeqDesignInferenceOrdinalPartialProportionalOddsAbstract-compute_treatment_estimate'><code>SeqExpMatch::SeqDesignInferenceOrdinalPartialProportionalOddsAbstract$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalPartialProportionalOdds-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalPartialProportionalOdds-new}{}}}
\subsection{Method \code{new()}}{
Initialize the backward-compatible multivariate PPO estimator.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalPartialProportionalOdds$new(
  seq_des_obj,
  nonparallel = character(0),
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A completed \code{SeqDesign} object with an ordinal
response.}

\item{\code{nonparallel}}{Covariate names that should vary across thresholds.}

\item{\code{num_cores}}{Number of CPU cores for bootstrap/randomization helpers.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalPartialProportionalOdds-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalPartialProportionalOdds-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalPartialProportionalOdds$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
