% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_stereotype_probit_univ.R
\name{SeqDesignInferenceOrdinalUniStereotypeProbitRegr}
\alias{SeqDesignInferenceOrdinalUniStereotypeProbitRegr}
\title{Stereotype Probit Inference for Ordinal Responses}
\description{
Stereotype probit inference for ordinal responses. The model exposes a common
treatment coefficient with category-specific score weights, and inference is
based on a probit link.
}
\section{Super class}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqDesignInferenceOrdinalUniStereotypeProbitRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-new}{\code{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$new()}}
\item \href{#method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate}{\code{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$compute_treatment_estimate()}}
\item \href{#method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_mle_confidence_interval}{\code{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$compute_mle_confidence_interval()}}
\item \href{#method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_mle_two_sided_pval_for_treatment_effect}{\code{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$compute_mle_two_sided_pval_for_treatment_effect()}}
\item \href{#method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-clone}{\code{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_mle_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_mle_confidence_interval}{}}}
\subsection{Method \code{compute_mle_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$compute_mle_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_mle_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-compute_mle_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_mle_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$compute_mle_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUniStereotypeProbitRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUniStereotypeProbitRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
