% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_cond_cont_ratio_univ.R
\name{SeqDesignInferenceOrdinalUnivCondContRatioRegr}
\alias{SeqDesignInferenceOrdinalUnivCondContRatioRegr}
\title{Univariate Conditional Continuation-Ratio Inference}
\description{
Fits a conditional continuation-ratio logit model for ordinal responses. Each
matched pair is treated as a stratum.
Reservoir subjects each form their own unique stratum. The continuation-ratio
model logit(P(Y = k | Y >= k)) is fit by expanding the ordinal response into
binary continuation trials and applying conditional logistic regression.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- SeqDesignCRD$
  new(
  n = nrow(x_dat),
  response_type = "ordinal",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- SeqDesignInferenceOrdinalUnivCondContRatioRegr$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferenceKKPassThrough} -> \code{SeqDesignInferenceOrdinalUnivCondContRatioRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-new}{\code{SeqDesignInferenceOrdinalUnivCondContRatioRegr$new()}}
\item \href{#method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_treatment_estimate}{\code{SeqDesignInferenceOrdinalUnivCondContRatioRegr$compute_treatment_estimate()}}
\item \href{#method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_mle_confidence_interval}{\code{SeqDesignInferenceOrdinalUnivCondContRatioRegr$compute_mle_confidence_interval()}}
\item \href{#method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_mle_two_sided_pval_for_treatment_effect}{\code{SeqDesignInferenceOrdinalUnivCondContRatioRegr$compute_mle_two_sided_pval_for_treatment_effect()}}
\item \href{#method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-clone}{\code{SeqDesignInferenceOrdinalUnivCondContRatioRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInferenceKKPassThrough.html#method-SeqDesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-new}{}}}
\subsection{Method \code{new()}}{
Initialize a univariate conditional continuation-ratio inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUnivCondContRatioRegr$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object.}

\item{\code{num_cores}}{Number of CPU cores.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Returns the estimated treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUnivCondContRatioRegr$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_mle_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_mle_confidence_interval}{}}}
\subsection{Method \code{compute_mle_confidence_interval()}}{
Computes the MLE-based confidence interval.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUnivCondContRatioRegr$compute_mle_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{Significance level.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_mle_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-compute_mle_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_mle_two_sided_pval_for_treatment_effect()}}{
Computes the MLE-based p-value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUnivCondContRatioRegr$compute_mle_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{Null value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceOrdinalUnivCondContRatioRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceOrdinalUnivCondContRatioRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
