% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_proportion_fractional_logit_univ.R
\name{SeqDesignInferencePropUniFractionalLogit}
\alias{SeqDesignInferencePropUniFractionalLogit}
\title{Univariate Fractional Logit Inference for Proportion Responses}
\description{
Fits a fractional logit model for proportion responses using a binomial logit
quasi-likelihood with sandwich-robust variance. The treatment effect is reported
on the log-odds scale.
}
\details{
This is the standard Papke-Wooldridge style fractional-response model for
outcomes in \eqn{[0, 1]}. The point estimate is the treatment log-odds
coefficient from the working logit model, and inference uses a robust sandwich
covariance rather than the model-based binomial variance.
}
\examples{

## ------------------------------------------------
## Method `SeqDesignInferencePropUniFractionalLogit$new`
## ------------------------------------------------

\dontrun{
seq_des = SeqDesignBernoulli$new(n = 20, response_type = "proportion")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	mu_i = plogis(-0.5 + 0.8 * w_i)
	seq_des$add_subject_response(i, rbeta(1, shape1 = 8 * mu_i, shape2 = 8 * (1 - mu_i)))
}
seq_des_inf = SeqDesignInferencePropUniFractionalLogit$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\section{Super class}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqDesignInferencePropUniFractionalLogit}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferencePropUniFractionalLogit-new}{\code{SeqDesignInferencePropUniFractionalLogit$new()}}
\item \href{#method-SeqDesignInferencePropUniFractionalLogit-compute_treatment_estimate}{\code{SeqDesignInferencePropUniFractionalLogit$compute_treatment_estimate()}}
\item \href{#method-SeqDesignInferencePropUniFractionalLogit-compute_asymp_confidence_interval}{\code{SeqDesignInferencePropUniFractionalLogit$compute_asymp_confidence_interval()}}
\item \href{#method-SeqDesignInferencePropUniFractionalLogit-compute_asymp_two_sided_pval_for_treatment_effect}{\code{SeqDesignInferencePropUniFractionalLogit$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-SeqDesignInferencePropUniFractionalLogit-clone}{\code{SeqDesignInferencePropUniFractionalLogit$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniFractionalLogit-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniFractionalLogit-new}{}}}
\subsection{Method \code{new()}}{
Initialize a fractional-logit inference object for a completed design
with a proportion response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniFractionalLogit$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A completed \code{SeqDesign} object with a proportion response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = SeqDesignBernoulli$new(n = 20, response_type = "proportion")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	mu_i = plogis(-0.5 + 0.8 * w_i)
	seq_des$add_subject_response(i, rbeta(1, shape1 = 8 * mu_i, shape2 = 8 * (1 - mu_i)))
}
seq_des_inf = SeqDesignInferencePropUniFractionalLogit$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniFractionalLogit-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniFractionalLogit-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the fractional-logit estimate of the treatment effect on the
log-odds scale.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniFractionalLogit$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniFractionalLogit-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniFractionalLogit-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval using the sandwich
standard error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniFractionalLogit$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is
1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniFractionalLogit-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniFractionalLogit-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the treatment effect.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniFractionalLogit$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null treatment effect on the log-odds scale.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniFractionalLogit-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniFractionalLogit-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniFractionalLogit$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
