% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_proportion_gcomp_univ.R
\name{SeqDesignInferencePropUniGCompMeanDiff}
\alias{SeqDesignInferencePropUniGCompMeanDiff}
\title{Univariate G-Computation Mean-Difference Inference for Proportion Responses}
\description{
Fits a univariate fractional-logit working model for a proportion outcome and
then estimates the marginal mean difference by standardizing predicted mean
proportions under all-treated and all-control assignments over the empirical
covariate distribution.
}
\examples{

## ------------------------------------------------
## Method `SeqDesignInferencePropUniGCompMeanDiff$new`
## ------------------------------------------------

\dontrun{
seq_des = SeqDesignCRD$new(n = 20, response_type = "proportion")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	mu_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbeta(1, shape1 = 8 * mu_i, shape2 = 8 * (1 - mu_i)))
}
seq_des_inf = SeqDesignInferencePropUniGCompMeanDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferencePropGCompAbstract} -> \code{SeqDesignInferencePropUniGCompMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferencePropUniGCompMeanDiff-new}{\code{SeqDesignInferencePropUniGCompMeanDiff$new()}}
\item \href{#method-SeqDesignInferencePropUniGCompMeanDiff-compute_treatment_estimate}{\code{SeqDesignInferencePropUniGCompMeanDiff$compute_treatment_estimate()}}
\item \href{#method-SeqDesignInferencePropUniGCompMeanDiff-compute_asymp_confidence_interval}{\code{SeqDesignInferencePropUniGCompMeanDiff$compute_asymp_confidence_interval()}}
\item \href{#method-SeqDesignInferencePropUniGCompMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect}{\code{SeqDesignInferencePropUniGCompMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-SeqDesignInferencePropUniGCompMeanDiff-compute_bootstrap_two_sided_pval}{\code{SeqDesignInferencePropUniGCompMeanDiff$compute_bootstrap_two_sided_pval()}}
\item \href{#method-SeqDesignInferencePropUniGCompMeanDiff-clone}{\code{SeqDesignInferencePropUniGCompMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_exact_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferencePropGCompAbstract" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInferencePropGCompAbstract.html#method-SeqDesignInferencePropGCompAbstract-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInferencePropGCompAbstract$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniGCompMeanDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniGCompMeanDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize the g-computation mean-difference inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniGCompMeanDiff$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A completed \code{SeqDesign} object with a proportion response.}

\item{\code{num_cores}}{The number of CPU cores to use for bootstrap and randomization inference.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = SeqDesignCRD$new(n = 20, response_type = "proportion")
for (i in 1:20) {
	x_i = data.frame(x1 = rnorm(1), x2 = rnorm(1))
	w_i = seq_des$add_subject_to_experiment_and_assign(x_i)
	mu_i = plogis(-0.5 + 0.6 * w_i)
	seq_des$add_subject_response(i, rbeta(1, shape1 = 8 * mu_i, shape2 = 8 * (1 - mu_i)))
}
seq_des_inf = SeqDesignInferencePropUniGCompMeanDiff$new(seq_des)
seq_des_inf$compute_treatment_estimate()
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniGCompMeanDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniGCompMeanDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the standardized marginal mean-difference estimate.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniGCompMeanDiff$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniGCompMeanDiff-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniGCompMeanDiff-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1 - \code{alpha} confidence interval for the marginal mean difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniGCompMeanDiff$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is 1 - \code{alpha}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniGCompMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniGCompMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a two-sided p-value for the marginal mean difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniGCompMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null mean difference to test against. The default is 0.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniGCompMeanDiff-compute_bootstrap_two_sided_pval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniGCompMeanDiff-compute_bootstrap_two_sided_pval}{}}}
\subsection{Method \code{compute_bootstrap_two_sided_pval()}}{
Computes a bootstrap two-sided p-value for the marginal mean difference.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniGCompMeanDiff$compute_bootstrap_two_sided_pval(
  delta = 0,
  B = 501,
  na.rm = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null mean difference to test against. The default is 0.}

\item{\code{B}}{Number of bootstrap samples.}

\item{\code{na.rm}}{Whether to remove non-finite bootstrap replicates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferencePropUniGCompMeanDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferencePropUniGCompMeanDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferencePropUniGCompMeanDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
