% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_survival_KK_weibull_frailty_multi_combined_likelihood.R
\name{SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood}
\alias{SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood}
\title{Multivariate Weibull Frailty Combined-Likelihood Compound Inference for KK Designs (stub)}
\description{
Stub placeholder. Methods are not yet implemented.
}
\section{Super classes}{
\code{SeqExpMatch::SeqDesignInference} -> \code{SeqExpMatch::SeqDesignInferenceKKPassThrough} -> \code{SeqExpMatch::SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood} -> \code{SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood-new}{\code{SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood$new()}}
\item \href{#method-SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood-clone}{\code{SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInferenceKKPassThrough.html#method-SeqDesignInferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood" data-id="compute_mle_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood.html#method-SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood-compute_mle_confidence_interval'><code>SeqExpMatch::SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood$compute_mle_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood" data-id="compute_mle_two_sided_pval_for_treatment_effect"><a href='../../SeqExpMatch/html/SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood.html#method-SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood-compute_mle_two_sided_pval_for_treatment_effect'><code>SeqExpMatch::SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood$compute_mle_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood" data-id="compute_treatment_estimate"><a href='../../SeqExpMatch/html/SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood.html#method-SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood-compute_treatment_estimate'><code>SeqExpMatch::SeqDesignInferenceAbstractKKWeibullFrailtyCombinedLikelihood$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood-new}{}}}
\subsection{Method \code{new()}}{
Initialize the inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object (must be a KK design).}

\item{\code{num_cores}}{Number of CPU cores for parallel processing.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceSurvivalMultiKKWeibullFrailtyCombinedLikelihood$clone(
  deep = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
