% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_survival__restricted_mean_diff.R
\name{SeqDesignInferenceSurvivalRestrictedMeanDiff}
\alias{SeqDesignInferenceSurvivalRestrictedMeanDiff}
\title{Simple Mean Difference Inference based on Maximum Likelihood}
\description{
The methods that support confidence intervals and testing for the mean difference
in all response types (except Weibull with censoring) sequential experimental design estimation and test object after the sequential design is completed.
}
\examples{
\dontrun{
# Setup design and add responses
seq_des = SeqDesignCRD$new(n = 6, response_type = "survival")
X = data.table::data.table(MASS::biopsy[1:6, 2:10])
for (i in 1:6) seq_des$add_subject_to_experiment_and_assign(X[i, ])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

# Initialize inference object
seq_des_inf = SeqDesignInferenceSurvivalRestrictedMeanDiff$new(seq_des)

# compute_treatment_estimate
seq_des_inf$compute_treatment_estimate()

# compute_mle_confidence_interval
seq_des_inf$compute_mle_confidence_interval()

# compute_mle_two_sided_pval_for_treatment_effect
seq_des_inf$compute_mle_two_sided_pval_for_treatment_effect()
}

}
\section{Super class}{
\code{\link[SeqExpMatch:SeqDesignInference]{SeqExpMatch::SeqDesignInference}} -> \code{SeqDesignInferenceSurvivalRestrictedMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignInferenceSurvivalRestrictedMeanDiff-new}{\code{SeqDesignInferenceSurvivalRestrictedMeanDiff$new()}}
\item \href{#method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_treatment_estimate}{\code{SeqDesignInferenceSurvivalRestrictedMeanDiff$compute_treatment_estimate()}}
\item \href{#method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_mle_confidence_interval}{\code{SeqDesignInferenceSurvivalRestrictedMeanDiff$compute_mle_confidence_interval()}}
\item \href{#method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_mle_two_sided_pval_for_treatment_effect}{\code{SeqDesignInferenceSurvivalRestrictedMeanDiff$compute_mle_two_sided_pval_for_treatment_effect()}}
\item \href{#method-SeqDesignInferenceSurvivalRestrictedMeanDiff-clone}{\code{SeqDesignInferenceSurvivalRestrictedMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-approximate_bootstrap_distribution_beta_hat_T'><code>SeqExpMatch::SeqDesignInference$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>SeqExpMatch::SeqDesignInference$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_confidence_interval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_confidence_interval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_bootstrap_two_sided_pval"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_bootstrap_two_sided_pval'><code>SeqExpMatch::SeqDesignInference$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_confidence_interval_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_confidence_interval_rand'><code>SeqExpMatch::SeqDesignInference$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-compute_two_sided_pval_for_treatment_effect_rand'><code>SeqExpMatch::SeqDesignInference$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="duplicate"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-duplicate'><code>SeqExpMatch::SeqDesignInference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesignInference" data-id="set_custom_randomization_statistic_function"><a href='../../SeqExpMatch/html/SeqDesignInference.html#method-SeqDesignInference-set_custom_randomization_statistic_function'><code>SeqExpMatch::SeqDesignInference$set_custom_randomization_statistic_function()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceSurvivalRestrictedMeanDiff-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceSurvivalRestrictedMeanDiff-new}{}}}
\subsection{Method \code{new()}}{
Initialize a sequential experimental design estimation and test object after the sequential design is completed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceSurvivalRestrictedMeanDiff$new(
  seq_des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{seq_des_obj}}{A SeqDesign object whose entire n subjects are assigned and response y is recorded within.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize the sampling during randomization-based inference
(which is very slow). The default is 1 for serial computation. This parameter is ignored
for \code{test_type = "MLE-or-KM-based"}.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed to the user. Default is \code{TRUE}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the appropriate estimate for mean difference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceSurvivalRestrictedMeanDiff$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The setting-appropriate (see description) numeric estimate of the treatment effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_mle_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_mle_confidence_interval}{}}}
\subsection{Method \code{compute_mle_confidence_interval()}}{
Computes a 1-alpha level frequentist confidence interval differently for all response types, estimate types and test types.

Here we use the theory that MLE's computed for GLM's are asymptotically normal.
Hence these confidence intervals are asymptotically valid and thus approximate for any sample size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceSurvivalRestrictedMeanDiff$compute_mle_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence interval is 1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A (1 - alpha)-sized frequentist confidence interval for the treatment effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_mle_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceSurvivalRestrictedMeanDiff-compute_mle_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_mle_two_sided_pval_for_treatment_effect()}}{
Computes a 2-sided p-value via the log rank test
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceSurvivalRestrictedMeanDiff$compute_mle_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null difference to test against. For any treatment effect at all this is set to zero (the default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The approximate frequentist p-value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignInferenceSurvivalRestrictedMeanDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignInferenceSurvivalRestrictedMeanDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignInferenceSurvivalRestrictedMeanDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
