% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_seq_pocock_simon.R
\name{SeqDesignPocockSimon}
\alias{SeqDesignPocockSimon}
\title{Pocock & Simon's Minimization Sequential Design}
\description{
An R6 Class encapsulating the data and functionality for a Pocock & Simon sequential experimental design.
This design minimizes the imbalance across treatments for multiple covariates.
}
\section{Super classes}{
\code{SeqExpMatch::Design} -> \code{SeqExpMatch::SeqDesign} -> \code{SeqDesignPocockSimon}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignPocockSimon-new}{\code{SeqDesignPocockSimon$new()}}
\item \href{#method-SeqDesignPocockSimon-clone}{\code{SeqDesignPocockSimon$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="add_all_subject_assignments"><a href='../../SeqExpMatch/html/Design.html#method-Design-add_all_subject_assignments'><code>SeqExpMatch::Design$add_all_subject_assignments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="add_all_subject_responses"><a href='../../SeqExpMatch/html/Design.html#method-Design-add_all_subject_responses'><code>SeqExpMatch::Design$add_all_subject_responses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="add_subject"><a href='../../SeqExpMatch/html/Design.html#method-Design-add_subject'><code>SeqExpMatch::Design$add_subject()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="add_subject_response"><a href='../../SeqExpMatch/html/Design.html#method-Design-add_subject_response'><code>SeqExpMatch::Design$add_subject_response()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="any_censoring"><a href='../../SeqExpMatch/html/Design.html#method-Design-any_censoring'><code>SeqExpMatch::Design$any_censoring()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="assert_all_subjects_arrived"><a href='../../SeqExpMatch/html/Design.html#method-Design-assert_all_subjects_arrived'><code>SeqExpMatch::Design$assert_all_subjects_arrived()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="assert_even_allocation"><a href='../../SeqExpMatch/html/Design.html#method-Design-assert_even_allocation'><code>SeqExpMatch::Design$assert_even_allocation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="assert_experiment_completed"><a href='../../SeqExpMatch/html/Design.html#method-Design-assert_experiment_completed'><code>SeqExpMatch::Design$assert_experiment_completed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="assert_fixed_sample"><a href='../../SeqExpMatch/html/Design.html#method-Design-assert_fixed_sample'><code>SeqExpMatch::Design$assert_fixed_sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="check_experiment_completed"><a href='../../SeqExpMatch/html/Design.html#method-Design-check_experiment_completed'><code>SeqExpMatch::Design$check_experiment_completed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="duplicate"><a href='../../SeqExpMatch/html/Design.html#method-Design-duplicate'><code>SeqExpMatch::Design$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_X"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_X'><code>SeqExpMatch::Design$get_X()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_X_imp"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_X_imp'><code>SeqExpMatch::Design$get_X_imp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_X_raw"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_X_raw'><code>SeqExpMatch::Design$get_X_raw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_dead"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_dead'><code>SeqExpMatch::Design$get_dead()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_n"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_n'><code>SeqExpMatch::Design$get_n()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_prob_T"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_prob_T'><code>SeqExpMatch::Design$get_prob_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_response_type"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_response_type'><code>SeqExpMatch::Design$get_response_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_t"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_t'><code>SeqExpMatch::Design$get_t()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_w"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_w'><code>SeqExpMatch::Design$get_w()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="get_y"><a href='../../SeqExpMatch/html/Design.html#method-Design-get_y'><code>SeqExpMatch::Design$get_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="Design" data-id="is_fixed_sample_size"><a href='../../SeqExpMatch/html/Design.html#method-Design-is_fixed_sample_size'><code>SeqExpMatch::Design$is_fixed_sample_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="add_subject_to_experiment_and_assign"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-add_subject_to_experiment_and_assign'><code>SeqExpMatch::SeqDesign$add_subject_to_experiment_and_assign()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="draw_ws_according_to_design"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-draw_ws_according_to_design'><code>SeqExpMatch::SeqDesign$draw_ws_according_to_design()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="print_current_subject_assignment"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-print_current_subject_assignment'><code>SeqExpMatch::SeqDesign$print_current_subject_assignment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="randomize"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-randomize'><code>SeqExpMatch::SeqDesign$randomize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="supports_resampling"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-supports_resampling'><code>SeqExpMatch::SeqDesign$supports_resampling()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignPocockSimon-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignPocockSimon-new}{}}}
\subsection{Method \code{new()}}{
Initialize a Pocock & Simon sequential experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignPocockSimon$new(
  strata_cols,
  weights = NULL,
  p_best = 0.8,
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strata_cols}}{The names of the covariates to be used for minimization. These must be factor or categorical variables.}

\item{\code{weights}}{A numeric vector of weights for each covariate. Defaults to 1 for all.}

\item{\code{p_best}}{The probability of assigning the treatment that minimizes the imbalance. Defaults to 0.8.}

\item{\code{response_type}}{The data type of response values.}

\item{\code{prob_T}}{The probability of the treatment assignment.}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size.}

\item{\code{num_cores}}{The number of CPU cores.}

\item{\code{verbose}}{Flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SeqDesignPocockSimon` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignPocockSimon-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignPocockSimon-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignPocockSimon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
