% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_cluster.R
\name{stop_global_fork_cluster}
\alias{stop_global_fork_cluster}
\title{Stop the Global Fork Cluster}
\usage{
stop_global_fork_cluster()
}
\value{
Invisible \code{NULL}.
}
\description{
This function stops the global fork cluster stored in the package environment
and clears the reference.
}
