% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{var_cpp}
\alias{var_cpp}
\title{Fast Variance Calculation}
\arguments{
\item{x}{A numeric vector.}
}
\value{
The variance of the vector.
}
\description{
Calculates the variance of a numeric vector using Rcpp for speed.
}
\examples{
\dontrun{
x <- rnorm(100)
var_cpp(x)
}
}
