% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wilson_score_interval_cpp}
\alias{wilson_score_interval_cpp}
\title{Wilson Score Interval for a Single Proportion}
\usage{
wilson_score_interval_cpp(x, n, alpha)
}
\description{
Wilson Score Interval for a Single Proportion
}
\keyword{internal}
