% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_abstract.R
\name{Design}
\alias{Design}
\title{An Abstract Experimental Design}
\description{
Internal method.
An abstract R6 Class encapsulating the data and functionality for an experimental design.
This class takes care of data storage and response handling.
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Design-new}{\code{Design$new()}}
\item \href{#method-Design-add_subject}{\code{Design$add_subject()}}
\item \href{#method-Design-add_subject_response}{\code{Design$add_subject_response()}}
\item \href{#method-Design-add_all_subject_responses}{\code{Design$add_all_subject_responses()}}
\item \href{#method-Design-add_all_subject_assignments}{\code{Design$add_all_subject_assignments()}}
\item \href{#method-Design-is_fixed_sample_size}{\code{Design$is_fixed_sample_size()}}
\item \href{#method-Design-assert_all_subjects_arrived}{\code{Design$assert_all_subjects_arrived()}}
\item \code{Design$assert_all_responses_recorded()}
\item \href{#method-Design-check_experiment_completed}{\code{Design$check_experiment_completed()}}
\item \href{#method-Design-assert_even_allocation}{\code{Design$assert_even_allocation()}}
\item \href{#method-Design-assert_fixed_sample}{\code{Design$assert_fixed_sample()}}
\item \href{#method-Design-any_censoring}{\code{Design$any_censoring()}}
\item \href{#method-Design-get_t}{\code{Design$get_t()}}
\item \href{#method-Design-get_X_raw}{\code{Design$get_X_raw()}}
\item \href{#method-Design-get_X_imp}{\code{Design$get_X_imp()}}
\item \href{#method-Design-get_X}{\code{Design$get_X()}}
\item \href{#method-Design-get_y}{\code{Design$get_y()}}
\item \href{#method-Design-get_w}{\code{Design$get_w()}}
\item \href{#method-Design-get_block_ids}{\code{Design$get_block_ids()}}
\item \href{#method-Design-get_n}{\code{Design$get_n()}}
\item \href{#method-Design-get_dead}{\code{Design$get_dead()}}
\item \href{#method-Design-get_prob_T}{\code{Design$get_prob_T()}}
\item \href{#method-Design-get_response_type}{\code{Design$get_response_type()}}
\item \href{#method-Design-duplicate}{\code{Design$duplicate()}}
\item \href{#method-Design-resample_design}{\code{Design$resample_design()}}
\item \href{#method-Design-clone}{\code{Design$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-new"></a>}}
\subsection{Method \code{new()}}{
Initialize an experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$new(
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{"continuous", "incidence", "proportion", "count", "survival", or
"ordinal".}

\item{\code{prob_T}}{Probability of treatment assignment.}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size (if fixed).}

\item{\code{num_cores}}{Number of CPU cores.}

\item{\code{verbose}}{Flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Design` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_subject"></a>}}
\subsection{Method \code{add_subject()}}{
Add subject-specific measurements for the next subject entrant
\subsection{Usage}{
\preformatted{Design$add_subject(x_new, allow_new_cols = TRUE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x_new}}{A row of the data frame corresponding to the new subject.}

\item{\code{allow_new_cols}}{Allow new features in the new subject's covariates.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_subject_response"></a>}}
\subsection{Method \code{add_subject_response()}}{
For CARA designs, add subject response for the a subject
\subsection{Usage}{
\preformatted{Design$add_subject_response(t, y, dead = 1)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{The subject index.}

\item{\code{y}}{The response value.}

\item{\code{dead}}{If the response is censored (0 for survival).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_all_subject_responses"></a>}}
\subsection{Method \code{add_all_subject_responses()}}{
For non-CARA designs, add all subject responses
\subsection{Usage}{
\preformatted{Design$add_all_subject_responses(ys, deads = NULL)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ys}}{The responses as a numeric vector.}

\item{\code{deads}}{The binary vector indicating if dead/censored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_all_subject_assignments"></a>}}
\subsection{Method \code{add_all_subject_assignments()}}{
For analysis on already-completed experimental data
\subsection{Usage}{
\preformatted{Design$add_all_subject_assignments(w)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{The binary responses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-is_fixed_sample_size"></a>}}
\subsection{Method \code{is_fixed_sample_size()}}{
Check if this design was initialized with a fixed sample size n
\subsection{Usage}{
\preformatted{Design$is_fixed_sample_size()}
}

\subsection{Returns}{
TRUE if fixed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_all_subjects_arrived"></a>}}
\subsection{Method \code{assert_all_subjects_arrived()}}{
Asserts if all subjects arrived.
\subsection{Usage}{
\preformatted{Design$assert_all_subjects_arrived()}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_all_responses_recorded"></a>}}
\subsection{Method \code{assert_all_responses_recorded()}}{
Asserts if all responses are recorded.
\subsection{Usage}{
\preformatted{Design$assert_all_responses_recorded()}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-check_experiment_completed"></a>}}
\subsection{Method \code{check_experiment_completed()}}{
Checks if the experiment is completed.
\subsection{Usage}{
\preformatted{Design$check_experiment_completed()}
}

\subsection{Returns}{
\code{TRUE} if experiment is complete, \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_even_allocation"></a>}}
\subsection{Method \code{assert_even_allocation()}}{
Checks if the experiment has a 50-50 allocation.
\subsection{Usage}{
\preformatted{Design$assert_even_allocation()}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_fixed_sample"></a>}}
\subsection{Method \code{assert_fixed_sample()}}{
Checks if the experiment has a fixed sample size.
\subsection{Usage}{
\preformatted{Design$assert_fixed_sample()}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-any_censoring"></a>}}
\subsection{Method \code{any_censoring()}}{
Checks if the experiment has any censored responses
\subsection{Usage}{
\preformatted{Design$any_censoring()}
}

\subsection{Returns}{
\code{TRUE} if any censored.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_t"></a>}}
\subsection{Method \code{get_t()}}{
Get t
\subsection{Usage}{
\preformatted{Design$get_t()}
}

\subsection{Returns}{
The current number of subjects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X_raw"></a>}}
\subsection{Method \code{get_X_raw()}}{
Get raw X information
\subsection{Usage}{
\preformatted{Design$get_X_raw()}
}

\subsection{Returns}{
A data frame of subject data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X_imp"></a>}}
\subsection{Method \code{get_X_imp()}}{
Get imputed X information
\subsection{Usage}{
\preformatted{Design$get_X_imp()}
}

\subsection{Returns}{
Same as \code{Xraw} except with imputations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X"></a>}}
\subsection{Method \code{get_X()}}{
Get X matrix
\subsection{Usage}{
\preformatted{Design$get_X()}
}

\subsection{Returns}{
A numeric matrix of subject data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_y"></a>}}
\subsection{Method \code{get_y()}}{
Get y
\subsection{Usage}{
\preformatted{Design$get_y()}
}

\subsection{Returns}{
A numeric vector of subject responses.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_w"></a>}}
\subsection{Method \code{get_w()}}{
Get w
\subsection{Usage}{
\preformatted{Design$get_w()}
}

\subsection{Returns}{
A binary vector of subject assignments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_block_ids"></a>}}
\subsection{Method \code{get_block_ids()}}{
If the design is a block design, get block identifiers (otherwise halts)
\subsection{Usage}{
\preformatted{Design$get_block_ids()}
}

\subsection{Returns}{
An integer vector of block identifiers.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_n"></a>}}
\subsection{Method \code{get_n()}}{
Get n, the sample size
\subsection{Usage}{
\preformatted{Design$get_n()}
}

\subsection{Returns}{
The number of subjects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_dead"></a>}}
\subsection{Method \code{get_dead()}}{
Get dead
\subsection{Usage}{
\preformatted{Design$get_dead()}
}

\subsection{Returns}{
A binary vector of whether the subject is dead.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_prob_T"></a>}}
\subsection{Method \code{get_prob_T()}}{
Get probability of treatment
\subsection{Usage}{
\preformatted{Design$get_prob_T()}
}

\subsection{Returns}{
The specified probability.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_response_type"></a>}}
\subsection{Method \code{get_response_type()}}{
Get response type
\subsection{Usage}{
\preformatted{Design$get_response_type()}
}

\subsection{Returns}{
The specified response type.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-duplicate"></a>}}
\subsection{Method \code{duplicate()}}{
Duplicate this design object
\subsection{Usage}{
\preformatted{Design$duplicate(verbose = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{A flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Design` object with the same data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-resample_design"></a>}}
\subsection{Method \code{resample_design()}}{
Resample this design for bootstrap inference.
Draws n subjects with replacement and updates w, y, dead (and m for KK designs).
\subsection{Usage}{
\preformatted{Design$resample_design()}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{Design$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
