% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_abstract.R
\name{Design}
\alias{Design}
\title{An Abstract Experimental Design}
\description{
Internal method.
An abstract R6 Class encapsulating the data and functionality for an experimental design.
This class takes care of data storage and response handling.
}
\keyword{internal}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{num_cores}}{Current number of cores in the global budget.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Design-new}{\code{Design$new()}}
\item \href{#method-Design-add_one_subject_response}{\code{Design$add_one_subject_response()}}
\item \href{#method-Design-add_all_subject_responses}{\code{Design$add_all_subject_responses()}}
\item \href{#method-Design-overwrite_all_subject_assignments}{\code{Design$overwrite_all_subject_assignments()}}
\item \href{#method-Design-is_fixed_sample_size}{\code{Design$is_fixed_sample_size()}}
\item \href{#method-Design-assert_all_subjects_arrived}{\code{Design$assert_all_subjects_arrived()}}
\item \href{#method-Design-assert_all_responses_recorded}{\code{Design$assert_all_responses_recorded()}}
\item \href{#method-Design-check_experiment_completed}{\code{Design$check_experiment_completed()}}
\item \href{#method-Design-assert_even_allocation}{\code{Design$assert_even_allocation()}}
\item \href{#method-Design-assert_blocking_design}{\code{Design$assert_blocking_design()}}
\item \href{#method-Design-assert_equal_block_sizes}{\code{Design$assert_equal_block_sizes()}}
\item \href{#method-Design-assert_fixed_sample}{\code{Design$assert_fixed_sample()}}
\item \href{#method-Design-any_censoring}{\code{Design$any_censoring()}}
\item \href{#method-Design-get_t}{\code{Design$get_t()}}
\item \href{#method-Design-get_X_raw}{\code{Design$get_X_raw()}}
\item \href{#method-Design-get_X_imp}{\code{Design$get_X_imp()}}
\item \href{#method-Design-get_X}{\code{Design$get_X()}}
\item \href{#method-Design-get_y}{\code{Design$get_y()}}
\item \href{#method-Design-get_w}{\code{Design$get_w()}}
\item \href{#method-Design-get_block_ids}{\code{Design$get_block_ids()}}
\item \href{#method-Design-get_n}{\code{Design$get_n()}}
\item \href{#method-Design-get_dead}{\code{Design$get_dead()}}
\item \href{#method-Design-get_prob_T}{\code{Design$get_prob_T()}}
\item \href{#method-Design-get_response_type}{\code{Design$get_response_type()}}
\item \href{#method-Design-duplicate}{\code{Design$duplicate()}}
\item \href{#method-Design-clone}{\code{Design$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-new"></a>}}
\if{latex}{\out{\hypertarget{method-Design-new}{}}}
\subsection{Method \code{new()}}{
Initialize an experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$new(
  response_type,
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = FALSE,
  n = NULL,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{"continuous", "incidence", "proportion", "count", "survival", or
"ordinal".}

\item{\code{prob_T}}{Probability of treatment assignment.}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size (if fixed).}

\item{\code{verbose}}{Flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Design` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_one_subject_response"></a>}}
\if{latex}{\out{\hypertarget{method-Design-add_one_subject_response}{}}}
\subsection{Method \code{add_one_subject_response()}}{
For CARA designs, add a single subject response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$add_one_subject_response(t, y, dead = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{t}}{The subject index.}

\item{\code{y}}{The response value.}

\item{\code{dead}}{If the response is censored (0 for survival).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-add_all_subject_responses"></a>}}
\if{latex}{\out{\hypertarget{method-Design-add_all_subject_responses}{}}}
\subsection{Method \code{add_all_subject_responses()}}{
For non-CARA designs, add all subject responses.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$add_all_subject_responses(ys, deads = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ys}}{The responses as a numeric vector.}

\item{\code{deads}}{The binary vector indicating if dead/censored.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-overwrite_all_subject_assignments"></a>}}
\if{latex}{\out{\hypertarget{method-Design-overwrite_all_subject_assignments}{}}}
\subsection{Method \code{overwrite_all_subject_assignments()}}{
For analysis on already-completed experimental data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$overwrite_all_subject_assignments(w)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{w}}{The binary responses.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-is_fixed_sample_size"></a>}}
\if{latex}{\out{\hypertarget{method-Design-is_fixed_sample_size}{}}}
\subsection{Method \code{is_fixed_sample_size()}}{
Check if this design was initialized with a fixed sample size n
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$is_fixed_sample_size()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE if fixed.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_all_subjects_arrived"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_all_subjects_arrived}{}}}
\subsection{Method \code{assert_all_subjects_arrived()}}{
Asserts if all subjects arrived.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_all_subjects_arrived()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_all_responses_recorded"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_all_responses_recorded}{}}}
\subsection{Method \code{assert_all_responses_recorded()}}{
Asserts if all responses are recorded.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_all_responses_recorded()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-check_experiment_completed"></a>}}
\if{latex}{\out{\hypertarget{method-Design-check_experiment_completed}{}}}
\subsection{Method \code{check_experiment_completed()}}{
Checks if the experiment is completed.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$check_experiment_completed()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if experiment is complete, \code{FALSE} otherwise.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_even_allocation"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_even_allocation}{}}}
\subsection{Method \code{assert_even_allocation()}}{
Checks if the experiment has a 50-50 allocation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_even_allocation()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_blocking_design"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_blocking_design}{}}}
\subsection{Method \code{assert_blocking_design()}}{
Checks whether this design is a blocking design.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_blocking_design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_equal_block_sizes"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_equal_block_sizes}{}}}
\subsection{Method \code{assert_equal_block_sizes()}}{
Checks whether all blocks have the same number of subjects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_equal_block_sizes()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
`TRUE` invisibly if the block sizes are equal.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-assert_fixed_sample"></a>}}
\if{latex}{\out{\hypertarget{method-Design-assert_fixed_sample}{}}}
\subsection{Method \code{assert_fixed_sample()}}{
Checks if the experiment has a fixed sample size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$assert_fixed_sample()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-any_censoring"></a>}}
\if{latex}{\out{\hypertarget{method-Design-any_censoring}{}}}
\subsection{Method \code{any_censoring()}}{
Checks if the experiment has any censored responses
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$any_censoring()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if any censored.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_t"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_t}{}}}
\subsection{Method \code{get_t()}}{
Get t
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_t()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The current number of subjects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X_raw"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_X_raw}{}}}
\subsection{Method \code{get_X_raw()}}{
Get raw X information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_X_raw()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A data frame of subject data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X_imp"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_X_imp}{}}}
\subsection{Method \code{get_X_imp()}}{
Get imputed X information
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_X_imp()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Same as \code{Xraw} except with imputations.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_X"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_X}{}}}
\subsection{Method \code{get_X()}}{
Get X matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_X()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric matrix of subject data.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_y"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_y}{}}}
\subsection{Method \code{get_y()}}{
Get y
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_y()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric vector of subject responses.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_w"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_w}{}}}
\subsection{Method \code{get_w()}}{
Get w
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_w()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A binary vector of subject assignments.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_block_ids"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_block_ids}{}}}
\subsection{Method \code{get_block_ids()}}{
If the design is a block design, get block identifiers (otherwise halts)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_block_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An integer vector of block identifiers.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_n"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_n}{}}}
\subsection{Method \code{get_n()}}{
Get n, the sample size
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_n()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The number of subjects.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_dead"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_dead}{}}}
\subsection{Method \code{get_dead()}}{
Get dead
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_dead()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A binary vector of whether the subject is dead.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_prob_T"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_prob_T}{}}}
\subsection{Method \code{get_prob_T()}}{
Get probability of treatment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_prob_T()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The specified probability.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-get_response_type"></a>}}
\if{latex}{\out{\hypertarget{method-Design-get_response_type}{}}}
\subsection{Method \code{get_response_type()}}{
Get response type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$get_response_type()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The specified response type.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-duplicate"></a>}}
\if{latex}{\out{\hypertarget{method-Design-duplicate}{}}}
\subsection{Method \code{duplicate()}}{
Duplicate this design object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$duplicate(verbose = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{verbose}}{A flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `Design` object with the same data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Design-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Design-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Design$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
