% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_logistic_regression_with_var_cpp}
\alias{fast_logistic_regression_with_var_cpp}
\title{Fast Logistic Regression with variance using Rcpp and IRLS}
\usage{
fast_logistic_regression_with_var_cpp(Xmm, y)
}
\arguments{
\item{Xmm}{Design matrix.}

\item{y}{Response vector.}
}
\value{
A list with coefficients and specific treatment variance.
}
\description{
Fast Logistic Regression with variance using Rcpp and IRLS
}
