% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fast_quasipoisson_regression_with_var_cpp}
\alias{fast_quasipoisson_regression_with_var_cpp}
\title{Fast Quasi-Poisson Regression with variance using Rcpp and IRLS}
\usage{
fast_quasipoisson_regression_with_var_cpp(
  Xmm,
  y,
  j = 2L,
  maxit = 100L,
  tol = 1e-08
)
}
\arguments{
\item{Xmm}{Design matrix.}

\item{y}{Response vector.}

\item{j}{Index of the coefficient for which to compute the variance.}

\item{maxit}{Maximum IRLS iterations.}

\item{tol}{Convergence tolerance for coefficient updates.}
}
\value{
A list with coefficients, dispersion estimate, and selected coefficient variances.
}
\description{
Fast Quasi-Poisson Regression with variance using Rcpp and IRLS
}
