% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fixed_binary_match.R
\name{FixedDesignBinaryMatch}
\alias{FixedDesignBinaryMatch}
\title{A Binary Match Fixed Design}
\description{
An R6 Class encapsulating the data and functionality for a fixed binary match experimental design.
This design pairs subjects based on covariate distances and randomizes within pairs.
}
\section{Super class}{
\code{SeqExpMatch::FixedDesign} -> \code{FixedDesignBinaryMatch}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FixedDesignBinaryMatch-new}{\code{FixedDesignBinaryMatch$new()}}
\item \href{#method-FixedDesignBinaryMatch-redraw_w_according_to_design}{\code{FixedDesignBinaryMatch$redraw_w_according_to_design()}}
\item \href{#method-FixedDesignBinaryMatch-draw_ws_according_to_design}{\code{FixedDesignBinaryMatch$draw_ws_according_to_design()}}
\item \href{#method-FixedDesignBinaryMatch-clone}{\code{FixedDesignBinaryMatch$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_all_subject_assignments"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_all_subject_assignments'><code>SeqExpMatch::FixedDesign$add_all_subject_assignments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_all_subject_responses"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_all_subject_responses'><code>SeqExpMatch::FixedDesign$add_all_subject_responses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_subject"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_subject'><code>SeqExpMatch::FixedDesign$add_subject()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_subject_response"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_subject_response'><code>SeqExpMatch::FixedDesign$add_subject_response()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="any_censoring"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-any_censoring'><code>SeqExpMatch::FixedDesign$any_censoring()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="assert_even_allocation"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-assert_even_allocation'><code>SeqExpMatch::FixedDesign$assert_even_allocation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="assert_experiment_completed"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-assert_experiment_completed'><code>SeqExpMatch::FixedDesign$assert_experiment_completed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="assert_fixed_sample"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-assert_fixed_sample'><code>SeqExpMatch::FixedDesign$assert_fixed_sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="duplicate"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-duplicate'><code>SeqExpMatch::FixedDesign$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_X"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_X'><code>SeqExpMatch::FixedDesign$get_X()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_X_imp"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_X_imp'><code>SeqExpMatch::FixedDesign$get_X_imp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_X_raw"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_X_raw'><code>SeqExpMatch::FixedDesign$get_X_raw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_dead"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_dead'><code>SeqExpMatch::FixedDesign$get_dead()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_n"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_n'><code>SeqExpMatch::FixedDesign$get_n()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_prob_T"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_prob_T'><code>SeqExpMatch::FixedDesign$get_prob_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_response_type"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_response_type'><code>SeqExpMatch::FixedDesign$get_response_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_t"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_t'><code>SeqExpMatch::FixedDesign$get_t()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_w"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_w'><code>SeqExpMatch::FixedDesign$get_w()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_y"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_y'><code>SeqExpMatch::FixedDesign$get_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="is_fixed_sample_size"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-is_fixed_sample_size'><code>SeqExpMatch::FixedDesign$is_fixed_sample_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="randomize"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-randomize'><code>SeqExpMatch::FixedDesign$randomize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="supports_resampling"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-supports_resampling'><code>SeqExpMatch::FixedDesign$supports_resampling()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignBinaryMatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesignBinaryMatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize a binary match fixed experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignBinaryMatch$new(
  response_type = "continuous",
  prob_T = 0.5,
  mahal_match = FALSE,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{The data type of response values.}

\item{\code{prob_T}}{The probability of the treatment assignment. Must be 0.5.}

\item{\code{mahal_match}}{Match using Mahalanobis distance. Default is \code{FALSE} (Euclidean).}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size.}

\item{\code{num_cores}}{The number of CPU cores.}

\item{\code{verbose}}{Flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FixedDesignBinaryMatch` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignBinaryMatch-redraw_w_according_to_design"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesignBinaryMatch-redraw_w_according_to_design}{}}}
\subsection{Method \code{redraw_w_according_to_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignBinaryMatch$redraw_w_according_to_design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignBinaryMatch-draw_ws_according_to_design"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesignBinaryMatch-draw_ws_according_to_design}{}}}
\subsection{Method \code{draw_ws_according_to_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignBinaryMatch$draw_ws_according_to_design(r = 100)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignBinaryMatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesignBinaryMatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignBinaryMatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
