% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_fixed_binary_match.R
\name{FixedDesignBinaryMatch}
\alias{FixedDesignBinaryMatch}
\title{A Binary Match Fixed Design}
\description{
A Binary Match Fixed Design

A Binary Match Fixed Design
}
\details{
An R6 Class encapsulating the data and functionality for a fixed binary match
experimental design.
This design pairs subjects based on covariate distances and randomizes within pairs.
Uses the \pkg{GreedyExperimentalDesign} package for distance computation and random allocation.
}
\section{Super classes}{
\code{\link[EDI:Design]{EDI::Design}} -> \code{\link[EDI:FixedDesign]{EDI::FixedDesign}} -> \code{FixedDesignBinaryMatch}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-FixedDesignBinaryMatch-new}{\code{FixedDesignBinaryMatch$new()}}
\item \href{#method-FixedDesignBinaryMatch-draw_ws_according_to_design}{\code{FixedDesignBinaryMatch$draw_ws_according_to_design()}}
\item \href{#method-FixedDesignBinaryMatch-clone}{\code{FixedDesignBinaryMatch$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="add_one_subject_response"><a href='../../EDI/html/Design.html#method-Design-add_one_subject_response'><code>EDI::Design$add_one_subject_response()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="any_censoring"><a href='../../EDI/html/Design.html#method-Design-any_censoring'><code>EDI::Design$any_censoring()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_all_responses_recorded"><a href='../../EDI/html/Design.html#method-Design-assert_all_responses_recorded'><code>EDI::Design$assert_all_responses_recorded()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_all_subjects_arrived"><a href='../../EDI/html/Design.html#method-Design-assert_all_subjects_arrived'><code>EDI::Design$assert_all_subjects_arrived()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_blocking_design"><a href='../../EDI/html/Design.html#method-Design-assert_blocking_design'><code>EDI::Design$assert_blocking_design()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_equal_block_sizes"><a href='../../EDI/html/Design.html#method-Design-assert_equal_block_sizes'><code>EDI::Design$assert_equal_block_sizes()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_even_allocation"><a href='../../EDI/html/Design.html#method-Design-assert_even_allocation'><code>EDI::Design$assert_even_allocation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="assert_fixed_sample"><a href='../../EDI/html/Design.html#method-Design-assert_fixed_sample'><code>EDI::Design$assert_fixed_sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="check_experiment_completed"><a href='../../EDI/html/Design.html#method-Design-check_experiment_completed'><code>EDI::Design$check_experiment_completed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="duplicate"><a href='../../EDI/html/Design.html#method-Design-duplicate'><code>EDI::Design$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_X"><a href='../../EDI/html/Design.html#method-Design-get_X'><code>EDI::Design$get_X()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_X_imp"><a href='../../EDI/html/Design.html#method-Design-get_X_imp'><code>EDI::Design$get_X_imp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_X_raw"><a href='../../EDI/html/Design.html#method-Design-get_X_raw'><code>EDI::Design$get_X_raw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_block_ids"><a href='../../EDI/html/Design.html#method-Design-get_block_ids'><code>EDI::Design$get_block_ids()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_dead"><a href='../../EDI/html/Design.html#method-Design-get_dead'><code>EDI::Design$get_dead()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_n"><a href='../../EDI/html/Design.html#method-Design-get_n'><code>EDI::Design$get_n()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_prob_T"><a href='../../EDI/html/Design.html#method-Design-get_prob_T'><code>EDI::Design$get_prob_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_response_type"><a href='../../EDI/html/Design.html#method-Design-get_response_type'><code>EDI::Design$get_response_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_t"><a href='../../EDI/html/Design.html#method-Design-get_t'><code>EDI::Design$get_t()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_w"><a href='../../EDI/html/Design.html#method-Design-get_w'><code>EDI::Design$get_w()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="get_y"><a href='../../EDI/html/Design.html#method-Design-get_y'><code>EDI::Design$get_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Design" data-id="is_fixed_sample_size"><a href='../../EDI/html/Design.html#method-Design-is_fixed_sample_size'><code>EDI::Design$is_fixed_sample_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="FixedDesign" data-id="add_all_subject_responses"><a href='../../EDI/html/FixedDesign.html#method-FixedDesign-add_all_subject_responses'><code>EDI::FixedDesign$add_all_subject_responses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="FixedDesign" data-id="add_all_subjects_to_experiment"><a href='../../EDI/html/FixedDesign.html#method-FixedDesign-add_all_subjects_to_experiment'><code>EDI::FixedDesign$add_all_subjects_to_experiment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="FixedDesign" data-id="assign_w_to_all_subjects"><a href='../../EDI/html/FixedDesign.html#method-FixedDesign-assign_w_to_all_subjects'><code>EDI::FixedDesign$assign_w_to_all_subjects()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="FixedDesign" data-id="overwrite_all_subject_assignments"><a href='../../EDI/html/FixedDesign.html#method-FixedDesign-overwrite_all_subject_assignments'><code>EDI::FixedDesign$overwrite_all_subject_assignments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="FixedDesign" data-id="supports_resampling"><a href='../../EDI/html/FixedDesign.html#method-FixedDesign-supports_resampling'><code>EDI::FixedDesign$supports_resampling()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignBinaryMatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesignBinaryMatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize a binary match fixed experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignBinaryMatch$new(
  response_type,
  prob_T = 0.5,
  mahal_match = FALSE,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{response_type}}{The data type of response values.}

\item{\code{prob_T}}{The probability of the treatment assignment. Must be 0.5.}

\item{\code{mahal_match}}{Match using Mahalanobis distance. Default is \code{FALSE} (Euclidean).}

\item{\code{include_is_missing_as_a_new_feature}}{Flag for missingness indicators.}

\item{\code{n}}{The sample size.}

\item{\code{verbose}}{Flag for verbosity.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `FixedDesignBinaryMatch` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignBinaryMatch-draw_ws_according_to_design"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesignBinaryMatch-draw_ws_according_to_design}{}}}
\subsection{Method \code{draw_ws_according_to_design()}}{
Draw multiple treatment assignment vectors according to binary matching.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignBinaryMatch$draw_ws_according_to_design(r = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{r}}{The number of designs to draw.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A matrix of size n x r.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-FixedDesignBinaryMatch-clone"></a>}}
\if{latex}{\out{\hypertarget{method-FixedDesignBinaryMatch-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{FixedDesignBinaryMatch$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
