% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_restricted_mean_se_for_group}
\alias{get_restricted_mean_se_for_group}
\title{Calculates the standard error of the restricted mean survival time for a single group.}
\usage{
get_restricted_mean_se_for_group(y, dead)
}
\arguments{
\item{y}{Numeric vector of survival times.}

\item{dead}{Integer vector of event indicators (1=event, 0=censored).}
}
\value{
The standard error of the restricted mean.
}
\description{
Calculates the standard error of the restricted mean survival time for a single group.
}
\keyword{internal}
