% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract.R
\name{Inference}
\alias{Inference}
\title{Inference for A Sequential Design}
\description{
An abstract R6 Class that estimates, tests and provides intervals for a treatment effect in a completed design.
This class takes a completed \code{Design} object as an input where this object
contains data for a fully completed experiment (i.e. all treatment
assignments were allocated and all responses were collected).
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Inference-new}{\code{Inference$new()}}
\item \href{#method-Inference-duplicate}{\code{Inference$duplicate()}}
\item \href{#method-Inference-clone}{\code{Inference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-new"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$new(des_obj, num_cores = 1, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-duplicate"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-duplicate}{}}}
\subsection{Method \code{duplicate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$duplicate(verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
