% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract.R
\name{Inference}
\alias{Inference}
\title{Inference for A Sequential Design}
\description{
Inference for A Sequential Design

Inference for A Sequential Design
}
\details{
An abstract R6 Class that estimates, tests and provides intervals for a
treatment effect in a completed design.
This class takes a completed \code{Design} object as an input where this object
contains data for a fully completed experiment (i.e. all treatment
assignments were allocated and all responses were collected).
}
\keyword{internal}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Inference-new}{\code{Inference$new()}}
\item \href{#method-Inference-compute_exact_two_sided_pval_for_treatment_effect}{\code{Inference$compute_exact_two_sided_pval_for_treatment_effect()}}
\item \href{#method-Inference-compute_exact_confidence_interval}{\code{Inference$compute_exact_confidence_interval()}}
\item \href{#method-Inference-compute_treatment_estimate}{\code{Inference$compute_treatment_estimate()}}
\item \href{#method-Inference-duplicate}{\code{Inference$duplicate()}}
\item \href{#method-Inference-clone}{\code{Inference$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-new"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$new(des_obj, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-compute_exact_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-compute_exact_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_exact_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$compute_exact_two_sided_pval_for_treatment_effect(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-compute_exact_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-compute_exact_confidence_interval}{}}}
\subsection{Method \code{compute_exact_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$compute_exact_confidence_interval(...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$compute_treatment_estimate(estimate_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-duplicate"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-duplicate}{}}}
\subsection{Method \code{duplicate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$duplicate(verbose = FALSE, make_fork_cluster = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Inference-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Inference-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Inference$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
