% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_KK_clogit_ivwc_abstract.R
\name{InferenceAbstractKKClogitIVWC}
\alias{InferenceAbstractKKClogitIVWC}
\title{Abstract class for Conditional Logistic Compound Inference}
\description{
Abstract class for Conditional Logistic Compound Inference

Abstract class for Conditional Logistic Compound Inference
}
\keyword{internal}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{\link[EDI:InferenceKKPassThrough]{EDI::InferenceKKPassThrough}} -> \code{InferenceAbstractKKClogitIVWC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceAbstractKKClogitIVWC-new}{\code{InferenceAbstractKKClogitIVWC$new()}}
\item \href{#method-InferenceAbstractKKClogitIVWC-compute_treatment_estimate}{\code{InferenceAbstractKKClogitIVWC$compute_treatment_estimate()}}
\item \href{#method-InferenceAbstractKKClogitIVWC-compute_asymp_confidence_interval}{\code{InferenceAbstractKKClogitIVWC$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceAbstractKKClogitIVWC-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceAbstractKKClogitIVWC$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceAbstractKKClogitIVWC-clone}{\code{InferenceAbstractKKClogitIVWC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>EDI::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClogitIVWC-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClogitIVWC-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClogitIVWC$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE,
  make_fork_cluster = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClogitIVWC-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClogitIVWC-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClogitIVWC$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClogitIVWC-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClogitIVWC-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClogitIVWC$compute_asymp_confidence_interval(alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClogitIVWC-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClogitIVWC-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClogitIVWC$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKClogitIVWC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAbstractKKClogitIVWC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAbstractKKClogitIVWC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
