% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_KK_marginal_abstract.R
\name{InferenceAbstractKKMarginalIncid}
\alias{InferenceAbstractKKMarginalIncid}
\title{Abstract class for all-subject marginal incidence inference in KK designs}
\description{
Abstract class for all-subject marginal incidence inference in KK designs

Abstract class for all-subject marginal incidence inference in KK designs
}
\keyword{internal}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \\code{EDI::InferenceKKPassThrough}
-> \code{InferenceAbstractKKMarginalIncid}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceAbstractKKMarginalIncid$new()}
\item \code{InferenceAbstractKKMarginalIncid$clone()}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_confidence_interval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::InferenceAsymp$compute_treatment_estimate()</code></li>
<li><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKMarginalIncid-new"></a>}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\preformatted{InferenceAbstractKKMarginalIncid$new(des_obj, num_cores = 1, verbose = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{Design} object.}

\item{\code{num_cores}}{The number of CPU cores to use.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAbstractKKMarginalIncid-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceAbstractKKMarginalIncid$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
