% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_mean_diff.R
\name{InferenceAllSimpleMeanDiff}
\alias{InferenceAllSimpleMeanDiff}
\title{Simple Mean Difference Inference based on Maximum Likelihood}
\description{
Simple Mean Difference Inference based on Maximum Likelihood

Simple Mean Difference Inference based on Maximum Likelihood
}
\details{
The methods that support confidence intervals and testing for the mean difference
in all response types (except Weibull with censoring)
sequential experimental design estimation and test object
after the sequential design is completed.
For incidence responses coded as 0/1, the asymptotic inference is numerically
equivalent to Welch-Satterthwaite unpooled two-sample inference.
}
\examples{
\dontrun{
seq_des = DesignSeqOneByOneBernoulli$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceAllSimpleMeanDiff$
  new(seq_des)
seq_des_inf$compute_treatment_estimate()
seq_des_inf$compute_asymp_confidence_interval()
seq_des_inf$compute_asymp_two_sided_pval_for_treatment_effect()
}
}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \code{InferenceAllSimpleMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \code{InferenceAllSimpleMeanDiff$compute_treatment_estimate()}
\item \code{InferenceAllSimpleMeanDiff$compute_confidence_interval_rand()}
\item \href{#method-InferenceAllSimpleMeanDiff-clone}{\code{InferenceAllSimpleMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::Inference$initialize()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_confidence_interval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleMeanDiff-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the appropriate estimate for mean difference
\subsection{Usage}{
\preformatted{InferenceAllSimpleMeanDiff$compute_treatment_estimate(estimate_only = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The setting-appropriate (see description) numeric estimate of the treatment effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleMeanDiff-compute_confidence_interval_rand"></a>}}
\subsection{Method \code{compute_confidence_interval_rand()}}{
Computes a 1-alpha level frequentist confidence interval for the randomization test
\subsection{Usage}{
\preformatted{InferenceAllSimpleMeanDiff$compute_confidence_interval_rand(
  alpha = 0.05,
  r = 501,
  pval_epsilon = 0.005,
  show_progress = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}

\item{\code{r}}{The number of randomization vectors. The default is 501.}

\item{\code{pval_epsilon}}{The bisection algorithm tolerance. The default is 0.005.}

\item{\code{show_progress}}{Show a text progress indicator.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 1 - alpha sized frequentist confidence interval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleMeanDiff-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceAllSimpleMeanDiff$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
