% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_simple_wilcox.R
\name{InferenceAllSimpleWilcox}
\alias{InferenceAllSimpleWilcox}
\title{Wilcoxon Rank-Sum Inference for Any Sequential Design}
\description{
Wilcoxon Rank-Sum Inference for Any Sequential Design

Wilcoxon Rank-Sum Inference for Any Sequential Design
}
\details{
Non-parametric inference based on the two-sample Wilcoxon rank-sum test
(Mann-Whitney U) for any sequential experimental design (Bernoulli, Efron, iBCRD,
or KK). The treatment effect estimate is the Hodges-Lehmann pseudo-median of
all pairwise treatment-minus-control response differences. Inference uses the
asymptotic normal approximation for the HL estimator.

This class supports \code{continuous}, \code{count}, \code{proportion}, and
uncensored \code{survival} response types. It throws an informative error for
\code{incidence} (binary) responses and for censored survival data.

The HL estimator satisfies the exact linearity property
\eqn{\hat{\Delta}(\mathbf{y}_T - \delta) = \hat{\Delta}(\mathbf{y}_T) - \delta},
so the randomization confidence interval inversion is exact (not approximate)
on the original response scale for continuous responses.
}
\section{Super classes}{
\code{::Inference} -> \code{::InferenceRand} -> \code{::InferenceRandCI} -> \code{::InferenceBoot} -> \code{::InferenceAsymp} -> \code{InferenceAllSimpleWilcox}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceAllSimpleWilcox-new}{\code{InferenceAllSimpleWilcox$new()}}
\item \href{#method-InferenceAllSimpleWilcox-compute_treatment_estimate}{\code{InferenceAllSimpleWilcox$compute_treatment_estimate()}}
\item \href{#method-InferenceAllSimpleWilcox-compute_asymp_confidence_interval}{\code{InferenceAllSimpleWilcox$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceAllSimpleWilcox-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceAllSimpleWilcox$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceAllSimpleWilcox-clone}{\code{InferenceAllSimpleWilcox$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../..//html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../..//html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../..//html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../..//html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleWilcox-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAllSimpleWilcox-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAllSimpleWilcox$new(des_obj, verbose = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleWilcox-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAllSimpleWilcox-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAllSimpleWilcox$compute_treatment_estimate(estimate_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleWilcox-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAllSimpleWilcox-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAllSimpleWilcox$compute_asymp_confidence_interval(alpha = 0.05)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleWilcox-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAllSimpleWilcox-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAllSimpleWilcox$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceAllSimpleWilcox-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceAllSimpleWilcox-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceAllSimpleWilcox$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
