% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_continuous_robust_regr_multi.R
\name{InferenceContinMultiRobustRegr}
\alias{InferenceContinMultiRobustRegr}
\title{Multivariate Robust Regression Inference for Continuous Responses}
\description{
Fits a robust linear regression via \code{MASS::rlm} for continuous responses
using the treatment indicator and all recorded covariates as predictors.
This provides a Huber/MM-style robustness upgrade over ordinary least squares
when outcomes are heavy-tailed or outlier-prone. Inference is based on the
coefficient table returned by \code{summary.rlm()}.
}
\examples{

## ------------------------------------------------
## Method `InferenceContinMultiRobustRegr$new`
## ------------------------------------------------

set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$new(n = nrow(x_dat), response_type = "continuous", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(c(1.2, 0.9, 1.5, 1.8, 2.1, 1.7, 2.6, 2.2))
infer <- InferenceContinMultiRobustRegr$new(seq_des, verbose = FALSE)
infer

}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{EDI::InferenceBoot} -> \code{EDI::InferenceAsymp} -> \code{\link[EDI:InferenceContinUnivRobustRegr]{EDI::InferenceContinUnivRobustRegr}} -> \code{InferenceContinMultiRobustRegr}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceContinMultiRobustRegr-new}{\code{InferenceContinMultiRobustRegr$new()}}
\item \href{#method-InferenceContinMultiRobustRegr-clone}{\code{InferenceContinMultiRobustRegr$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceContinUnivRobustRegr" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceContinUnivRobustRegr.html#method-InferenceContinUnivRobustRegr-compute_asymp_confidence_interval'><code>EDI::InferenceContinUnivRobustRegr$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceContinUnivRobustRegr" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceContinUnivRobustRegr.html#method-InferenceContinUnivRobustRegr-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceContinUnivRobustRegr$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceContinUnivRobustRegr" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceContinUnivRobustRegr.html#method-InferenceContinUnivRobustRegr-compute_treatment_estimate'><code>EDI::InferenceContinUnivRobustRegr$compute_treatment_estimate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiRobustRegr-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiRobustRegr-new}{}}}
\subsection{Method \code{new()}}{
Initialize a multivariate robust-regression inference object for a completed design
with a continuous response.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiRobustRegr$new(
  des_obj,
  method = "MM",
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A DesignSeqOneByOne object whose entire n subjects are assigned and
whose continuous response y is recorded.}

\item{\code{method}}{Robust-regression fitting method for \code{MASS::rlm}; one
of \code{"M"} or \code{"MM"}. The default is \code{"MM"}.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize
the sampling during randomization-based inference and
bootstrap resampling.
                          and bootstrap resampling.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed to
the user. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$new(n = nrow(x_dat), response_type = "continuous", verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(c(1.2, 0.9, 1.5, 1.8, 2.1, 1.7, 2.6, 2.2))
infer <- InferenceContinMultiRobustRegr$new(seq_des, verbose = FALSE)
infer

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceContinMultiRobustRegr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceContinMultiRobustRegr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceContinMultiRobustRegr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
