% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/inference_count_poisson_KK_cpoisson_ivwc_multi.R
\name{InferenceCountPoissonMultiKKCPoissonIVWC}
\alias{InferenceCountPoissonMultiKKCPoissonIVWC}
\title{Multivariate Conditional Poisson / Negative Binomial Compound Inference for KK Designs}
\description{
Fits a compound estimator for KK matching-on-the-fly designs with count responses
using the treatment indicator and all recorded covariates. For matched pairs, it uses
conditional Poisson regression. For reservoir subjects, it uses Negative Binomial regression.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneKK14$
  new(
  n = nrow(x_dat),
  response_type = "count",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 1, 2, 2, 3, 3, 4))
infer <- InferenceCountPoissonMultiKKCPoissonIVWC$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{EDI::Inference} -> \code{EDI::InferenceRand} -> \code{EDI::InferenceRandCI} -> \code{EDI::InferenceBoot} -> \code{EDI::InferenceAsymp} -> \code{EDI::InferenceKKPassThrough} -> \code{EDI::InferenceAbstractKKPoissonCPoissonIVWC} -> \code{InferenceCountPoissonMultiKKCPoissonIVWC}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceCountPoissonMultiKKCPoissonIVWC-clone}{\code{InferenceCountPoissonMultiKKCPoissonIVWC$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>EDI::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceKKPassThrough" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceKKPassThrough.html#method-InferenceKKPassThrough-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKPoissonCPoissonIVWC" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceAbstractKKPoissonCPoissonIVWC.html#method-InferenceAbstractKKPoissonCPoissonIVWC-compute_asymp_confidence_interval'><code>EDI::InferenceAbstractKKPoissonCPoissonIVWC$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKPoissonCPoissonIVWC" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceAbstractKKPoissonCPoissonIVWC.html#method-InferenceAbstractKKPoissonCPoissonIVWC-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceAbstractKKPoissonCPoissonIVWC$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKPoissonCPoissonIVWC" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceAbstractKKPoissonCPoissonIVWC.html#method-InferenceAbstractKKPoissonCPoissonIVWC-compute_treatment_estimate'><code>EDI::InferenceAbstractKKPoissonCPoissonIVWC$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAbstractKKPoissonCPoissonIVWC" data-id="initialize"><a href='../../EDI/html/InferenceAbstractKKPoissonCPoissonIVWC.html#method-InferenceAbstractKKPoissonCPoissonIVWC-initialize'><code>EDI::InferenceAbstractKKPoissonCPoissonIVWC$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceCountPoissonMultiKKCPoissonIVWC-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceCountPoissonMultiKKCPoissonIVWC-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceCountPoissonMultiKKCPoissonIVWC$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
