% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract_exact.R
\name{InferenceExact}
\alias{InferenceExact}
\title{Exact Inference API}
\description{
Internal method.
Thin interface for exact inference methods.
}
\keyword{internal}
\section{Super class}{
\code{::Inference} -> \code{InferenceExact}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceExact-compute_exact_confidence_interval}{\code{InferenceExact$compute_exact_confidence_interval()}}
\item \href{#method-InferenceExact-compute_exact_two_sided_pval_for_treatment_effect}{\code{InferenceExact$compute_exact_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceExact-clone}{\code{InferenceExact$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_treatment_estimate"><a href='../..//html/Inference.html#method-Inference-compute_treatment_estimate'><code>::Inference$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="initialize"><a href='../..//html/Inference.html#method-Inference-initialize'><code>::Inference$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceExact-compute_exact_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceExact-compute_exact_confidence_interval}{}}}
\subsection{Method \code{compute_exact_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceExact$compute_exact_confidence_interval(
  alpha = 0.05,
  pval_epsilon = 0.005,
  type = NULL,
  args_for_type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceExact-compute_exact_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceExact-compute_exact_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_exact_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceExact$compute_exact_two_sided_pval_for_treatment_effect(
  delta = 0,
  type = NULL,
  args_for_type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceExact-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceExact-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceExact$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
