% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_indicidence_exact_fisher.R
\name{InferenceIncidExactFisher}
\alias{InferenceIncidExactFisher}
\title{Exact Fisher Incidence Inference}
\description{
Exact Fisher Incidence Inference

Exact Fisher Incidence Inference
}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceExact}
-> \code{InferenceIncidExactFisher}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidExactFisher-new}{\code{InferenceIncidExactFisher$new()}}
\item \code{InferenceIncidExactFisher$compute_treatment_estimate()}
\item \href{#method-InferenceIncidExactFisher-clone}{\code{InferenceIncidExactFisher$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::InferenceExact$compute_exact_confidence_interval()</code></li>
<li><code>EDI::InferenceExact$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactFisher-new"></a>}}
\subsection{Method \code{new()}}{
Initialize exact Fisher inference for incidence outcomes.
\subsection{Usage}{
\preformatted{InferenceIncidExactFisher$new(des_obj, num_cores = 1, verbose = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed design object.}

\item{\code{num_cores}}{Number of CPU cores to use.}

\item{\code{verbose}}{Whether to print progress messages.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{InferenceIncidExactFisher} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactFisher-compute_treatment_estimate"></a>}}
\subsection{Method \code{compute_treatment_estimate()}}{
Compute the Fisher exact treatment estimate on the log-odds scale.
\subsection{Usage}{
\preformatted{InferenceIncidExactFisher$compute_treatment_estimate(estimate_only = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{Ignored for this estimator.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The treatment estimate.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactFisher-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceIncidExactFisher$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
