% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_indicidence_exact_fisher.R
\name{InferenceIncidExactFisher}
\alias{InferenceIncidExactFisher}
\title{Exact Fisher Incidence Inference}
\description{
Exact Fisher Incidence Inference

Exact Fisher Incidence Inference
}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceExact]{EDI::InferenceExact}} -> \code{InferenceIncidExactFisher}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidExactFisher-new}{\code{InferenceIncidExactFisher$new()}}
\item \href{#method-InferenceIncidExactFisher-compute_treatment_estimate}{\code{InferenceIncidExactFisher$compute_treatment_estimate()}}
\item \href{#method-InferenceIncidExactFisher-clone}{\code{InferenceIncidExactFisher$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceExact" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/InferenceExact.html#method-InferenceExact-compute_exact_confidence_interval'><code>EDI::InferenceExact$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceExact" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceExact.html#method-InferenceExact-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::InferenceExact$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactFisher-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactFisher-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactFisher$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE,
  make_fork_cluster = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactFisher-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactFisher-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactFisher$compute_treatment_estimate(estimate_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactFisher-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactFisher-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactFisher$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
