% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract_rand_ci.R
\name{InferenceIncidExactZhang}
\alias{InferenceIncidExactZhang}
\title{Exact Zhang Incidence Inference}
\description{
Exact Zhang Incidence Inference

Exact Zhang Incidence Inference
}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{InferenceIncidExactZhang}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidExactZhang-new}{\code{InferenceIncidExactZhang$new()}}
\item \href{#method-InferenceIncidExactZhang-compute_treatment_estimate}{\code{InferenceIncidExactZhang$compute_treatment_estimate()}}
\item \href{#method-InferenceIncidExactZhang-compute_exact_confidence_interval}{\code{InferenceIncidExactZhang$compute_exact_confidence_interval()}}
\item \href{#method-InferenceIncidExactZhang-compute_exact_two_sided_pval_for_treatment_effect}{\code{InferenceIncidExactZhang$compute_exact_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceIncidExactZhang-clone}{\code{InferenceIncidExactZhang$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE,
  make_fork_cluster = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$compute_treatment_estimate(estimate_only = FALSE)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-compute_exact_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-compute_exact_confidence_interval}{}}}
\subsection{Method \code{compute_exact_confidence_interval()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$compute_exact_confidence_interval(
  alpha = 0.05,
  pval_epsilon = 0.005,
  args_for_type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-compute_exact_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-compute_exact_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_exact_two_sided_pval_for_treatment_effect()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$compute_exact_two_sided_pval_for_treatment_effect(
  delta = 0,
  args_for_type = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
