% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_exact_abstract.R
\name{InferenceIncidExactZhang}
\alias{InferenceIncidExactZhang}
\title{Exact Zhang Incidence Randomization CI for KK or Bernoulli Designs}
\description{
Implements the exact test-inversion randomization CI of Zhang (2026). Partitions
subjects into matched pairs (analyzed via binomial test) and reservoir subjects
(analyzed via Fisher's Exact Test), then combines via a p-value combination rule.
}
\details{
Applicable to binary (incidence) outcomes in KK matching-on-the-fly or Bernoulli
designs. A bisection algorithm inverts the combined p-value to construct the CI.
}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{EDI::InferenceExact} -> \code{InferenceIncidExactZhang}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidExactZhang-new}{\code{InferenceIncidExactZhang$new()}}
\item \href{#method-InferenceIncidExactZhang-compute_treatment_estimate}{\code{InferenceIncidExactZhang$compute_treatment_estimate()}}
\item \href{#method-InferenceIncidExactZhang-clone}{\code{InferenceIncidExactZhang$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceExact" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/InferenceExact.html#method-InferenceExact-compute_exact_confidence_interval'><code>EDI::InferenceExact$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceExact" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceExact.html#method-InferenceExact-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::InferenceExact$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Zhang inference object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE,
  make_fork_cluster = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{Design} object with incidence response.}

\item{\code{num_cores}}{Number of CPU cores for parallel bisection search.}

\item{\code{verbose}}{Flag for progress messages.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Returns the incidence treatment estimate (log odds ratio).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidExactZhang-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidExactZhang-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidExactZhang$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
