% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_incidence_gcomp_multi.R
\name{InferenceIncidMultiGCompRiskDiff}
\alias{InferenceIncidMultiGCompRiskDiff}
\title{Multivariate G-Computation Risk-Difference Inference for Binary Responses}
\description{
Fits a logistic working model for an incidence outcome using treatment and all
recorded covariates, then estimates the marginal risk difference by
standardizing predicted risks under all-treated and all-control assignments
over the empirical covariate distribution.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$
  new(
  n = nrow(x_dat),
  response_type = "incidence",
  verbose = FALSE
)
for (i in seq_len(nrow(x_dat))) {
  seq_des$
  add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$
  add_all_subject_responses(c(0, 1, 0, 1, 0, 1, 1, 0))
infer <- InferenceIncidMultiGCompRiskDiff$
  new(
  seq_des,
  verbose = FALSE
)
infer

}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{EDI::InferenceIncidGCompAbstract} -> \code{\link[EDI:InferenceIncidUnivGCompRiskDiff]{EDI::InferenceIncidUnivGCompRiskDiff}} -> \code{InferenceIncidMultiGCompRiskDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceIncidMultiGCompRiskDiff-clone}{\code{InferenceIncidMultiGCompRiskDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>EDI::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="approximate_randomization_distribution_beta_hat_T"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-approximate_randomization_distribution_beta_hat_T'><code>EDI::InferenceRand$approximate_randomization_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidGCompAbstract" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceIncidGCompAbstract.html#method-InferenceIncidGCompAbstract-compute_asymp_confidence_interval'><code>EDI::InferenceIncidGCompAbstract$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidGCompAbstract" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceIncidGCompAbstract.html#method-InferenceIncidGCompAbstract-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceIncidGCompAbstract$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidGCompAbstract" data-id="compute_bootstrap_two_sided_pval"><a href='../../EDI/html/InferenceIncidGCompAbstract.html#method-InferenceIncidGCompAbstract-compute_bootstrap_two_sided_pval'><code>EDI::InferenceIncidGCompAbstract$compute_bootstrap_two_sided_pval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidGCompAbstract" data-id="compute_treatment_estimate"><a href='../../EDI/html/InferenceIncidGCompAbstract.html#method-InferenceIncidGCompAbstract-compute_treatment_estimate'><code>EDI::InferenceIncidGCompAbstract$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceIncidGCompAbstract" data-id="initialize"><a href='../../EDI/html/InferenceIncidGCompAbstract.html#method-InferenceIncidGCompAbstract-initialize'><code>EDI::InferenceIncidGCompAbstract$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceIncidMultiGCompRiskDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceIncidMultiGCompRiskDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceIncidMultiGCompRiskDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
