% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract_KK_passthrough.R
\name{InferenceKKPassThrough}
\alias{InferenceKKPassThrough}
\title{A class that provides for relevant methods when the designs are KK matching-on-the-fly}
\description{
A class that provides for relevant methods when the designs are KK matching-on-the-fly

A class that provides for relevant methods when the designs are KK matching-on-the-fly
}
\details{
An abstract class
}
\examples{

## ------------------------------------------------
## Method `InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T`
## ------------------------------------------------

\dontrun{
seq_des = DesignSeqOneByOneKK14$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLSKK$new(seq_des)
beta_hat_T_bs = seq_des_inf$approximate_bootstrap_distribution_beta_hat_T(B = 5)
beta_hat_T_bs
}

}
\keyword{internal}
\section{Super classes}{
\\code{EDI::Inference}
-> \\code{EDI::InferenceRand}
-> \\code{EDI::InferenceRandCI}
-> \\code{EDI::InferenceBoot}
-> \\code{EDI::InferenceAsymp}
-> \code{InferenceKKPassThrough}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceKKPassThrough-new}{\code{InferenceKKPassThrough$new()}}
\item \code{InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T()}
\item \href{#method-InferenceKKPassThrough-clone}{\code{InferenceKKPassThrough$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><code>EDI::Inference$compute_exact_confidence_interval()</code></li>
<li><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::Inference$duplicate()</code></li>
<li><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></li>
<li><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></li>
<li><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></li>
<li><code>EDI::InferenceRandCI$compute_two_sided_pval_for_treatment_effect_rand()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></li>
<li><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_confidence_interval()</code></li>
<li><code>EDI::InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()</code></li>
<li><code>EDI::InferenceAsymp$compute_treatment_estimate()</code></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceKKPassThrough-new"></a>}}
\subsection{Method \code{new()}}{
Initialize
\subsection{Usage}{
\preformatted{InferenceKKPassThrough$new(des_obj, num_cores = 1, verbose = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A DesignSeqOneByOne object whose entire n subjects are assigned
and response y is recorded within.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize
  the sampling during randomization-based inference
and bootstrap resampling. The default is 1 for serial computation. For simple
estimators (e.g. mean difference
and KK compound), parallelization is achieved with zero-overhead C++ OpenMP. For
complex models (e.g. GLMs),
parallelization falls back to R's \code{parallel::mclapply} which incurs
session-forking overhead.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed
to the user. Default is \code{TRUE}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\subsection{Method \code{approximate_bootstrap_distribution_beta_hat_T()}}{
Creates the boostrap distribution of the estimate for the treatment effect
\subsection{Usage}{
\preformatted{InferenceKKPassThrough$approximate_bootstrap_distribution_beta_hat_T(
  B = 501,
  show_progress = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{B}}{Number of bootstrap samples. The default is 501.}

\item{\code{show_progress}}{Description for show_progress}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A vector of length \code{B} with the bootstrap values of the estimates of the
  treatment effect
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
seq_des = DesignSeqOneByOneKK14$new(n = 6, response_type = "continuous")
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[1, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[2, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[3, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[4, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[5, 2 : 10])
seq_des$add_subject_to_experiment_and_assign(MASS::biopsy[6, 2 : 10])
seq_des$add_all_subject_responses(c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43))

seq_des_inf = InferenceContinMultOLSKK$new(seq_des)
beta_hat_T_bs = seq_des_inf$approximate_bootstrap_distribution_beta_hat_T(B = 5)
beta_hat_T_bs
}

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceKKPassThrough-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\preformatted{InferenceKKPassThrough$clone(deep = FALSE)}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
