% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_ordinal_jonckheere_terpstra_test.R
\name{InferenceOrdinalJonckheereTerpstraTest}
\alias{InferenceOrdinalJonckheereTerpstraTest}
\title{Jonckheere-Terpstra (JT) Test for Ordinal Responses}
\description{
Exact Jonckheere-Terpstra (JT) rank test for a two-arm ordered alternative with an
ordinal response. For treatment versus control, the test statistic is the
sum of Mann-Whitney U counts across groups. This class provides the exact
distribution-based p-value.
}
\examples{
set.seed(1)
x_dat <- data.frame(
  x1 = c(-1.2, -0.7, -0.2, 0.3, 0.8, 1.3, 1.8, 2.3),
  x2 = c(0, 1, 0, 1, 0, 1, 0, 1)
)
seq_des <- DesignSeqOneByOneBernoulli$new(n = nrow(x_dat), response_type = "ordinal",
  verbose = FALSE)
for (i in seq_len(nrow(x_dat))) {
  seq_des$add_subject_to_experiment_and_assign(x_dat[i, , drop = FALSE])
}
seq_des$add_all_subject_responses(as.integer(c(1, 2, 2, 3, 3, 4, 4, 5)))
infer <- InferenceOrdinalJonckheereTerpstraTest$
  new(seq_des, verbose = FALSE)
infer

}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{InferenceOrdinalJonckheereTerpstraTest}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceOrdinalJonckheereTerpstraTest-new}{\code{InferenceOrdinalJonckheereTerpstraTest$new()}}
\item \href{#method-InferenceOrdinalJonckheereTerpstraTest-compute_treatment_estimate}{\code{InferenceOrdinalJonckheereTerpstraTest$compute_treatment_estimate()}}
\item \href{#method-InferenceOrdinalJonckheereTerpstraTest-compute_exact_two_sided_pval_for_treatment_effect}{\code{InferenceOrdinalJonckheereTerpstraTest$compute_exact_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceOrdinalJonckheereTerpstraTest-clone}{\code{InferenceOrdinalJonckheereTerpstraTest$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRandCI" data-id="compute_confidence_interval_rand"><a href='../../EDI/html/InferenceRandCI.html#method-InferenceRandCI-compute_confidence_interval_rand'><code>EDI::InferenceRandCI$compute_confidence_interval_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAsymp" data-id="compute_asymp_confidence_interval"><a href='../../EDI/html/InferenceAsymp.html#method-InferenceAsymp-compute_asymp_confidence_interval'><code>EDI::InferenceAsymp$compute_asymp_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceAsymp" data-id="compute_asymp_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/InferenceAsymp.html#method-InferenceAsymp-compute_asymp_two_sided_pval_for_treatment_effect'><code>EDI::InferenceAsymp$compute_asymp_two_sided_pval_for_treatment_effect()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalJonckheereTerpstraTest-new"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalJonckheereTerpstraTest-new}{}}}
\subsection{Method \code{new()}}{
Initialize the JT test object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalJonckheereTerpstraTest$new(
  des_obj,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{des_obj}}{A completed \code{DesignSeqOneByOne} object.}

\item{\code{num_cores}}{Number of CPU cores.}

\item{\code{verbose}}{Whether to print progress.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalJonckheereTerpstraTest-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalJonckheereTerpstraTest-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Returns the estimated treatment effect (JT superiority measure).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalJonckheereTerpstraTest$compute_treatment_estimate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalJonckheereTerpstraTest-compute_exact_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalJonckheereTerpstraTest-compute_exact_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_exact_two_sided_pval_for_treatment_effect()}}{
Returns the exact two-sided p-value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalJonckheereTerpstraTest$compute_exact_two_sided_pval_for_treatment_effect(
  
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceOrdinalJonckheereTerpstraTest-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceOrdinalJonckheereTerpstraTest-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceOrdinalJonckheereTerpstraTest$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
