% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_all_abstract_rand.R
\name{InferenceRand}
\alias{InferenceRand}
\title{Randomization-based Inference}
\description{
Randomization-based Inference

Randomization-based Inference
}
\details{
Abstract class for randomization-based inference.
}
\keyword{internal}
\section{Super class}{
\code{::Inference} -> \code{InferenceRand}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceRand-set_custom_randomization_statistic_function}{\code{InferenceRand$set_custom_randomization_statistic_function()}}
\item \href{#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null}{\code{InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()}}
\item \href{#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand}{\code{InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()}}
\item \href{#method-InferenceRand-clone}{\code{InferenceRand$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../..//html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../..//html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="compute_treatment_estimate"><a href='../..//html/Inference.html#method-Inference-compute_treatment_estimate'><code>::Inference$compute_treatment_estimate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="duplicate"><a href='../..//html/Inference.html#method-Inference-duplicate'><code>::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="" data-topic="Inference" data-id="initialize"><a href='../..//html/Inference.html#method-Inference-initialize'><code>::Inference$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceRand-set_custom_randomization_statistic_function"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceRand-set_custom_randomization_statistic_function}{}}}
\subsection{Method \code{set_custom_randomization_statistic_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceRand$set_custom_randomization_statistic_function(
  custom_randomization_statistic_function
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null}{}}}
\subsection{Method \code{compute_beta_hat_T_randomization_distr_under_sharp_null()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null(
  r = 501,
  delta = 0,
  transform_responses = "none",
  show_progress = TRUE,
  permutations = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand}{}}}
\subsection{Method \code{compute_two_sided_pval_for_treatment_effect_rand()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceRand$compute_two_sided_pval_for_treatment_effect_rand(
  r = 501,
  delta = 0,
  transform_responses = "none",
  na.rm = TRUE,
  show_progress = TRUE,
  permutations = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceRand-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceRand-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceRand$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
