% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference_survival_rmst.R
\name{InferenceSurvivalRestrictedMeanDiff}
\alias{InferenceSurvivalRestrictedMeanDiff}
\title{Simple Mean Difference Inference based on Maximum Likelihood}
\description{
Simple Mean Difference Inference based on Maximum Likelihood

Simple Mean Difference Inference based on Maximum Likelihood
}
\details{
The methods that support confidence intervals and testing for the mean difference
in all response types (except Weibull with censoring)
sequential experimental design estimation and test object
after the sequential design is completed.
}
\examples{
\dontrun{
# Setup design and add responses
seq_des = DesignSeqOneByOneBernoulli$
  new(n = 6, response_type = "survival")
X = data.table::data.table(MASS::biopsy[1:6, 2:10])
for (i in 1:6) seq_des$
  add_subject_to_experiment_and_assign(X[i, ])
seq_des$
  add_all_subject_responses(
  ys = c(4.71, 1.23, 4.78, 6.11, 5.95, 8.43),
  deads = c(1L, 0L, 1L, 1L, 0L, 1L)
)

# Initialize inference object
seq_des_inf = InferenceSurvivalRestrictedMeanDiff$
  new(seq_des)

# compute_treatment_estimate
seq_des_inf$
  compute_treatment_estimate()

# compute_asymp_confidence_interval
seq_des_inf$
  compute_asymp_confidence_interval()

# compute_asymp_two_sided_pval_for_treatment_effect
seq_des_inf$
  compute_asymp_two_sided_pval_for_treatment_effect()
}

}
\section{Super classes}{
\code{\link[EDI:Inference]{EDI::Inference}} -> \code{\link[EDI:InferenceRand]{EDI::InferenceRand}} -> \code{\link[EDI:InferenceRandCI]{EDI::InferenceRandCI}} -> \code{\link[EDI:InferenceBoot]{EDI::InferenceBoot}} -> \code{\link[EDI:InferenceAsymp]{EDI::InferenceAsymp}} -> \code{InferenceSurvivalRestrictedMeanDiff}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-InferenceSurvivalRestrictedMeanDiff-compute_treatment_estimate}{\code{InferenceSurvivalRestrictedMeanDiff$compute_treatment_estimate()}}
\item \href{#method-InferenceSurvivalRestrictedMeanDiff-compute_asymp_confidence_interval}{\code{InferenceSurvivalRestrictedMeanDiff$compute_asymp_confidence_interval()}}
\item \href{#method-InferenceSurvivalRestrictedMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect}{\code{InferenceSurvivalRestrictedMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect()}}
\item \href{#method-InferenceSurvivalRestrictedMeanDiff-compute_confidence_interval_rand}{\code{InferenceSurvivalRestrictedMeanDiff$compute_confidence_interval_rand()}}
\item \href{#method-InferenceSurvivalRestrictedMeanDiff-clone}{\code{InferenceSurvivalRestrictedMeanDiff$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_confidence_interval"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_confidence_interval'><code>EDI::Inference$compute_exact_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="compute_exact_two_sided_pval_for_treatment_effect"><a href='../../EDI/html/Inference.html#method-Inference-compute_exact_two_sided_pval_for_treatment_effect'><code>EDI::Inference$compute_exact_two_sided_pval_for_treatment_effect()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="duplicate"><a href='../../EDI/html/Inference.html#method-Inference-duplicate'><code>EDI::Inference$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="Inference" data-id="initialize"><a href='../../EDI/html/Inference.html#method-Inference-initialize'><code>EDI::Inference$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_beta_hat_T_randomization_distr_under_sharp_null"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_beta_hat_T_randomization_distr_under_sharp_null'><code>EDI::InferenceRand$compute_beta_hat_T_randomization_distr_under_sharp_null()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="compute_two_sided_pval_for_treatment_effect_rand"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-compute_two_sided_pval_for_treatment_effect_rand'><code>EDI::InferenceRand$compute_two_sided_pval_for_treatment_effect_rand()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceRand" data-id="set_custom_randomization_statistic_function"><a href='../../EDI/html/InferenceRand.html#method-InferenceRand-set_custom_randomization_statistic_function'><code>EDI::InferenceRand$set_custom_randomization_statistic_function()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="approximate_bootstrap_distribution_beta_hat_T"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-approximate_bootstrap_distribution_beta_hat_T'><code>EDI::InferenceBoot$approximate_bootstrap_distribution_beta_hat_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_confidence_interval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_confidence_interval'><code>EDI::InferenceBoot$compute_bootstrap_confidence_interval()</code></a></span></li>
<li><span class="pkg-link" data-pkg="EDI" data-topic="InferenceBoot" data-id="compute_bootstrap_two_sided_pval"><a href='../../EDI/html/InferenceBoot.html#method-InferenceBoot-compute_bootstrap_two_sided_pval'><code>EDI::InferenceBoot$compute_bootstrap_two_sided_pval()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceSurvivalRestrictedMeanDiff-compute_treatment_estimate"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceSurvivalRestrictedMeanDiff-compute_treatment_estimate}{}}}
\subsection{Method \code{compute_treatment_estimate()}}{
Computes the appropriate estimate for mean difference
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceSurvivalRestrictedMeanDiff$compute_treatment_estimate(
  estimate_only = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{estimate_only}}{If TRUE, skip variance component calculations.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The setting-appropriate (see description) numeric estimate of the treatment effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceSurvivalRestrictedMeanDiff-compute_asymp_confidence_interval"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceSurvivalRestrictedMeanDiff-compute_asymp_confidence_interval}{}}}
\subsection{Method \code{compute_asymp_confidence_interval()}}{
Computes a 1-alpha level frequentist confidence interval
differently for all response types, estimate types, and
test types.

Here we use the theory that MLE's computed for GLM's are asymptotically normal.
Hence these confidence intervals are asymptotically valid
and thus approximate for any sample size.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceSurvivalRestrictedMeanDiff$compute_asymp_confidence_interval(
  alpha = 0.05
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A (1 - alpha)-sized frequentist confidence interval for the treatment effect
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceSurvivalRestrictedMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceSurvivalRestrictedMeanDiff-compute_asymp_two_sided_pval_for_treatment_effect}{}}}
\subsection{Method \code{compute_asymp_two_sided_pval_for_treatment_effect()}}{
Computes a 2-sided p-value via the log rank test
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceSurvivalRestrictedMeanDiff$compute_asymp_two_sided_pval_for_treatment_effect(
  delta = 0
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delta}}{The null difference to test against. For any
treatment effect at all this is set to zero (the default).}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The approximate frequentist p-value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceSurvivalRestrictedMeanDiff-compute_confidence_interval_rand"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceSurvivalRestrictedMeanDiff-compute_confidence_interval_rand}{}}}
\subsection{Method \code{compute_confidence_interval_rand()}}{
Computes a 1-alpha level frequentist confidence interval for the randomization test
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceSurvivalRestrictedMeanDiff$compute_confidence_interval_rand(
  alpha = 0.05,
  r = 501,
  pval_epsilon = 0.005,
  show_progress = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{The confidence level in the computed confidence
interval is 1 - \code{alpha}. The default is 0.05.}

\item{\code{r}}{The number of randomization vectors. The default is 501.}

\item{\code{pval_epsilon}}{The bisection algorithm tolerance. The default is 0.005.}

\item{\code{show_progress}}{Show a text progress indicator.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A 1 - alpha sized frequentist confidence interval
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-InferenceSurvivalRestrictedMeanDiff-clone"></a>}}
\if{latex}{\out{\hypertarget{method-InferenceSurvivalRestrictedMeanDiff-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{InferenceSurvivalRestrictedMeanDiff$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
