% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_seq_random_block_size.R
\name{SeqDesignRandomBlockSize}
\alias{SeqDesignRandomBlockSize}
\title{A sequential blocking allocation design with random block sizes}
\description{
An R6 Class encapsulating the data and functionality for a sequential blocking experimental design
where block sizes are randomly chosen from a specified set for each stratum. This design
is commonly used in clinical trials to prevent predictability of treatment assignments.
}
\section{Super classes}{
\code{SeqExpMatch::FixedDesign} -> \code{SeqExpMatch::SeqDesign} -> \code{SeqDesignRandomBlockSize}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SeqDesignRandomBlockSize-new}{\code{SeqDesignRandomBlockSize$new()}}
\item \href{#method-SeqDesignRandomBlockSize-assign_wt}{\code{SeqDesignRandomBlockSize$assign_wt()}}
\item \href{#method-SeqDesignRandomBlockSize-redraw_w_according_to_design}{\code{SeqDesignRandomBlockSize$redraw_w_according_to_design()}}
\item \href{#method-SeqDesignRandomBlockSize-clone}{\code{SeqDesignRandomBlockSize$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_all_subject_assignments"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_all_subject_assignments'><code>SeqExpMatch::FixedDesign$add_all_subject_assignments()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_all_subject_responses"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_all_subject_responses'><code>SeqExpMatch::FixedDesign$add_all_subject_responses()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_subject"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_subject'><code>SeqExpMatch::FixedDesign$add_subject()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="add_subject_response"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-add_subject_response'><code>SeqExpMatch::FixedDesign$add_subject_response()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="any_censoring"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-any_censoring'><code>SeqExpMatch::FixedDesign$any_censoring()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="assert_even_allocation"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-assert_even_allocation'><code>SeqExpMatch::FixedDesign$assert_even_allocation()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="assert_experiment_completed"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-assert_experiment_completed'><code>SeqExpMatch::FixedDesign$assert_experiment_completed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="assert_fixed_sample"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-assert_fixed_sample'><code>SeqExpMatch::FixedDesign$assert_fixed_sample()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="draw_ws_according_to_design"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-draw_ws_according_to_design'><code>SeqExpMatch::FixedDesign$draw_ws_according_to_design()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="duplicate"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-duplicate'><code>SeqExpMatch::FixedDesign$duplicate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_X"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_X'><code>SeqExpMatch::FixedDesign$get_X()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_X_imp"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_X_imp'><code>SeqExpMatch::FixedDesign$get_X_imp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_X_raw"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_X_raw'><code>SeqExpMatch::FixedDesign$get_X_raw()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_dead"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_dead'><code>SeqExpMatch::FixedDesign$get_dead()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_n"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_n'><code>SeqExpMatch::FixedDesign$get_n()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_prob_T"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_prob_T'><code>SeqExpMatch::FixedDesign$get_prob_T()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_response_type"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_response_type'><code>SeqExpMatch::FixedDesign$get_response_type()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_t"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_t'><code>SeqExpMatch::FixedDesign$get_t()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_w"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_w'><code>SeqExpMatch::FixedDesign$get_w()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="get_y"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-get_y'><code>SeqExpMatch::FixedDesign$get_y()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="is_fixed_sample_size"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-is_fixed_sample_size'><code>SeqExpMatch::FixedDesign$is_fixed_sample_size()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="FixedDesign" data-id="randomize"><a href='../../SeqExpMatch/html/FixedDesign.html#method-FixedDesign-randomize'><code>SeqExpMatch::FixedDesign$randomize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="add_subject_to_experiment_and_assign"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-add_subject_to_experiment_and_assign'><code>SeqExpMatch::SeqDesign$add_subject_to_experiment_and_assign()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="print_current_subject_assignment"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-print_current_subject_assignment'><code>SeqExpMatch::SeqDesign$print_current_subject_assignment()</code></a></span></li>
<li><span class="pkg-link" data-pkg="SeqExpMatch" data-topic="SeqDesign" data-id="supports_resampling"><a href='../../SeqExpMatch/html/SeqDesign.html#method-SeqDesign-supports_resampling'><code>SeqExpMatch::SeqDesign$supports_resampling()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignRandomBlockSize-new"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignRandomBlockSize-new}{}}}
\subsection{Method \code{new()}}{
Initialize a random block size sequential experimental design
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignRandomBlockSize$new(
  strata_cols = NULL,
  block_sizes = c(4, 6, 8),
  response_type = "continuous",
  prob_T = 0.5,
  include_is_missing_as_a_new_feature = TRUE,
  n = NULL,
  num_cores = 1,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{strata_cols}}{A character vector of column names to use for stratification. If NULL, simple blocking is used.}

\item{\code{block_sizes}}{A vector of positive integers representing the possible block sizes to choose from. 
Each must be a multiple of the inverse of \code{prob_T} to ensure integer treatment/control counts.}

\item{\code{response_type}}{The data type of response values which must be one of the following:
"continuous" (the default),
"incidence",
"proportion",
"count",
"survival",
"ordinal".}

\item{\code{prob_T}}{The probability of the treatment assignment. This defaults to \code{0.5}.}

\item{\code{include_is_missing_as_a_new_feature}}{If missing data is present in a variable,
should we include another dummy variable for its missingness? Default is \code{TRUE}.}

\item{\code{n}}{The sample size (if fixed). Default is \code{NULL} for not fixed.}

\item{\code{num_cores}}{The number of CPU cores to use to parallelize the sampling.}

\item{\code{verbose}}{A flag indicating whether messages should be displayed. Default is \code{FALSE}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new `SeqDesignRandomBlockSize` object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignRandomBlockSize-assign_wt"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignRandomBlockSize-assign_wt}{}}}
\subsection{Method \code{assign_wt()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignRandomBlockSize$assign_wt()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignRandomBlockSize-redraw_w_according_to_design"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignRandomBlockSize-redraw_w_according_to_design}{}}}
\subsection{Method \code{redraw_w_according_to_design()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignRandomBlockSize$redraw_w_according_to_design()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SeqDesignRandomBlockSize-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SeqDesignRandomBlockSize-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SeqDesignRandomBlockSize$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
