% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqExpMatch.R
\docType{package}
\name{SeqExpMatch}
\alias{SeqExpMatch-package}
\alias{SeqExpMatch}
\title{Sequential Experimental Designs via Matching On-the-Fly}
\description{
SeqExpMatch
}
\details{
Generates the following sequential two-arm experimental designs
(1) completely randomized (Bernoulli)
(2) balanced completely randomized
(3) Efron's (1971) Biased Coin
(4) Atkinson's (1982) Covariate-Adjusted Biased Coin
(5) Kapelner and Krieger's (2014) Covariate-Adjusted Matching on the Fly
(6) Kapelner and Krieger's (2021) CARA Matching on the Fly with Weighted Covariates
(7) Kapelner and Krieger's (2021) CARA Matching on the Fly with Weighted Covariates Stepwise
}
\references{
Adam Kapelner and Abba Krieger, A Matching Procedure for Sequential Experiments that Iteratively Learns which Covariates Improve Power, Biometrics (2021), \doi{10.1111/biom.13561}
}
\author{
Adam Kapelner \email{kapelner@qc.cuny.edu}
}
\keyword{design}
\keyword{htest}
