% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_cluster.R
\name{set_num_cores}
\alias{set_num_cores}
\title{Set the number of cores for parallelization}
\usage{
set_num_cores(num_cores, force_mirai = FALSE)
}
\arguments{
\item{force_mirai}{If \code{TRUE}, forces the use of the \code{mirai} package
even on systems where forking is available.}
}
\value{
Invisible \code{NULL}.
}
\description{
This function initializes a persistent parallel cluster (either a fork cluster
on Unix-like systems or a mirai cluster on others) to be used by all Design
and Inference objects. This avoids the overhead of creating clusters
repeatedly.
}
