% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{summary_glm_lean}
\alias{summary_glm_lean}
\title{Lean GLM Summary}
\usage{
summary_glm_lean(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  ...
)
}
\arguments{
\item{object}{The GLM object}

\item{dispersion}{See GLM documentation}

\item{correlation}{See GLM documentation}

\item{symbolic.cor}{See GLM documentation}

\item{...}{Other parameters (currently unused)}
}
\value{
The summary of the GLM
}
\description{
Same as summary.glm except it doesn't calculate the residuals as this takes a long time
}
